Beverton Holt function to calculate density-dependent reproduction rate

Takes the density-independent rates \(R_{di}\) of egg production (as calculated by getRDI()) and returns reduced, density-dependent reproduction rates \(R_{dd}\) given as $$R_{dd} = R_{di} \frac{R_{max}}{R_{di} + R_{max}}$$ where \(R_{max}\) are the maximum possible reproduction rates that must be specified in a column in the species parameter dataframe. (All quantities in the above equation are species-specific but we dropped the species index for simplicity.)

Usage

BevertonHoltRDD(rdi, species_params, ...)

Arguments

rdi

Vector of density-independent reproduction rates \(R_{di}\) for all species.

species_params

A species parameter dataframe. Must contain a column R_max holding the maximum reproduction rate \(R_{max}\) for each species.

...

Unused

Value

Vector of density-dependent reproduction rates.

Details

This is only one example of a density-dependence. You can write your own function based on this example, returning different density-dependent reproduction rates. Three other examples provided are RickerRDD(), SheperdRDD(), noRDD() and constantRDD(). For more explanation see setReproduction().

See also