Example MizerParams object for the North East US Continential Shelf (NEUSCS) with 24 species.

Created with purely size-based predation, i.e., no species-specific interactions. Set up with three fishing gears targeting small, medium and large species. Vulnerabilities are represented by changing the clearance rate constant (gamma) between species. Calibrated to efforts effort = c(small = 0.4, medium = 0.3, large = 0.25).

Format

A MizerParams object

Source

N.S. Jacobsen, M. Burgess and K.H. Andersen (2017): Efficiency of fisheries is increasing at the ecosystem level. Fish and Fisheries 18(2) 199- 211. doi:10.1111/faf.12171.

See also

Examples

if (FALSE) {
sim = project(NEUSCS_params, effort = c(small = 0.4, medium = 0.3, large = 0.25))
plot(sim)
}