Set up plankton

Set up plankton

changePlankton(params, kappa = params@kappa, lambda = params@lambda,
  r_pp = 10, w_pp_cutoff = 10, plankton_dynamics = NULL)

Arguments

params
A MizerParams object

kappa
Carrying capacity of the plankton spectrum.

lambda
Exponent of the plankton spectrum.

r_pp
Growth rate of the primary productivity. Default is 10 g/year.

w_pp_cutoff
The upper cut off size of the plankton spectrum. Default is 10 g.

plankton_dynamics
Function that determines plankton dynamics by calculating the plankton spectrum at the next time step from the current state.

Value

A MizerParams object

Setting plankton dynamics

Still need to document

See also