Set proportion of energy that is invested into reproduction

Sets the proportion of the total energy available for reproduction and growth that is invested into reproduction as a function of the size of the individual and sets the reproductive efficiency.

changeReproduction(params, maturity = NULL, repro_prop = NULL)

Arguments

params
A MizerParams object

maturity
Optional. An array (species x size) that holds the proportion of individuals of each species at size that are mature. If not supplied, a default is set as described in the section "Setting reproduction".

repro_prop
Optional. An array (species x size) that holds the proportion of consumed energy that a mature individual allocates to reproduction for each species at size. If not supplied, a default is set as described in the section "Setting reproduction".

Value

The MizerParams object.

Setting reproduction

For each species and at each size, the proportion of the available energy that is invested into reproduction is the product of two factors: the proportion maturity of individuals that are mature and the proportion repro_prop of the energy available to a mature individual that is invested into reproduction.

If the maturity argument is not supplied, then it is set to a sigmoidal maturity ogive that changes from 0 to 1 at around the maturity size: $${\tt maturity}(w) = \left[1+\left(\frac{w}{w_{mat}}\right)^{-U}\right]^{-1}.$$ (To avoid clutter, we are not showing the species index in the equations.) The maturity weights are taken from the w_mat column of the species_params data frame. Any missing maturity weights are set to 1/4 of the asymptotic weight in the w_inf column. The exponent \(U\) determines the steepness of the maturity ogive. By default it is chosen as \(U = 10\), however this can be overridden by including a column w_mat25 in the species parameter dataframe that specifies the weight at which 25% of individuals are mature, which sets \(U = \log(3) / \log(w_{mat} / w_{25}).\)

The sigmoidal function given above would strictly reach 1 only asymptotically. Mizer instead sets the function equal to 1 already at the species' maximum size, taken from the compulsory w_inf column in the species_params data frame.

If the repro_prop argument is not supplied, it is set to the allometric form $${\tt repro\_prop}(w) = \left(\frac{w}{w_{inf}}\right)^{m-n}.$$ Here \(n\) is the scaling exponent of the energy income rate. Hence the exponent \(m\) determines the scaling of the investment into reproduction for mature individuals. By default it is chosen to be \(m = 1\) so that the rate at which energy is invested into reproduction scales linearly with the size. This default can be overridden by including a column m in the species parameter dataframe. The asymptotic sizes are taken from the compulsory w_inf column in the species_params data frame.

The reproductive efficiency, i.e., the proportion of energy allocated to reproduction that results in offspring biomass, is set from the erepro column in the species_params data frame. If that is not provided the default is set to 1 (which you will want to override).

See also