Set default reproduction proportion

Sets the proportion of the energy available for reproduction and growth that is invested into reproduction as a function of the size of the individual.

defaultReproProp(params)

Arguments

params
A MizerParams object

Value

A MizerParams object

Details

The proportion is set to the product of a sigmoidal maturity ogive that gives the proportion of individuals of a given species and size that are mature, and a factor that describes how investment into reproduction by mature individuals scales with size. In formulas: $$\psi(w) = \left[1+\left(\frac{w}{w_{mat}}\right)^{-U}\right]^{-1} \left(\frac{w}{w_{inf}}\right)^{m-n}.$$ Here \(n\) is the scaling exponent of the energy income rate. The exponent \(m\) determines the scaling of the investment into reproduction for mature individuals. By default it is chosen to be \(m = 1\) so that the rate at which energy is invested into reproduction scales linearly with the size. This default can be overridden by including a column m in the species parameter dataframe.

The exponent \(U\) determines the steepness of the maturity ogive. By default it is chosen as \(U = 10\), however this can be overridden by including a column w_mat25 in the species parameter dataframe that specifies the weight at which 25% of individuals are mature, which sets $$U = \frac{\log(3)}{\log(w_{mat} / w_{25})}$$

The result for \(\psi(w)\) for all species is stored in the psi slot of the params object and this object is returned.