Detritus dynamics

Calculates the detritus biomass at the next timestep from the current detritus biomass.

detritus_dynamics(params, n, n_pp, B, rates, dt,
  detritus_external = params@resource_params$detritus_external,
  detritus_proportion = params@resource_params$detritus_proportion, ...)

Arguments

params
A MizerParams object

n
A matrix of current species abundances (species x size)

n_pp
A vector of current plankton abundance by size

B
A vector of current resource biomasses

rates
A list of rates as returned by getRates()

dt
Time step

detritus_external
Rate of change from external sources

detritus_proportion
Proportion of consumption by fish that flows into the detritus component.

...
Unused

Value

A single number giving the biomass of detritus at next time step

Details

The equation for the time evolution of the detritus biomass \(B\) is assumed to be of the form $$dB/dt = inflow - consumption * B + external$$ where

inflow comes from feces, calculated as a proportion detritus_proportion of the biomass consumed by all consumers.

consumption is by detritivorous species, where the encounter rate is specified by params@rho[, "detritus", ].

external is an influx from external sources. It can be negative in which case it represents a loss to external sources.

This equation is solved analytically to $$B(t+dt) = B(t)\exp(-consumption \cdot dt) +\frac{inflow + external}{consumption} (1-\exp(-consumption \cdot dt)).$$ This avoids the stability problems that would arise if we used the Euler method to solve the equation numerically.

See also