Get information about other ecosystem components

Get information about other ecosystem components

Usage

getComponent(params, component)

Arguments

params

A MizerParams object

component

Name of the component of interest. If missing, a list of all components will be returned.

Value

A list with the entries initial_value, dynamics_fun, encounter_fun, mort_fun, component_params for the requested component. If the requested component does not exist, NULL is returned. If no component argument is given, then a list of lists for all components is returned.