Get growth curves giving weight as a function of age

If given a MizerSim object, uses the growth rates at the final time of a simulation to calculate the size at age. If given a MizerParams object, uses the initial growth rates instead.

getGrowthCurves(object, max_age = 24)

Arguments

object
MizerSim or MizerParams object

max_age
The age up to which the weight is to be plotted. Default is 20

Value

An array (species x age) containing the weight in grams.

Examples

data(NS_species_params_gears)
data(inter)
params <- suppressMessages(set_multispecies_model(NS_species_params_gears, inter))
getGrowthCurves(params)

#>          Age
#> Species       0 0.489795918367347 0.979591836734694 1.46938775510204
#>   Sprat   0.001         0.3022132         1.6802528         4.743035
#>   Sandeel 0.001         0.8322811         4.4587948         7.823354
#>   N.pout  0.001         1.2903992         7.6145909        21.162349
#>   Herring 0.001         1.5274228         9.0555478        25.983261
#>   Dab     0.001         1.0908586         6.4081541        18.174056
#>   Whiting 0.001         0.9071907         5.2972941        15.138424
#>   Sole    0.001         0.5087551         2.9051276         8.263291
#>   Gurnard 0.001         0.3517496         1.9720635         5.580511
#>   Plaice  0.001         0.1801163         0.9674818         2.702355
#>   Haddock 0.001         1.7255450        10.2594485        29.433381
#>   Cod     0.001         6.4458336        38.9488064       109.298971
#>   Saithe  0.001         3.7320080        22.4787134        63.925297
#>          Age
#> Species   1.95918367346939 2.44897959183673 2.93877551020408 3.42857142857143
#>   Sprat           9.932817         14.83905         17.48829         19.44856
#>   Sandeel        11.218248         14.53658         17.62694         20.41406
#>   N.pout         31.506472         38.86652         45.46705         51.47487
#>   Herring        54.941626         94.36491        122.24836        141.53462
#>   Dab            30.976362         43.57647         57.09078         71.16661
#>   Whiting        31.485263         54.09719         78.68002         98.85516
#>   Sole           17.441846         31.15996         50.00236         73.01013
#>   Gurnard        11.758428         21.00035         33.45092         46.09430
#>   Plaice          5.663479         10.09367         16.20762         24.19176
#>   Haddock        62.097973        110.32774        170.19997        226.30928
#>   Cod           216.359206        346.04973        479.90597        603.62332
#>   Saithe        129.338276        210.78100        295.20584        371.71767
#>          Age
#> Species   3.91836734693878 4.40816326530612 4.89795918367347 5.38775510204082
#>   Sprat           21.06527         22.48612         23.73947         24.84670
#>   Sandeel         22.86799         25.00027         26.81983         28.36918
#>   N.pout          56.90179         61.77835         66.13063         70.01297
#>   Herring        158.00553        173.01144        186.91004        199.82365
#>   Dab             85.46166         99.71009        113.70391        127.30740
#>   Whiting        117.08320        134.57160        151.43754        167.60914
#>   Sole            93.67704        111.83009        129.73861        147.94098
#>   Gurnard         57.55631         69.29325         81.57341         94.34721
#>   Plaice          34.20480         46.37699         60.80349         77.43791
#>   Haddock        282.94675        342.93013        406.02365        471.64532
#>   Cod            709.59646        795.64817        863.56260        915.25392
#>   Saithe         435.03221        484.30571        521.01003        547.61729
#>          Age
#> Species   5.87755102040816 6.36734693877551 6.85714285714286 7.3469387755102
#>   Sprat           25.82717         26.69528         27.46394        28.14454
#>   Sandeel         29.67405         30.76046         31.66498        32.41806
#>   N.pout          73.45843         76.51170         79.21738        81.61506
#>   Herring        211.77193        222.79681        232.96969       242.33674
#>   Dab            140.40535        152.92462        164.82795       176.09919
#>   Whiting        183.02398        197.65834        211.50784       224.56321
#>   Sole           166.44532        185.17017        204.03292       222.91504
#>   Gurnard        107.52509        121.01188        134.70988       148.56071
#>   Plaice          95.52966        113.29183        130.03474       146.43714
#>   Haddock        539.24919        608.33575        678.40353       749.11747
#>   Cod            954.51766        984.61347       1009.83473      1031.45888
#>   Saithe         566.90439        581.37749        593.89457       604.87620
#>          Age
#> Species   7.83673469387755 8.3265306122449 8.81632653061224 9.30612244897959
#>   Sprat           28.74837        29.28796         29.77047         30.20196
#>   Sandeel         33.04507        33.56709         34.00511         34.40581
#>   N.pout          83.73838        85.61609         87.27680         88.74554
#>   Herring        250.89065       258.69593        265.81807        272.31681
#>   Dab            186.70363       196.67997        206.03717        214.76417
#>   Whiting        236.87036       248.45046        259.32689        269.54280
#>   Sole           241.77026       260.49182        279.04057        297.37303
#>   Gurnard        162.47349       176.40957        190.28823        204.09628
#>   Plaice         163.03571       179.96645        197.28018        214.94508
#>   Haddock        820.12436       891.11762        961.85906       1032.18761
#>   Cod           1049.99927      1065.89495       1079.52347       1091.20816
#>   Saithe         614.51089       622.96345        630.37910        636.88503
#>          Age
#> Species   9.79591836734694 10.2857142857143 10.7755102040816 11.265306122449
#>   Sprat           30.58780         30.93283         31.24136        31.51726
#>   Sandeel         34.78134         35.13335         35.46329        35.77254
#>   N.pout          90.04449         91.19328         92.20927        93.10781
#>   Herring        278.24665        283.65653        288.57478       293.03934
#>   Dab            222.90240        230.49157        237.56834       244.13917
#>   Whiting        279.13816        288.15073        296.61583       304.56671
#>   Sole           315.40841        333.14861        350.56518       367.59215
#>   Gurnard        217.75794        231.25779        244.58196       257.66877
#>   Plaice         232.91522        251.15778        269.59853       288.23076
#>   Haddock       1101.81956       1170.75049       1238.76609      1305.74406
#>   Cod           1101.22609       1109.81521       1117.17937      1123.49319
#>   Saithe         642.59279        647.60034        651.99362       655.84795
#>          Age
#> Species   11.7551020408163 12.2448979591837 12.734693877551 13.2244897959184
#>   Sprat           31.76398         31.98461        32.18189         32.35831
#>   Sandeel         36.06240         36.33409        36.58873         36.82742
#>   N.pout          93.90248         94.60528        95.22683         95.77654
#>   Herring        297.09206        300.77096       304.11046        307.14191
#>   Dab            250.22424        255.85955       261.07832        265.91148
#>   Whiting        312.03460        319.04886       325.63703        331.82501
#>   Sole           384.23573        400.50458       416.37342        431.80345
#>   Gurnard        270.51569        283.12687       295.46805        307.51453
#>   Plaice         306.97565        325.80655       344.71488        363.62612
#>   Haddock       1371.70023       1436.56809      1500.19177       1562.58695
#>   Cod           1128.90648       1133.54770      1137.52696       1140.93866
#>   Saithe         659.22946        662.19614       664.79889        667.08234
#>          Age
#> Species   13.7142857142857 14.2040816326531 14.6938775510204 15.1836734693878
#>   Sprat           32.51607         32.65715         32.78330         32.89611
#>   Sandeel         37.05113         37.26081         37.45735         37.64157
#>   N.pout          96.26270         96.69265         97.07291         97.40921
#>   Herring        309.89371        312.39165        314.65917        316.71752
#>   Dab            270.38751        274.53268        278.37147        281.92654
#>   Whiting        337.63832        343.10838        348.26683        353.13022
#>   Sole           446.80666        461.39476        475.57922        489.35733
#>   Gurnard        319.27352        330.75183        341.95086        352.83923
#>   Plaice         382.52209        401.40268        420.22764        438.95121
#>   Haddock       1623.77830       1683.76274       1742.44595       1799.83051
#>   Cod           1143.86377       1146.37167       1148.52188       1150.36542
#>   Saithe         669.08568        670.84325        672.38522        673.73802
#>          Age
#> Species   15.6734693877551 16.1632653061224 16.6530612244898 17.1428571428571
#>   Sprat           32.99699         33.08719         33.16786         33.23999
#>   Sandeel         37.81423         37.97606         38.12775         38.26993
#>   N.pout          97.70686         97.97884         98.24604         98.50942
#>   Herring        318.58599        320.28210        321.82175        323.21937
#>   Dab            285.23711        288.34820        291.27188        294.01994
#>   Whiting        357.71534        362.03813        366.11357        369.95584
#>   Sole           502.69998        515.61775        528.12478        540.23390
#>   Gurnard        363.41783        373.69548        383.68076        393.38197
#>   Plaice         457.57354        476.09521        494.49446        512.72689
#>   Haddock       1855.94516       1910.81798       1964.47579       2016.92616
#>   Cod           1151.94601       1153.30117       1154.46305       1155.45922
#>   Saithe         674.92488        675.96613        676.87965        677.68110
#>          Age
#> Species   17.6326530612245 18.1224489795918 18.6122448979592 19.1020408163265
#>   Sprat           33.30450         33.36218         33.41376         33.45988
#>   Sandeel         38.40319         38.52809         38.64516         38.75489
#>   N.pout          98.76903         99.02494         99.27719         99.52583
#>   Herring        324.48807        325.63973        326.68515        327.63413
#>   Dab            296.60274        299.03026        301.31183        303.45619
#>   Whiting        373.57826        376.99341        380.21316        383.24869
#>   Sole           551.95780        563.30872        574.29732        584.91533
#>   Gurnard        402.80713        411.94503        420.78994        429.35133
#>   Plaice         530.79227        548.69213        566.42798        583.99050
#>   Haddock       2068.12279       2118.10573       2166.90166       2214.53912
#>   Cod           1156.31330       1157.04557       1157.67339       1158.21168
#>   Saithe         678.38424        679.00112        679.54232        680.01714
#>          Age
#> Species   19.5918367346939 20.0816326530612 20.5714285714286 21.0612244897959
#>   Sprat           33.50113         33.53801          33.5710         33.60049
#>   Sandeel         38.85774         38.95415          39.0445         39.12919
#>   N.pout          99.77091        100.01249         100.2506        100.48535
#>   Herring        328.49558        329.27756         329.9874        330.63176
#>   Dab            305.47161        307.36583         309.1462        310.81942
#>   Whiting        386.11053        388.80863         391.3524        393.75053
#>   Sole           595.16239        605.05147         614.5951        623.80529
#>   Gurnard        437.63845        445.66000         453.4245        460.94021
#>   Plaice         601.34506        618.49070         635.4300        652.16544
#>   Haddock       2261.04563       2306.44803        2350.7724       2394.03775
#>   Cod           1158.67318       1159.06887        1159.4081       1159.69898
#>   Saithe         680.43370        680.79916         681.1198        681.40109
#>          Age
#> Species   21.5510204081633 22.0408163265306 22.530612244898 23.0204081632653
#>   Sprat           33.62686         33.65045        33.67154         33.69040
#>   Sandeel         39.20857         39.28297        39.35271         39.41808
#>   N.pout         100.71672        100.94478       101.16958        101.39117
#>   Herring        331.21668        331.74765       332.22963        332.66715
#>   Dab            312.39207        313.87015       315.25935        316.56501
#>   Whiting        396.01150        398.14309       400.15273        402.04740
#>   Sole           632.69373        641.27166       649.54991        657.53897
#>   Gurnard        468.21505        475.25678       482.07193        488.65550
#>   Plaice         668.69945        685.03425       701.15364        717.04759
#>   Haddock       2436.26028       2477.46424      2517.67458       2556.91518
#>   Cod           1159.94836       1160.16217      1160.34549       1160.50266
#>   Saithe         681.64788        681.86439       682.05435        682.22100
#>          Age
#> Species   23.5102040816327         24
#>   Sprat           33.70726   33.72234
#>   Sandeel         39.47934   39.53676
#>   N.pout         101.60959  101.82489
#>   Herring        333.06432  333.42484
#>   Dab            317.79216  318.94553
#>   Whiting        403.86237  405.63501
#>   Sole           665.24893  672.68954
#>   Gurnard        495.01051  501.14556
#>   Plaice         732.71927  748.17185
#>   Haddock       2595.20946 2632.58025
#>   Cod           1160.63741 1160.75294
#>   Saithe         682.36720  682.49548

sim <- project(params, effort=1, t_max=20, t_save = 2, progress_bar = FALSE)
getGrowthCurves(sim, max_age = 24)

#>          Age
#> Species       0 0.489795918367347 0.979591836734694 1.46938775510204
#>   Sprat   0.001         0.2993893         1.6332965         4.515792
#>   Sandeel 0.001         0.8270020         4.4027066         7.706426
#>   N.pout  0.001         0.8669787         4.7089976        15.856929
#>   Herring 0.001         1.5045243         8.7433378        24.591701
#>   Dab     0.001         0.8857705         4.4159471        11.418070
#>   Whiting 0.001         0.6200264         3.0509787         9.215514
#>   Sole    0.001         0.4489029         2.2826839         5.838531
#>   Gurnard 0.001         0.3121939         1.5725175         4.041732
#>   Plaice  0.001         0.1588358         0.7626962         1.921199
#>   Haddock 0.001         1.4569440         7.6251975        20.268913
#>   Cod     0.001         3.8810468        23.1880366        99.052230
#>   Saithe  0.001         2.1297511         9.5580091        26.054034
#>          Age
#> Species   1.95918367346939 2.44897959183673 2.93877551020408 3.42857142857143
#>   Sprat           9.287000        14.238144        16.968269         18.90032
#>   Sandeel        11.016086        14.251343        17.271916         20.00614
#>   N.pout         31.024870        41.618668        51.471604         60.38953
#>   Herring        51.017054        87.736425       116.717390        135.84884
#>   Dab            22.188395        32.477220        43.738865         56.51383
#>   Whiting        24.467547        58.734769       104.175268        151.32079
#>   Sole           11.250204        18.671902        28.350017         40.64279
#>   Gurnard         7.859118        13.189341        20.272820         29.38406
#>   Plaice          3.691813         6.152248         9.428694         13.71200
#>   Haddock        41.681153        75.533931       127.039640        191.37526
#>   Cod           339.641484       888.810624      1782.807473       2684.96457
#>   Saithe         61.052542       131.320369       259.718839        471.97241
#>          Age
#> Species   3.91836734693878 4.40816326530612 4.89795918367347 5.38775510204082
#>   Sprat           20.48712         21.87369         23.10620         24.20218
#>   Sandeel         22.42790         24.54445         26.36416         27.92167
#>   N.pout          68.12609         74.64142         79.98381         84.33318
#>   Herring        151.60178        165.94647        179.18552        191.54104
#>   Dab             70.65271         85.88550        101.89349        118.33365
#>   Whiting        205.27814        264.20400        326.03497        388.95689
#>   Sole            55.95754         73.55296         89.96107        104.92886
#>   Gurnard         39.72877         49.38461         58.93359         69.05629
#>   Plaice          19.27312         26.48034         35.80904         47.86478
#>   Haddock        257.24129        332.73859        419.77292        518.09542
#>   Cod           3696.51300       4821.93378       6035.45789       7312.08470
#>   Saithe         794.57009       1201.38271       1616.36118       2077.82155
#>          Age
#> Species   5.87755102040816 6.36734693877551 6.85714285714286 7.3469387755102
#>   Sprat           25.18015         26.05361         26.83371        27.53045
#>   Sandeel         29.24867         30.36271         31.29699        32.08053
#>   N.pout          87.75897         90.44604         92.55356        94.20656
#>   Herring        203.06648        213.77401        223.71259       232.93748
#>   Dab            134.85886        151.18645        167.02691       182.15862
#>   Whiting        451.49469        512.52336        571.12689       626.76737
#>   Sole           119.89638        135.42275        151.67915       168.68423
#>   Gurnard         79.92848         91.57662        103.98164       117.09977
#>   Plaice          63.38800         83.05297        105.72865       128.60288
#>   Haddock        626.93110        745.14975        871.39728      1004.14316
#>   Cod           8629.78741       9969.36965      11314.12130     12652.18160
#>   Saithe        2596.20385       3168.33390       3789.06737      4452.82280
#>          Age
#> Species   7.83673469387755 8.3265306122449 8.81632653061224 9.30612244897959
#>   Sprat           28.15274        28.70970         29.21179         29.66483
#>   Sandeel         32.73761        33.28866         33.75079         34.15093
#>   N.pout          95.50303        96.51985         97.31746         97.94882
#>   Herring        241.48705       249.35624        256.59319        263.24873
#>   Dab            196.44177       209.79279        222.07020        233.35754
#>   Whiting        679.05561       727.83215        772.86083        814.42343
#>   Sole           186.37928       204.72549        223.60753        242.98817
#>   Gurnard        130.88098       145.27598        160.19209        175.58481
#>   Plaice         152.24368       177.87631        205.93054        236.48420
#>   Haddock       1141.76380      1282.84031       1425.89599       1569.32089
#>   Cod          13972.27420     15265.14496      16525.78746      17748.32615
#>   Saithe        5154.03762      5887.23359       6646.96871       7429.18604
#>          Age
#> Species   9.79591836734694 10.2857142857143 10.7755102040816 11.265306122449
#>   Sprat           30.07363         30.44249         30.77532        31.07564
#>   Sandeel         34.52578         34.87813         35.20932        35.52063
#>   N.pout          98.52946         99.09209         99.63727       100.16554
#>   Herring        269.36953        274.99857        280.17535       284.93394
#>   Dab            243.67149        252.92271        261.21017       268.63436
#>   Whiting        852.50029        887.07656        918.47356       946.98344
#>   Sole           262.72494        282.77495        303.02267       323.35867
#>   Gurnard        191.33613        207.40861        223.66357       240.08003
#>   Plaice         269.50889        304.93546        342.62801       382.43749
#>   Haddock       1712.32252       1853.27741       1991.81515      2126.75589
#>   Cod          18928.34936      20066.67709      21157.68448     22201.50139
#>   Saithe        8229.15555       9042.42399       9866.22388     10695.95349
#>          Age
#> Species   11.7551020408163 12.2448979591837 12.734693877551 13.2244897959184
#>   Sprat           31.34663         31.59114        31.81177         32.01085
#>   Sandeel         35.81326         36.08831        36.34686         36.58988
#>   N.pout         100.67743        101.17345       101.65408        102.11980
#>   Herring        289.29526        293.29090       296.95169        300.30568
#>   Dab            275.28505        281.24289       286.59524        291.43233
#>   Whiting        972.87169        996.21303      1017.08997       1035.76582
#>   Sole           343.76957        364.08225       384.24389        404.25409
#>   Gurnard        256.52521        272.93634       289.29484        305.45189
#>   Plaice         424.25845        467.84888       513.11056        559.78349
#>   Haddock       2257.45049       2383.98237      2505.33120       2621.21117
#>   Cod          23200.10496      24153.50807     25059.02934      25918.78314
#>   Saithe       11529.46662      12363.70932     13195.11256      14023.41636
#>          Age
#> Species   13.7142857142857 14.2040816326531 14.6938775510204 15.1836734693878
#>   Sprat           32.19048         32.35257         32.49882         32.63079
#>   Sandeel         36.81831         37.03304         37.23487         37.42458
#>   N.pout         102.57108        103.00836        103.43209        103.84267
#>   Herring        303.37858        306.19396        308.77339        311.13665
#>   Dab            295.80502        299.75790        303.33128        306.56161
#>   Whiting       1052.47220       1067.41659       1080.78460       1092.74239
#>   Sole           423.97233        443.30903        462.27163        480.86654
#>   Gurnard        321.37660        337.07203        352.44534        367.40973
#>   Plaice         607.77205        656.78861        706.78900        757.44965
#>   Haddock       2731.86874       2837.49382       2937.39277       3031.51218
#>   Cod          26735.08193      27510.12209      28245.67173      28940.42329
#>   Saithe       14844.29576      15656.13648      16459.03905      17250.22352
#>          Age
#> Species   15.6734693877551 16.1632653061224 16.6530612244898 17.1428571428571
#>   Sprat           32.74987         32.85731         32.95426         33.04174
#>   Sandeel         37.60291         37.77053         37.92809         38.07619
#>   N.pout         104.24051        104.62602        104.99957        105.36154
#>   Herring        313.30185        315.28560        317.10310        318.76828
#>   Dab            309.48172        312.12143        314.50767        316.66478
#>   Whiting       1103.43887       1113.00707       1121.56596       1129.22205
#>   Sole           498.98019        516.52780        533.52709        549.99518
#>   Gurnard        381.97626        396.15550        409.93004        423.18846
#>   Plaice         808.71389        860.42222        912.30907        964.37021
#>   Haddock       3120.19312       3203.74969       3282.47683       3356.65363
#>   Cod          29595.69899      30213.77447      30796.76699      31346.66017
#>   Saithe       18026.93014      18789.42288      19537.96100      20270.90430
#>          Age
#> Species   17.6326530612245 18.1224489795918 18.6122448979592 19.1020408163265
#>   Sprat           33.12068         33.19190         33.25617         33.31416
#>   Sandeel         38.21540         38.34625         38.46924         38.58486
#>   N.pout         105.71228        106.05214        106.38147        106.70057
#>   Herring        320.29391        321.69168        322.97231        324.14561
#>   Dab            318.61475        320.37749        321.97097        323.41144
#>   Whiting       1136.07059       1142.19673       1147.67669       1152.57862
#>   Sole           565.94869        581.37359        596.15203        610.29519
#>   Gurnard        435.93525        448.19023        459.97230        471.29973
#>   Plaice        1016.46289       1068.32179       1119.94175       1171.31520
#>   Haddock       3426.08617       3490.67628       3550.76325       3606.66719
#>   Cod          31865.33859      32354.56965      32816.02360      33251.27841
#>   Saithe       20985.34766      21681.69697      22360.41059      23021.93518
#>          Age
#> Species   19.5918367346939 20.0816326530612 20.5714285714286 21.0612244897959
#>   Sprat           33.36649         33.41370         33.45630         33.49474
#>   Sandeel         38.69353         38.79568         38.89169         38.98195
#>   N.pout         107.00979        107.30941        107.59974        107.88106
#>   Herring        325.22059        326.20547        327.10782        327.93455
#>   Dab            324.71359        325.89071        326.95481        327.91672
#>   Whiting       1156.96349       1160.88584       1164.51489       1168.08128
#>   Sole           623.83023        636.78328        649.17945        661.04259
#>   Gurnard        482.18679        492.56620        502.42371        511.78559
#>   Plaice        1222.19765       1272.46949       1322.13813       1371.21091
#>   Haddock       3658.67652       3707.06326       3752.07941       3793.95926
#>   Cod          33661.32121      34046.01253      34406.79624      34745.15857
#>   Saithe       23666.45080      24291.25858      24896.04205      25481.44313
#>          Age
#> Species   21.5510204081633 22.0408163265306 22.530612244898 23.0204081632653
#>   Sprat           33.52943         33.56073        33.58897         33.61445
#>   Sandeel         39.06678         39.14652        39.22148         39.29193
#>   N.pout         108.15366        108.41781       108.67376        108.92178
#>   Herring        328.69199        329.38595       330.02175        330.60427
#>   Dab            328.78627        329.57233       330.28291        330.92526
#>   Whiting       1171.59025       1175.04273      1178.43962       1181.78182
#>   Sole           672.39562        683.25822       693.59554        703.37001
#>   Gurnard        520.67675        529.12086       537.14040        544.75671
#>   Plaice        1419.61194       1467.11891      1513.73134       1559.46608
#>   Haddock       3832.92145       3869.16931      3902.89190       3934.26515
#>   Cod          35062.49299      35360.10537     35639.22189      35900.99201
#>   Saithe       26048.08313      26596.56301     27127.46491      27641.35217
#>          Age
#> Species   23.5102040816327          24
#>   Sprat           33.63744    33.65819
#>   Sandeel         39.35816    39.42041
#>   N.pout         109.16210   109.39497
#>   Herring        331.13797   331.62695
#>   Dab            331.50593   332.03084
#>   Whiting       1185.07021  1188.30565
#>   Sole           712.61201   721.35057
#>   Gurnard        551.99007   558.85973
#>   Plaice        1604.33965  1648.36829
#>   Haddock       3963.45277  3990.60700
#>   Cod          36146.49364 36376.73795
#>   Saithe       28138.42078 28617.43936