Get default value for h

Sets h so that the species reaches maturity size w_mat at the maturity age age_mat if it feeds at feeding level f0.

Usage

get_h_default(params)

Arguments

params

A MizerParams object or a species parameter data frame

Value

A vector with the values of h for all species

Details

If age_mat is missing in the species parameter data frame, then it is calculated from the von Bertalanffy growth curve parameters k_vb and (optionally t0) taken from the species parameter data frame. This is not reliable and a warning is issued.

If no growth information is given at all for a species, the default is set to h = 30.

See also