Initial fishing effort

The fishing effort is a named vector, specifying for each fishing gear the effort invested into fishing with that gear. The effort value for each gear is multiplied by the catchability and the selectivity to determine the fishing mortality imposed by that gear, see setFishing() for more details.

A valid effort vector is a named vector with one effort value for each gear.

Usage

initial_effort(params)

initial_effort(params) <- value

validEffortVector(effort, params)

Arguments

params

A MizerParams object

value

The initial fishing effort

effort

A vector or scalar.

Value

Effort vector

Details

The initial effort you have set can be overruled when running a simulation by providing an effort argument to project() which allows you to specify a time-varying effort.

The function also accepts an effort that is not yet valid:

a scalar, which is then replicated for each gear

an unnamed vector, which is then assumed to be in the same order as the gears in the params object

a named vector in which the gear names have a different order than in the params object. This is then sorted correctly.

a named vector which only supplies values for some of the gears. The effort for the other gears is then set to zero.

An effort argument will lead to an error if it is either

unnamed and of the wrong length

named but where some names do not match any of the gears

not numeric