mizer: Multi-species size-based modelling in R

The mizer package implements multi-species size-based modelling in R. It has been designed for modelling marine ecosystems.

Details

Using mizer is relatively simple. There are three main stages:

Setting the model parameters. This is done by creating an object of class MizerParams. This includes model parameters such as the life history parameters of each species, and the range of the size spectrum. There are several setup functions that help to create a MizerParams objects for particular types of models:

newSingleSpeciesParams()

newCommunityParams()

newTraitParams()

newMultispeciesParams()

Running a simulation. This is done by calling the project() function with the model parameters. This produces an object of MizerSim that contains the results of the simulation.

Exploring results. After a simulation has been run, the results can be explored using a range of plotting_functions, summary_functions and indicator_functions.

See the mizer website for full details of the principles behind mizer and how the package can be used to perform size-based modelling.

See also

Author

Maintainer: Gustav Delius gustav.delius@york.ac.uk (ORCID) [copyright holder]

Authors:

Finlay Scott drfinlayscott@gmail.com [copyright holder]

Julia Blanchard julia.blanchard@utas.edu.au (ORCID) [copyright holder]

Ken Andersen kha@aqua.dtu.dk (ORCID) [copyright holder]

Other contributors:

Richard Southwell richard.southwell@york.ac.uk [contributor, copyright holder]