Get feeding level needed to project standard mizer model

Get feeding level needed to project standard mizer model

mizerFeedingLevel(params, encounter, ...)

Arguments

params
A MizerParams object

encounter
A two dimensional array (predator species x predator size) with the encounter rate.

...
Unused

Value

A two dimensional array (predator species x predator size) with the feeding level.

Feeding level

The feeding level \(f_i(w)\) is the proportion of its maximum intake rate at which the predator is actually taking in fish. It is calculated from the encounter rate \(E_i\) and the maximum intake rate \(h_i(w)\) as $$f_i(w) = \frac{E_i(w)}{E_i(w)+h_i(w)}.$$ The encounter rate \(E_i\) is passed as an argument or calculated with getEncounter(). The maximum intake rate \(h_i(w)\) is taken from the params object, and is set with setMaxIntakeRate(). As a consequence of the above expression for the feeding level, \(1-f_i(w)\) is the proportion of the food available to it that the predator actually consumes.

The feeding level is used in mizerEReproAndGrowth() and in mizerPredRate().

See also