Get predation rate needed to project standard mizer model

Calculates the potential rate (in units 1/year) at which a prey individual of a given size \(w\) is killed by predators from species \(j\). In formulas $${\tt pred\_rate}_j(w_p) = \int \phi_j(w,w_p) (1-f_j(w)) \gamma_j(w) N_j(w) \, dw.$$ This potential rate is used in the function mizerPredMort() to calculate the realised predation mortality rate on the prey individual.

mizerPredRate(params, n, n_pp, n_other, feeding_level, ...)

Arguments

params
A MizerParams object

n
A matrix of species abundances (species x size).

n_pp
A vector of the resource abundance by size

n_other
A list of abundances for other dynamical components of the ecosystem

feeding_level
The current feeding level (optional). A matrix of size no. species x no. size bins. If not supplied, is calculated internally using the getFeedingLevel() function.

...
Unused

Value

A two dimensional array (predator species x prey size) with the predation rate, where the prey size runs over fish community plus resource spectrum.

See also