Get all rates needed to project standard mizer model

Calls other rate functions in sequence and collects the results in a list.

mizerRates(params, n, n_pp, n_other, t = 0, effort, rates_fns, ...)

Arguments

params
A MizerParams object

n
A matrix of species abundances (species x size).

n_pp
A vector of the resource abundance by size

n_other
A list of abundances for other dynamical components of the ecosystem

t
The current time

effort
The effort for each fishing gear

rates_fns
Named list of the functions to call to calculate the rates. Note that this list holds the functions themselves, not their names.

...
Unused

Details

By default this function returns a list with the following components:

encounter from mizerEncounter()

feeding_level from mizerFeedingLevel()

pred_rate from mizerPredRate()

pred_mort from mizerPredMort()

f_mort from mizerFMort()

mort from mizerMort()

resource_mort from mizerResourceMort()

e from mizerEReproAndGrowth()

e_repro from mizerERepro()

e_growth from mizerEGrowth()

rdi from mizerRDI()

rdd from BevertonHoltRDD()

However you can replace any of these rate functions by your own rate function if you wish, see setRateFunction() for details.

See also