Construct MizerParams object for multispecies model

Provides default functional forms for all slots in the MizerParams object based on user-provided species parameters.

multispeciesParams(object, interaction, min_w = 0.001, max_w = NA,
  no_w = 100, min_w_pp = NA, no_w_pp = NA, n = 2/3, p = 0.7,
  q = 0.8, r_pp = 10, kappa = 1e+11, lambda = (2 + q - n),
  w_pp_cutoff = 10, f0 = 0.6, z0pre = 0.6, z0exp = n - 1,
  store_kernel = (no_w <= 100),
  plankton_dynamics = plankton_semichemostat, interaction_p = rep(1,
  nrow(object)), rho = NULL, resource_dynamics = list(),
  resource_params = list())

Arguments

object
A data frame of species specific parameter values (see notes below).

interaction
Optional argument to specify the interaction matrix of the species (predator by prey). If missing a default interaction is used where all interactions between species are set to 1. Note that any dimnames of the interaction matrix argument are ignored by the constructor. The dimnames of the interaction matrix in the returned MizerParams object are taken from the species names in the species_params slot. This means that the order of the columns and rows of the interaction matrix argument should be the same as the species name in the species_params slot.

min_w
The smallest size of the community spectrum. By default his is set to the smallest w_min given in the species_params data frame, or to 0.001g if the w_min slot is missing.

max_w
The largest size of the community spectrum. By default this is set to the largest w_inf specified in the species_params data frame.

no_w
The number of size bins in the community spectrum.

min_w_pp
The smallest size of the plankton spectrum. By default this is set so the smallest value at which any of the consumers can feed.

no_w_pp
Obsolete argument that is no longer used because the number of plankton size bins is determined because all size bins have to be logarithmically equally spaced.

n
Scaling of the intake. Default value is 2/3.

p
Scaling of the standard metabolism. Default value is 0.7.

q
Exponent of the search volume. Default value is 0.8.

r_pp
Growth rate of the primary productivity. Default value is 10.

kappa
Carrying capacity of the plankton spectrum. Default value is 1e11.

lambda
Exponent of the plankton spectrum. Default value is (2+q-n).

w_pp_cutoff
The cut off size of the plankton spectrum. Default value is 10.

f0
Average feeding level. Used to calculated h and gamma if those are not columns in the species data frame. Also requires k_vb (the von Bertalanffy K parameter) to be a column in the species data frame. If h and gamma are supplied then this argument is ignored. Default is 0.6.

z0pre
If z0, the mortality from other sources, is not a column in the species data frame, it is calculated as z0pre * w_inf ^ z0exp. Default value is 0.6.

z0exp
If z0, the mortality from other sources, is not a column in the species data frame, it is calculated as z0pre * w_inf ^ z0exp. Default value is n-1.

store_kernel
A boolean flag that determines whether the full feeding kernel is stored. If FALSE, only its Fourier transforms are stored. The default is TRUE if the number of size bins is no larger than 100 and FALSE otherwise.

plankton_dynamics
Function that determines plankton dynamics by calculating the plankton spectrum at the next time step from the current state. The default is "plankton_semichemostat".

interaction_p
Vector specifying for each species its interaction with plankton, similar to what the interaction matrix does for the interaction with other species. Entries should be numbers between 0 and 1. The default is a vector of 1s.

rho
Either NULL (default), in which case there will be no unstructured resources in the model, or an array (species x resource) that gives the rate \(\rho_{id}\) that determines the rate at which species \(i\) encounters biomass of resource \(d\). The rate is assumed to scale allometrically with size with exponent n. So the total contribution of the resources to the rate at which the individual encounters biomass is $$\sum_d\rho_{id} w^n B_d,$$ where \(B_d\) is the biomass of the d-th unstructured resource component. See resource_dynamics help for more details.

resource_dynamics
A named list of functions that determine the dynamics of the unstructured resources by calculating their biomasses at the next time step from the current state. See resource_dynamics for details. Ignored if rho is NULL.

resource_params
A list of parameters needed by the resource_dynamics functions. Ignored if rho is NULL.

...
Additional arguments.

Value

An object of type MizerParams

Note

The only essential argument is a data frame which contains the species data. The data frame is arranged species by parameter, so each column of the parameter data frame is a parameter and each row has the parameters for one of the species in the model.

There are some essential columns that must be included in the species parameter data.frame and that do not have default values. Other columns do have default values, so that if they are not included in the species parameter data frame, they will be automatically added when the MizerParams object is created. See the accompanying vignette for details of these columns.

The following table gives different components of a species parameters data.frame.

Column name	Description	Default value

species	Name of the species	Compulsory (no default)

w_inf	The asymptotic mass of the species	Compulsory (no default)

w_mat	The maturation mass of the species	Compulsory (no default)

beta	Preferred predator prey mass ratio	Compulsory (no default)

sigma	Width of prey size preference	Compulsory (no default)

h	Maximum food intake rate. If this is not provided, it is calculated using the k_vb column. Therefore, either h or k_vb must be provided.	Optional (no default)

k_vb	The von Bertalanffy K parameter. Only used to calculate h if that column is not provided	Optional (no default)

gamma	Volumetric search rate. If this is not provided, it is calculated using the h column and other parameters.	Optional (no default)

ks	Standard metabolism coefficient	h*0.2

z0	Background mortality (constant for all sizes). If this is not provided then z0 is calculated as z0pre ∗ w_inf^z0exp. Also z0pre and z0exp have default values of 0.6 and -1/3 respectively.	Optional (no default)

k	Activity coefficient	0

alpha	Assimilation efficiency	0.6

erepro	Reproductive efficiency	1

w_min	The size class that recruits are placed in.	smallest size class of the species size spectrum

See also

Examples

data(NS_species_params_gears)
data(inter)
params <- multispeciesParams(NS_species_params_gears, inter)

#> Note: No sel_func column in species data frame. Setting selectivity to be 'knife_edge' for all species.

#> Note: No knife_edge_size column in species data frame. Setting knife edge selectivity equal to w_mat.

#> Note: No h provided for some species, so using f0 and k_vb to calculate it.

#> Note: No gamma provided for some species, so using f0, h, beta, sigma, lambda and kappa to calculate it.

#> Note: Using z0 = z0pre * w_inf ^ z0exp for missing z0 values.

#> Note: No ks column in species data frame so using ks = h * 0.2.