Set up parameters for a single-species in a Sheldon power-law background

This functions creates a MizerParams object with a single species. This species is embedded in a fixed power-law community spectrum $$N_c(w) = \kappa w^{-\lambda}$$ This community provides the food income for the species. Cannibalism is switched off. The predation mortality arises only from the predators in the power-law community and it is assumed that the predators in the community have the same feeding parameters as the foreground species. The function has many arguments, all of which have default values.

newSheldonParams(
  w_inf = 100,
  w_min = 0.001,
  eta = 10^(-0.6),
  w_mat = w_inf * eta,
  no_w = log10(w_inf/w_min) * 50 + 1,
  n = 3/4,
  p = n,
  lambda = 2.05,
  kappa = 0.005,
  alpha = 0.4,
  ks = 4,
  k_vb = 1,
  beta = 100,
  sigma = 1.3,
  f0 = 0.6,
  gamma = NA,
  bmort_prop = 0,
  starv_coef = 10,
  rfac = 4,
  ...
)

Arguments

w_inf
Asymptotic size of species

w_min
Egg size of species

eta
Ratio between maturity size w_mat and asymptotic size w_inf. Default is 10^(-0.6), approximately 1/4.. Ignored if w_mat is supplied explicitly.

w_mat
Maturity size of species. Default value is eta * w_inf.

no_w
The number of size bins in the community spectrum. These bins will be equally spaced on a logarithmic scale. Default value is such that there are 50 bins for each factor of 10 in weight.

n
Scaling exponent of the maximum intake rate. Default value is 2/3.

p
Scaling exponent of the standard metabolic rate. By default this is equal to the exponent n.

lambda
Exponent of the abundance power law.

kappa
Coefficient in abundance power law. Default value is 0.005.

alpha
The assimilation efficiency of the community. The default value is 0.4.

ks
Standard metabolism coefficient. Default value is 4.

beta
Preferred predator prey mass ratio. Default value is 100.

sigma
Width of prey size preference. Default value is 1.3.

f0
Expected average feeding level. Used to set gamma, the coefficient in the search rate. The default value is 0.6. Ignored if gamma is given explicitly.

gamma
Volumetric search rate. If not provided, default is determined by get_gamma_default using the value of f0.

bmort_prop
The proportion of the total mortality that comes from background mortality, i.e., from sources other than predation or fishing. A number in the interval [0, 1). Default 0.

starv_coef
Proportionality constant for starvation mortality. When starv_coef is equal to 10 the instantaneous starvation mortality (1/year) is 1 when energy deficit is 10 is no starvation mortality

rfac
The factor such that R_max = rfac * R, where R_max is the maximum recruitment allowed and R is the steady-state recruitment. Thus the larger rfac the less the impact of the non-linear stock-recruitment curve. The default is 4.

...
Other arguments to pass to the MizerParams constructor.

Value

An object of type MizerParams

Details

In addition to setting up the parameters, this function also sets up an initial condition that is close to steady state, under the assumption of no fishing.

Although the steady state is often stable without imposing a stock recruitment relationship, the function can set a Beverton-Holt type stock recruitment relationship that imposes a maximal reproduction rate that is a multiple of the recruitment rate at steady state. That multiple is set by the argument rfac.

See also

Examples

if (FALSE) {
params <- newTraitParams()
sim <- project(params, t_max = 5, effort = 0)
plotSpectra(sim)
}