Create a hover-enabled plotly plot from a mizer object

Creates an interactive plotly version of a mizer plot. Can be called on any mizer array object (such as those returned by getEncounter(), getBiomass(), etc.) or on any mizer_plot object returned by the named plot functions such as plotBiomass(), plotSpectra(), etc.

Usage

# S3 method for class 'ArrayTimeBySpecies'
plotHover(x, ...)

plotHover(x, ...)

Arguments

x

A mizer_plot, ArraySpeciesBySize, ArrayTimeBySpecies, or ArrayTimeBySpeciesBySize object.

...

Arguments passed to the corresponding plot() method for mizer array objects, or to plotly::ggplotly() for mizer_plot objects.

tooltip

Character vector of aesthetic names to include in the tooltip, or "all" for all aesthetics. Defaults to the tooltip stored in the mizer_plot object.

Value

A plotly object.

See also

Examples

# \donttest{
plotHover(getEncounter(NS_params))

# }
# \donttest{
plotHover(getBiomass(NS_sim))

# }
# \donttest{
plotHover(getFMort(NS_sim))

# }