Methods used for projecting

The methods defined in the file project_methods calculate the various quantities needed to project the size-spectra forward in time, using the model described in section 3 of the mizer vignette.

List of Methods

In this list we relate the methods in this file to the quantities named in the mizer vignette.

Method name	Expression	Description	Section in vignette

getPhiPrey	\(E_{a.i}(w)\)	Available energy	3.2

getFeedingLevel	\(f_i(w)\)	Feeding level	3.3

getPredRate	\(\phi_i(w_p/w) (1-f_i(w)) \gamma_i w^q N_i(w) dw\)	Predation	3.7

getM2	\(\mu_{p.i}(w)\)	Predation mortality	3.7

getM2Background	\(\mu_{p}(w)\)	Predation mortality on background	3.8

getFMortGear	\(F_{g,i}(w)\)	Fishing mortality by gear	8.3

getFMort	\(\mu_{f.i}(w)\)	Total fishing mortality	8.3

getZ	\(\mu_{i}(w)\)	Total mortality	3.7

getEReproAndGrowth	\(E_{r.i}(w)\)	Energy put into growth and reproduction	3.4

getESpawning	\(\psi_i(w)E_{r.i}(w)\)	Energy put reproduction	3.5

getEGrowth	\(g_i(w)\)	Energy put growth	3.4

getRDI	\(R_{p.i}\)	Egg production	3.5

getRDD	\(R_i\)	Recruitment	3.6