Set background mortality rate

Set background mortality rate

setBMort(params, mu_b = NULL, z0pre = 0.6, z0exp = -1/4)

Arguments

params
MizerParams

mu_b
Optional. An array (species x size) holding the background mortality rate.

z0pre
If z0, the mortality from other sources, is not a column in the species data frame, it is calculated as z0pre * w_inf ^ z0exp. Default value is 0.6.

z0exp
If z0, the mortality from other sources, is not a column in the species data frame, it is calculated as z0pre * w_inf ^ z0exp. Default value is n-1.

Value

MizerParams

Setting background mortality rate

The background mortality is all the mortality that is not due to either predation or fishing. It is a rate with units 1/year.

The mu_b argument allows you to specify a background mortality rate that depends on species and body size. You can see an example of this in the Examples section of the help page for setBMort.

If the mu_b argument is not supplied, then the background mortality is assumed to depend only on the asymptotic size of the species, not on the size of the individual: \(\mu_{b.i}(w) = z_{0.i}\). The value of the constant \(z_0\) for each species is taken from the z0 column of the species_params data frame, if that column exists. Otherwise it is calculated as $$z_{0.i} = {\tt z0pre}_i\, w_{inf}^{\tt z0exp}.$$

See also

Examples

if (FALSE) {
data("NS_species_params")
params <- newMultispeciesParams(NS_species_params)

#### Setting allometric death rate #######################

# Set coefficient for each species. Here we choose 0.1 for each species
mu_b0 <- rep(0.1, nrow(params@species_params))

# Multiply by power of size with exponent, here chosen to be -1/4
# The outer() function makes it an array species x size
mu_b <- outer(mu_b0, params@w^(-1/4))

# Change the background death in the params object
params <- setBMort(params, mu_b = mu_b)
}