Set or change model parameters

Passes its arguments to the the setup functions

setPredKernel

setSearchVolume

setInteraction

setIntakeMax

setMetab

setBMort

setReproduction

setPlankton

setResources

setResourceEncounter

setFishing

setParams(params, interaction = NULL, pred_kernel = NULL,
  pred_kernel_type = params@pred_kernel_type, store_kernel = FALSE,
  search_vol = NULL, q = params@q, f0 = params@f0,
  intake_max = NULL, n = params@n, metab = NULL, p = params@p,
  z0pre = 0.6, z0exp = n - 1, psi = NULL, kappa = params@kappa,
  lambda = params@lambda, r_pp = 10, w_pp_cutoff = 10,
  plankton_dynamics = params@plankton_dynamics,
  resource_dynamics = params@resource_dynamics,
  resource_params = params@resource_params, rho = NULL)

Arguments

params
A MizerParams object

interaction
Interaction matrix of the species (predator by prey). Entries should be numbers between 0 and 1. See "Setting interactions" section below.

pred_kernel
Optional. An array (species x predator size x prey size) that holds the predation coefficient of each predator at size on each prey size. The dimensions are thus no_sp, no_w, no_w_full.

pred_kernel_type
Only used if pred_kernel is not supplied. A string or vector of strings that determines which predation kernel function is used for each species. If not supplied, is taken from species parameter dataframe if possible, or defaults to "lognormal".

store_kernel
Only used if pred_kernel is not supplied. A boolean flag that determines whether the full three dimensional predation kernel array is calculated and stored, even though it is not needed by mizer. Only useful if you have your own code that relies on this array. Default is FALSE.

q
Exponent of the allometric search volume. Only needed if gamma is not an array.

f0
Average feeding level. Used to calculated h and gamma if those are not columns in the species data frame. Also requires k_vb (the von Bertalanffy K parameter) to be a column in the species data frame. If h and gamma are supplied then this argument is ignored. Default is 0.6.

n
Scaling exponent of the intake rate.

metab
Optional. An array (species x size) holding the metabolic rate for each species at size. If not supplied, a default is set as described in the section "Setting metabolic rate".

p
Scaling exponent of the standard metabolic rate.

z0pre
If z0, the mortality from other sources, is not a column in the species data frame, it is calculated as z0pre * w_inf ^ z0exp. Default value is 0.6.

z0exp
If z0, the mortality from other sources, is not a column in the species data frame, it is calculated as z0pre * w_inf ^ z0exp. Default value is n-1.

psi
Optional. An array (species x size) that holds the allocation to reproduction for each species at size. If not supplied, a default is set as described in the section "Setting reproduction".

kappa
Carrying capacity of the plankton spectrum.

lambda
Exponent of the plankton spectrum.

r_pp
Growth rate of the primary productivity. Default is 10 g/year.

w_pp_cutoff
The upper cut off size of the plankton spectrum. Default is 10 g.

plankton_dynamics
Function that determines plankton dynamics by calculating the plankton spectrum at the next time step from the current state. The default is "plankton_semichemostat".

resource_dynamics
A named list of functions that determine the dynamics of the unstructured resources by calculating their biomasses at the next time step from the current state. See resource_dynamics for details. An empty list if the model does not have unstructured resources.

resource_params
A list of parameters needed by the resource_dynamics functions. An empty list if no parameters are needed.

rho
Either a 3-dim array (predator species x resource x predator size) or a 2-dim array (predator species x resource). In the latter case the size dependence is given by allometry.

Value

A MizerParams object

Setting interactions

The species interaction matrix \(\theta_{ij}\) is used when calculating the food encounter rate in getEncounter and the predation rate in getPredRate. This function checks that the supplied interaction matrix is valid and then stores it in the interaction slot of the params object before returning that object.

Any dimnames of the interaction matrix argument are ignored by the constructor. The dimnames of the interaction matrix in the returned MizerParams object are taken from the species names in the species_params slot. This means that the order of the columns and rows of the interaction matrix argument should be the same as the species name in the species_params slot.

Setting predation kernel

Kernel dependent on predator to prey size ratio
If the pred_kernel argument is not supplied, then this function sets a predation kernel that depends only on the ratio of predator mass to prey mass, not on the two masses independently. The shape of that kernel is then determined by the pred_kernel_type argument. The default pred_kernel_type is "lognormal". This will call the function lognormal_pred_kernel to calculate the predation kernel and the function lognormal_max_ppmr to return the maximal predator/prey mass ratio for each species. An alternative pred_kernel type is "box", implemented by the functions box_pred_kernel and box_max_ppmr. You can use any other string as the type. If for example you choose "my" then you need to define a function my_pred_kernel that you can model on the existing functions like lognormal_pred_kernel. You can also define a function my_max_ppmr if you want the smallest plankton size to be calculated automatically by set_multispecies_model, otherwise you need to specify the min_w_pp argument explicitly. When using a kernel dependent on the predator/prey size ratio only, mizer does not need to store the entire three dimensional array in the MizerParams object. Such an array can be very big when there is a large number of size bins. Instead mizer only needs to store two two-dimensional arrays that hold Fourier transforms of the feeding kernel function that allow the encounter rate and the predation rate to be calculated very efficiently. However mizer gives you the option of storing the full three-dimensional array anyway by setting store_kernel = TRUE. This might be useful if you have code for analysing the results of a mizer simulation that relies on the full array.
Kernel dependent on both predator and prey size
If you want to work with a feeding kernel that depends on predator mass and prey mass independently, you can specify the full feeding kernel as a three-dimensional array (predator species x predator size x prey size). You should use this option only if a kernel dependent only on the predator/prey mass ratio is not appropriate. Using a kernel dependent on predator/prey mass ratio only allows mizer to use fast Fourier transform methods to significantly reduce the running time of simulations. The order of the predator species in pred_kernel should be the same as the order in the species params dataframe. If you supply a named array then the function will check the order and warn if it is different.
Setting search volume

The search volume \(\gamma_i(w)\) of an individual of species \(i\) and weight \(w\) multiplies the predation kernel when calculating the encounter rate and the predation rate.

If the gamma argument is not supplied, then it is set to $$\gamma_i(w) = \gamma_i w^q.$$ The values of \(gamma_i\) are taken from the gamma column in the species parameter dataframe. If the gamma column is not supplied in the species parameter dataframe, it is calculated from f0, h, beta, sigma, lambda and kappa.

Setting maximum intake rate

The maximum intake rate \(h_i(w)\) of an individual of species \(i\) and weight \(w\) determines the feeding level, calculated with getFeedingLevel.

If the h argument is not supplied, then it is set to $$h_i(w) = h_i w^n.$$ The values of \(h_i\) are taken from the h column in the species parameter dataframe. If the h column is not supplied in the species parameter dataframe, it is calculated from f0 and the k_vb column, if they are supplied.

If \(h_i\) is set to Inf, fish will consume all encountered food.

Setting metabolic rate

To be documented

Setting background mortality rate

Still need to document

Setting reproduction

If the psi argument is not supplied, the proportion is set to the product of a sigmoidal maturity ogive that gives the proportion of individuals of a given species and size that are mature, and a factor that describes how investment into reproduction by mature individuals scales with size. In formulas: $$\psi(w) = \left[1+\left(\frac{w}{w_{mat}}\right)^{-U}\right]^{-1} \left(\frac{w}{w_{inf}}\right)^{m-n}.$$ Here \(n\) is the scaling exponent of the energy income rate. The exponent \(m\) determines the scaling of the investment into reproduction for mature individuals. By default it is chosen to be \(m = 1\) so that the rate at which energy is invested into reproduction scales linearly with the size. This default can be overridden by including a column m in the species parameter dataframe.

The exponent \(U\) determines the steepness of the maturity ogive. By default it is chosen as \(U = 10\), however this can be overridden by including a column w_mat25 in the species parameter dataframe that specifies the weight at which 25% of individuals are mature, which sets $$U = \frac{\log(3)}{\log(w_{mat} / w_{25})}$$

Setting plankton dynamics

Still need to document

Setting resource dynamics

Still need to document

Setting resource encounter rate

The resource encounter rate \(\rho_{id}(w)\) determines the rate at which an individual of species \(i\) encounters biomass of resource \(d\), $$\sum_d\rho_{id}(w) B_d,$$ where \(B_d\) is the biomass of the d-th unstructured resource component. See resource_dynamics for more details.

If only a two-dimensional array (species x resource) is given for the rho argument, then the size dependence is set to a power-law: $$\rho_{id}(w) = \rho_{id} w^n.$$

The ordering of the entries in the array rho is important. The order of the species in the first array dimension needs to be the same as that in the species parameter dataframe. The order of the resources in the second array dimension must be the same as in the list of resource dynamics. The third dimension, if given, is the size dimension.

Setting fishing

Needs to be documented