Set resource dynamics

Set resource dynamics

setResourceDynamics(params, resource_dynamics = NULL,
  resource_params = NULL)

Arguments

params
A MizerParams object

resource_dynamics
A named list of functions that determine the dynamics of the unstructured resources by calculating their biomasses at the next time step from the current state. An empty list if the model does not have unstructured resources.

resource_params
A named list of parameters needed by the resource_dynamics functions. An empty list if no parameters are needed.

Value

A MizerParams object

Setting resource dynamics

Besides the size-structured planktonic resource, mizer can also model any number of unstructured resource components. Such unstructured components are appropriate whenever the predation on these componente is not size based. Examples include detritus as a resource for detritovores, carrion as a resource for scavengers, or macroflora on which fish can graze.

During a simulation using project(), the biomasses of the resources are updated at each time step by calling the functions specified in the resource_dynamics list, which has one named entry for each unstructured resource component. By default a model is set up without unstructured resource components, so you do not need to provide this list. But if you do want to use it, you can see an example of how to set up a resource_dynamics list in the Examples section of setResourceDynamics().

Mizer provides two functions that you can use to model resource dynamics: detritus_dynamics() and carrion_dynamics(), but you can easily implement others by following those templates. As you can see in the documentation of these functions, their arguments are: the MizerParams object params, the current fish size spectra n, the current plankton spectrum n_pp, the current resource biomasses B and the current rates calculated by the getRates() function.

The other arguments to the resource dynamics functions are model parameters, like for example growth rates. These need to be provided in the resource_params argument which is a named list. One model parameter that should always be present in this list is the rate of change due to external causes. This should be given a name of the form resource_external where resource should be replaced by the name of the resource, see for example detritus_external in detritus_dynamics().

When writing your own resource dynamics functions, you can choose any names for your other model parameters, but you must make sure not to use the same name in the function for another resource component. One way to ensure this is to prefix all parameter namse with your resource name.

The dynamics for a resource should always have a loss term accounting for the consumption of the resource. This should always have the form used in the example function detritus_dynamics(), in order to be in agreement with the feeding by consumers that is set with setResourceEncounter.

See also