Set resource encounter rate

Set resource encounter rate

setResourceEncounter(params, rho = NULL, n = params@n)

Arguments

params
A MizerParams object

rho
Optional. An array (species x resource x size) holding the rate at which a consumer of a particular size feeds on each resource. Described in the section "Setting resource encounter rate".

n
Scaling exponent of the intake rate.

Value

A MizerParams object

Setting resource encounter rate

The resource encounter rate \(\rho_{id}(w)\) (units 1/year) determines the rate at which an individual of species \(i\) encounters biomass of resource \(d\), so that the contribution from all unstructured resources to the total encounter rate is $$E_{u.i}(w) = \sum_d\rho_{id}(w) B_d,$$ where \(B_d\) is the biomass of the d-th unstructured resource component.

Resource consumption is subject to satiation in the same way as other food, so that a consumer only consumes a fraction \(1-f_i(w)\) of the encountered resource biomass, where \(f_i(w)\) is the feeding level.

If the rho array is not supplied, then the resource encounter rate is set to a power law $$\rho_{id}(w) = \rho_{id} w^n.$$ The coefficients \(\rho_{id}\) are parameters in the species_params dataframe. For example if there is a resource called "detritus" then the species_params data frame needs to have a column called rho_detritus and similarly for each other resource.

If the rho array is supplied, the ordering of the entries in the array is important. The order of the species in the first array dimension needs to be the same as that in the species parameter dataframe. The order of the resources in the second array dimension must be the same as in the list of resource dynamics. The third dimension is the size dimension.

See also