Tune params object to be at steady state

This is done by running the dynamics for a specified number of years while keeping recruitment and resources constant to reach steady state. Then the reproductive efficiencies and the external resource rates are retuned to achieve that level of recruitment.

steady(params, effort = 0, t_max = 50, t_per = 2, tol = 10^(-2),
  dt = 0.1, progress_bar = TRUE)

Arguments

params
A MizerParams object

effort
The fishing effort. Default is 0

t_max
The maximum number of years to run the simulation. Default is 50.

t_per
The simulation is broken up into shorter runs of t_per years, after each of which we check for convergence. Default value is 2.

tol
The simulation stops when the relative change in the egg production RDI over t_per years is less than tol for every background species. Default value is 1/100.

dt
The time step to use in `project()`

progress_bar
A shiny progress object to implement a progress bar in a shiny app. Default FALSE