Alias for set_multispecies_model

An alias provided for backward compatibility with mizer version <= 1.0

MizerParams(species_params, interaction = matrix(1, nrow =
  nrow(species_params), ncol = nrow(species_params)), no_w = 100,
  min_w = 0.001, max_w = NA, min_w_pp = NA, no_w_pp = NA,
  n = 2/3, p = 0.7, q = 0.8, kappa = 1e+11, lambda = (2 + q - n),
  f0 = 0.6, pred_kernel = NULL, search_vol = NULL,
  intake_max = NULL, metab = NULL, z0pre = 0.6, z0exp = n - 1,
  maturity = NULL, repro_prop = NULL, r_pp = 10, w_pp_cutoff = 10,
  plankton_dynamics = plankton_semichemostat,
  resource_dynamics = list(), resource_params = list(), rho = NULL,
  srr = srrBevertonHolt)

Arguments

species_params
A data frame of species-specific parameter values.

interaction
Interaction matrix of the species (predator by prey). Entries should be numbers between 0 and 1. See "Setting interactions" section below.

no_w
The number of size bins in the consumer spectrum.

min_w
Sets the size of the eggs of all species for which this is not given in the w_min column of the species_params dataframe.

max_w
The largest size of the consumer spectrum. By default this is set to the largest w_inf specified in the species_params data frame.

min_w_pp
The smallest size of the plankton spectrum. By default this is set to the smallest value at which any of the consumers can feed.

no_w_pp
Obsolete argument that is no longer used because the number of plankton size bins is determined because all size bins have to be logarithmically equally spaced.

n
Scaling exponent of the intake rate.

p
Scaling exponent of the standard metabolic rate.

q
Exponent of the allometric search volume. Not needed if a search_vol array is specified.

kappa
Carrying capacity of the plankton spectrum.

lambda
Exponent of the plankton spectrum.

f0
Average feeding level. Used to calculated h and gamma if those are not columns in the species data frame. Also requires k_vb (the von Bertalanffy K parameter) to be a column in the species data frame.

pred_kernel
Optional. An array (species x predator size x prey size) that holds the predation coefficient of each predator at size on each prey size. If not supplied, a default is set as described in section "Setting predation kernel".

search_vol
Optional. An array (species x size) holding the search volume for each species at size. If not supplied, a default is set as described in the section "Setting search volume".

intake_max
Optional. An array (species x size) holding the maximum intake rate for each species at size. If not supplied, a default is set as described in the section "Setting maximum intake rate".

metab
Optional. An array (species x size) holding the metabolic rate for each species at size. If not supplied, a default is set as described in the section "Setting metabolic rate".

z0pre
If z0, the mortality from other sources, is not a column in the species data frame, it is calculated as z0pre * w_inf ^ z0exp. Default value is 0.6.

z0exp
If z0, the mortality from other sources, is not a column in the species data frame, it is calculated as z0pre * w_inf ^ z0exp. Default value is n-1.

maturity
Optional. An array (species x size) that holds the proportion of individuals of each species at size that are mature. If not supplied, a default is set as described in the section "Setting reproduction".

repro_prop
Optional. An array (species x size) that holds the proportion of consumed energy that a mature individual allocates to reproduction for each species at size. If not supplied, a default is set as described in the section "Setting reproduction".

r_pp
Growth rate of the primary productivity. Default is 10 g/year.

w_pp_cutoff
The upper cut off size of the plankton spectrum. Default is 10 g.

plankton_dynamics
Function that determines plankton dynamics by calculating the plankton spectrum at the next time step from the current state. The default is "plankton_semichemostat".

resource_dynamics
A named list of functions that determine the dynamics of the unstructured resources by calculating their biomasses at the next time step from the current state. See resource_dynamics for details. An empty list if the model does not have unstructured resources.

resource_params
A list of parameters needed by the resource_dynamics functions. An empty list if no parameters are needed.

rho
Optional. An array (species x resource x size) holding the rate at which a consumer of a particular size feeds on each resource. Described in the section "Setting resource encounter rate".

srr
The stock recruitment function. Default is srrBevertonHolt.

Value

An object of type MizerParams

Species parameters

The only essential argument is a data frame that contains the species parameters. The data frame is arranged species by parameter, so each column of the parameter data frame is a parameter and each row has the values of the parameters for one of the species in the model.

There are two essential columns that must be included in the species parameter data.frame and that do not have default values: the species column that should hold strings with the names of the species and the w_inf column with the asymptotic sizes of the species.

The species_params dataframe also needs to contain the parameters needed by any predation kernel function or size selectivity function. This will be mentioned in the appropriate sections below.

For all other species parameters, mizer will calculate default values if they are not included in the species parameter data frame. They will be automatically added when the MizerParams object is created. For these parameters you can also specify values for only some species and leave the other entries as NA and the missing values will be set to the defaults.

All the parameters will be mentioned in the following sections.

Setting interactions

The species interaction matrix \(\theta_{ij}\) is used when calculating the food encounter rate in getEncounter and the predation rate in getPredRate. This function checks that the supplied interaction matrix is valid and then stores it in the interaction slot of the params object before returning that object.

Any dimnames of the interaction matrix argument are ignored by the constructor. The dimnames of the interaction matrix in the returned MizerParams object are taken from the species names in the species_params slot. This means that the order of the columns and rows of the interaction matrix argument should be the same as the species name in the species_params slot.

The interaction of the species with the plankton are set via a colun interaction_p in the species_params data frame. Again the entries have to be numbers between 0 and 1. By default this column is set to all 1s.

Setting predation kernel

Kernel dependent on predator to prey size ratio

If the pred_kernel argument is not supplied, then this function sets a predation kernel that depends only on the ratio of predator mass to prey mass, not on the two masses independently. The shape of that kernel is then determined by the pred_kernel_type column in species_params.

The default pred_kernel_type is "lognormal". This will call the function lognormal_pred_kernel to calculate the predation kernel. An alternative pred_kernel type is "box", implemented by the functions box_pred_kernel. These functions require certain species parameters in the species_params data frame. For the lognormal kernel these are beta and sigma, for the box kernely they are ppmr_min and ppmr_max. They are explained in the help pages for the kernel functions. No defaults are set for these parameters. If they are missing from the species_params data frame then mizer will issue an error message.

You can use any other string as the type. If for example you choose "my" then you need to define a function my_pred_kernel that you can model on the existing functions like lognormal_pred_kernel.

When using a kernel that depends on the predator/prey size ratio only, mizer does not need to store the entire three dimensional array in the MizerParams object. Such an array can be very big when there is a large number of size bins. Instead mizer only needs to store two two-dimensional arrays that hold Fourier transforms of the feeding kernel function that allow the encounter rate and the predation rate to be calculated very efficiently. However, if you need the full three-dimensional array you can calculate it with the getPredKernel function.

Kernel dependent on both predator and prey size

If you want to work with a feeding kernel that depends on predator mass and prey mass independently, you can specify the full feeding kernel as a three-dimensional array (predator species x predator size x prey size). The dimensions are thus (no_sp, no_w, no_w_full).

You should use this option only if a kernel dependent only on the predator/prey mass ratio is not appropriate. Using a kernel dependent on predator/prey mass ratio only allows mizer to use fast Fourier transform methods to significantly reduce the running time of simulations.

The order of the predator species in pred_kernel should be the same as the order in the species params dataframe. If you supply a named array then the function will check the order and warn if it is different.

Setting search volume

The search volume \(\gamma_i(w)\) of an individual of species \(i\) and weight \(w\) multiplies the predation kernel when calculating the encounter rate in getEncounter and the predation rate in getPredRate.

If the search_vol argument is not supplied, then the search volume is set to $$\gamma_i(w) = \gamma_i w^q.$$ The values of \(\gamma_i\) are taken from the gamma column in the species parameter dataframe. If the gamma column is not supplied in the species parameter dataframe, a default is calculated by the get_gamma_default function.

Setting maximum intake rate

The maximum intake rate \(h_i(w)\) of an individual of species \(i\) and weight \(w\) determines the feeding level, calculated with getFeedingLevel.

If the intake_max argument is not supplied, then the maximum intake rate is set to $$h_i(w) = h_i w^n.$$ The values of \(h_i\) are taken from the h column in the species parameter dataframe. If the h column is not supplied in the species parameter dataframe, it is calculated by the get_h_default function, using f0 and the k_vb column, if they are supplied.

If \(h_i\) is set to Inf, fish will consume all encountered food.

Setting metabolic rate

The metabolic rate is subtracted from the energy income rate to calculate the rate at which energy is available for growth and reproduction, see getEReproAndGrowth.

If the metab argument is not supplied, then the metabolic rate \(k_i(w)\) for and individual of species \(i\) and size \(w\) is set to $$k_i(w) = ks_i\, w^p + k_i\, w,$$ where \(ks_i w^p\) represents the rate of standard metabolism and \(k_i w\) is the rate at which energy is expended on activity and movement. The values of \(ks_i\) and \(k_i\) are taken from the ks and k columns in the species parameter dataframe. If these parameters are not supplied, the defaults are \(ks_i = 0.2 h_i\) and \(k_i = 0\) where \(h_i\) is the coefficient of the maximum intake rate.

Setting background mortality rate

The background mortality is all the mortality that is not due to either predation or fishing.

If the mu_b argument is not supplied, then the background mortality is assumed to depend only on the asymptotic size of the species, not on the size of the individual: \(\mu_{b.i}(w) = z_{0.i}\). The value of the constant \(z_0\) for each species is taken from the z0 column of the species_params data frame, if that column exists. Otherwise it is calculated as $$z_{0.i} = {\tt z0pre}_i\, w_{inf}^{\tt z0exp}.$$

Setting reproduction

For each species and at each size, the proportion of the available energy that is invested into reproduction is the product of two factors: the proportion maturity of individuals that are mature and the proportion repro_prop of the energy available to a mature individual that is invested into reproduction.

If the maturity argument is not supplied, then it is set to a sigmoidal maturity ogive that changes from 0 to 1 at around the maturity size: $${\tt maturity}(w) = \left[1+\left(\frac{w}{w_{mat}}\right)^{-U}\right]^{-1}.$$ (To avoid clutter, we are not showing the species index in the equations.) The maturity weights are taken from the w_mat column of the species_params data frame. Any missing maturity weights are set to 1/4 of the asymptotic weight in the w_inf column. The exponent \(U\) determines the steepness of the maturity ogive. By default it is chosen as \(U = 10\), however this can be overridden by including a column w_mat25 in the species parameter dataframe that specifies the weight at which 25% of individuals are mature, which sets \(U = \log(3) / \log(w_{mat} / w_{25}).\)

The sigmoidal function given above would strictly reach 1 only asymptotically. Mizer instead sets the function equal to 1 already at the species' maximum size, taken from the compulsory w_inf column in the species_params data frame.

If the repro_prop argument is not supplied, it is set to the allometric form $${\tt repro\_prop}(w) = \left(\frac{w}{w_{inf}}\right)^{m-n}.$$ Here \(n\) is the scaling exponent of the energy income rate. Hence the exponent \(m\) determines the scaling of the investment into reproduction for mature individuals. By default it is chosen to be \(m = 1\) so that the rate at which energy is invested into reproduction scales linearly with the size. This default can be overridden by including a column m in the species parameter dataframe. The asymptotic sizes are taken from the compulsory w_inf column in the species_params data frame.

The reproductive efficiency, i.e., the proportion of energy allocated to reproduction that results in offspring biomass, is set from the erepro column in the species_params data frame. If that is not provided the default is set to 1 (which you will want to override).

Setting plankton dynamics

Still need to document

Setting resource dynamics

Still need to document

Setting resource encounter rate

The resource encounter rate \(\rho_{id}(w)\) determines the rate at which an individual of species \(i\) encounters biomass of resource \(d\), $$\sum_d\rho_{id}(w) B_d,$$ where \(B_d\) is the biomass of the d-th unstructured resource component. See resource_dynamics for more details.

The ordering of the entries in the array rho is important. The order of the species in the first array dimension needs to be the same as that in the species parameter dataframe. The order of the resources in the second array dimension must be the same as in the list of resource dynamics. The third dimension, if given, is the size dimension.

If the rho array is not supplied, then the resource encounter rate is set to a power law $$\rho_{id}(w) = \rho_{id} w^n.$$ The coefficients \(\rho_{id}\) are parameters in the species_params dataframe. For example if there is a resource called "detritus" then the species_params data frame needs to have a column called rho_detritus and similarly for each other resource.

Setting fishing

Needs to be documented

See also

Examples

# NOT RUN {
data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter)
# }