Alias for set_multispecies_model()

An alias provided for backward compatibility with mizer version <= 1.0

Usage

MizerParams(
  species_params,
  interaction = matrix(1, nrow = nrow(species_params), ncol = nrow(species_params)),
  min_w_pp = 1e-10,
  min_w = 0.001,
  max_w = NULL,
  no_w = 100,
  n = 2/3,
  q = 0.8,
  f0 = 0.6,
  kappa = 1e+11,
  lambda = 2 + q - n,
  r_pp = 10,
  ...
)

Arguments

species_params

A data frame of species-specific parameter values.

interaction

Optional interaction matrix of the species (predator species x prey species). By default all entries are 1. See "Setting interaction matrix" section below.

min_w_pp

The smallest size of the resource spectrum. By default this is set to the smallest value at which any of the consumers can feed.

min_w

Sets the size of the eggs of all species for which this is not given in the w_min column of the species_params dataframe.

max_w

The largest size of the consumer spectrum. By default this is set to the largest w_max specified in the species_params data frame.

no_w

The number of size bins in the consumer spectrum.

n

The allometric growth exponent. This can be overruled for individual species by including a n column in the species_params.

q

Allometric exponent of search volume

f0

Expected average feeding level. Used to set gamma, the coefficient in the search rate. Ignored if gamma is given explicitly.

kappa

The coefficient of the initial resource abundance power-law.

lambda

Used to set power-law exponent for resource capacity if the resource_capacity argument is given as a single number.

r_pp

. Use resource_rate argument instead.

...

Unused

Value

A MizerParams object