A class to hold the parameters for a size based model.

MizerParams objects can be created using a range of constructor functions.

Details

Dynamic simulations are performed using the project method on objects of this class.

Slots

w

The size grid for the fish part of the spectrum. An increasing vector of weights (in grams) running from the smallest egg size to the largest asymptotic size.

dw

The spacing in the size grid. So dw[i] = w[i+1] - w[i]. A vector the same length as the w_full slot. The last entry is not determined by the w slot but represents the size of the last size bin.

w_full

The size grid for the full size range including the plankton spectrum. An increasing vector of weights (in grams) running from the smallest plankton size to the largest asymptotic size of fish. The last entries of the vector have to be equal to the content of the w slot.

dw_full

The spacing in the full size grid. So dw_full[i] = w_full[i+1] - w_full[i]. The last entries have to be equal to the content of the dw slot.

w_min_idx

A vector holding the index of the weight of the egg size of each species

psi

An array (species x size) that holds the allocation to reproduction for each species at size, \(\psi_i(w)\)

intake_max

An array (species x size) that holds the maximum intake for each species at size.

search_vol

An array (species x size) that holds the search volume for each species at size.

rho

A 3-dim array (species x resource x size) holding the encounter rates for unstructured resources. See [resource_dynamics] for details.

metab

An array (species x size) that holds the metabolism for each species at size.

mu_b

An array (species x size) that holds the background death \(\mu_{b.i}(w)\)

pred_kernel

An array (species x predator size x prey size) that holds the predation coefficient of each predator at size on each prey size. If this is NA then the following two slots will be used.

ft_pred_kernel_e

An array (species x log of predator/prey size ratio) that holds the Fourier transform of the feeding kernel in a form appropriate for evaluating the encounter rate integral. If this is NA then the pred_kernel will be used to calculate the available energy integral.

ft_pred_kernel_p

An array (species x log of predator/prey size ratio) that holds the Fourier transform of the feeding kernel in a form appropriate for evaluating the predation mortality integral. If this is NA then the pred_kernel will be used to calculate the integral.

rr_pp

A vector the same length as the w_full slot. The size specific growth rate of the plankton spectrum.

cc_pp

A vector the same length as the w_full slot. The size specific carrying capacity of the plankton spectrum.

resource_dynamics

A named list of functions for projecting the biomasses in the unstructured resource components by one timestep. The names of the list entries are the resource names. See resource_dynamics for details.

plankton_dynamics

A function for projecting the plankton abundance density by one timestep. See plankton_semichemostat for an example.

resource_params

A list containing the parameters needed by the `resource_dynamics` functions, see resource_dynamics for details.

sc

The community abundance of the scaling community

species_params

A data.frame to hold the species specific parameters (see the mizer vignette, Table 2, for details)

interaction

The species specific interaction matrix, \(\theta_{ij}\)

interaction_p

The species specific interaction with plankton, \(\theta_{ip}\)

srr

Function to calculate the realised (density dependent) recruitment. Has two arguments which are rdi and species_params

selectivity

An array (gear x species x w) that holds the selectivity of each gear for species and size, \(S_{g,i,w}\)

catchability

An array (gear x species) that holds the catchability of each species by each gear, \(Q_{g,i}\)

initial_n

An array (species x size) that holds abundance of each species at each weight at our candidate steady state solution.

initial_n_pp

A vector the same length as the w_full slot that describes the plankton abundance at each weight.

initial_B

A vector containing the biomasses of the unstructured resource components, see resource_dynamics for details.

n

Exponent of maximum intake rate.

p

Exponent of metabolic cost.

lambda

Exponent of plankton spectrum.

q

Exponent for volumetric search rate.

f0

Initial feeding level.

kappa

Magnitude of plankton spectrum.

A

Abundance multipliers.

linecolour

A named vector of colour values, named by species. Used to give consistent colours to species in plots.

linetype

A named vector of linetypes, named by species. Used to give consistent line types to species in plots.

Note

The MizerParams class is fairly complex with a large number of slots, many of which are multidimensional arrays. The dimensions of these arrays is strictly enforced so that MizerParams objects are consistent in terms of number of species and number of size classes.

Although it is possible to build a MizerParams object by hand it is not recommended and several constructors are available.

The MizerParams class does not hold any dynamic information, e.g. abundances or harvest effort through time. These are held in MizerSim objects.

See also