A class to hold the results of a simulation

A class that holds the results of projecting a MizerParams object through time using project().

Details

A new MizerSim object can be created with the MizerSim() constructor, but you will never have to do that because the object is created automatically by project() when needed.

As a user you should never have to access the slots of a MizerSim object directly. Instead there are a range of functions to extract the information. N() and NPlankton() return arrays with the saved abundances of the species and the plankton population at size respectively. effort() returns the fishing effort of each gear through time. times() returns the vector of times at which simulation results were stored and idxFinalT() returns the index with which to access specifically the value at the final time in the arrays returned by the other functions. Tparams() returns the MizerParams object that was passed in to project(). There are also several summary_functions and plotting_functions available to explore the contents of a MizerSim object.

The arrays all have named dimensions. The names of the time dimension denote the time in years. The names of the w dimension are weights in grams rounded to three significant figures. The names of the sp dimension are the same as the species name in the order specified in the species_params data frame. The names of the gear dimension are the names of the gears, in the same order as specified when setting up the MizerParams object.

Extensions of mizer can use the n_other slot to store the abundances of other ecosystem compoents and these extensions should provide their own functions for accessing that information.

The MizerSim class has changed since previous versions of mizer. To use a MizerSim object created by a previous version, you need to upgrade it with upgradeSim().

Slots

params

An object of type MizerParams.

n

Three-dimensional array (time x species x size) that stores the projected community number densities.

n_pp

An array (time x size) that stores the projected plankton number densities.

n_other

A list array (time x component) that stores the projected values for other ecosystem components.

effort

An array (time x gear) that stores the fishing effort by time and gear.