Example species parameter set based on the North Sea with different gears

This data set is based on species in the North Sea (Blanchard et al.). It is similar to the data set NS_species_params except that this one has an additional column specifying the fishing gear that operates on each species.

Format

A data frame with 12 rows and 8 columns. Each row is a species.

species

Name of the species

w_inf

The von Bertalanffy W_infinity parameter

w_mat

Size at maturity

beta

Size preference ratio

sigma

Width of the size-preference

R_max

Maximum reproduction rate

k_vb

The von Bertalanffy k parameter

gear

Name of the fishing gear

Source

Blanchard et al.

Examples

if (FALSE) {
params <- MizerParams(NS_species_params_gears)
sim = project(params, effort = c(Industrial = 0, Pelagic = 1,
                                 Beam = 0.5, Otter = 0.5))
plot(sim)
}