Length based double-sigmoid selectivity function

A hump-shaped selectivity function with a sigmoidal rise and an independent sigmoidal drop-off. This drop-off is what distinguishes this from the function sigmoid_length() and it is intended to model the escape of large individuals from the fishing gear.

double_sigmoid_length(w, l25, l50, l50_right, l25_right, species_params, ...)

Arguments

w
the size of the individual.

l25
the length which gives a selectivity of 25%.

l50
the length which gives a selectivity of 50%.

l50_right
the length which gives a selectivity of 50%.

l25_right
the length which gives a selectivity of 25%.

species_params
A list with the species params for the current species. Used to get at the length-weight parameters a and b

...
Unused

Details

The selectivity is obtained as the product of two sigmoidal curves, one rising and one dropping. The sigmoidal rise is based on the two parameters l25 and l50 which determine the length at which 25% and 50% of the stock is selected respectively. The sigmoidal drop-off is based on the two parameters l50_right and l25_right which determine the length at which the selectivity curve has dropped back to 50% and 25% respectively.

As the size-based model is weight based, and this selectivity function is length based, it uses the length-weight parameters a and b to convert between length and weight.