Get energy rate available for growth

Calculates the energy rate \(g_i(w)\) (grams/year) available by species and size for growth after metabolism, movement and reproduction have been accounted for. Used by project for performing simulations.

getEGrowth(params, n = params@initial_n, n_pp = params@initial_n_pp,
  B = params@initial_B, e_repro = getERepro(params, n = n, n_pp = n_pp,
  B = B), e = getEReproAndGrowth(params, n = n, n_pp = n_pp, B = B))

Arguments

params
A MizerParams object

n
A matrix of species abundances (species x size)

n_pp
A vector of the plankton abundance by size

B
A vector of biomasses of unstructured resource components

e_repro
The energy available for reproduction (optional, although if specified, e must also be specified). A matrix of size no. species x no. size bins. If not supplied, is calculated internally using getERepro.

e
The energy available for reproduction and growth (optional, although if specified, e_repro must also be specified). A matrix of size no. species x no. size bins. If not supplied, is calculated internally using getEReproAndGrowth.

Value

A two dimensional array (prey species x prey size)

See also

Examples

# NOT RUN {
data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the energy at a particular time step
getEGrowth(params,sim@n[21,,],sim@n_pp[21,])
# }