Get the fishing mortality by time, gear, species and size

Calculates the fishing mortality rate \(F_{g,i,w}\) by gear, species and size at each time step in the effort argument. Used by the project method to perform simulations.

getFMortGear(object, effort, ...)

# S4 method for MizerParams,numeric
getFMortGear(object, effort, ...)

# S4 method for MizerParams,matrix
getFMortGear(object, effort, ...)

# S4 method for MizerSim,missing
getFMortGear(object, effort,
  time_range = dimnames(object@effort)$time, ...)

Arguments

object
A MizerParams object or a MizerSim object.

effort
The effort of each fishing gear. Only needed if the object argument is of class MizerParams. See notes below.

...
Other arguments (currently unused).

time_range
Subset the returned fishing mortalities by time. The time range is either a vector of values, a vector of min and max time, or a single value. Default is the whole time range. Only used if the object argument is of type MizerSim.

Value

An array. If the effort argument has a time dimension, or a MizerSim is passed in, the output array has four dimensions (time x gear x species x size). If the effort argument does not have a time dimension (i.e. it is a vector or a single numeric), the output array has three dimensions (gear x species x size).

Note

Here: fishing mortality = catchability x selectivity x effort.

The effort argument is only used if a MizerParams object is passed in. The effort argument can be a two dimensional array (time x gear), a vector of length equal to the number of gears (each gear has a different effort that is constant in time), or a single numeric value (each gear has the same effort that is constant in time). The order of gears in the effort argument must be the same the same as in the MizerParams object.

If the object argument is of class MizerSim then the effort slot of the MizerSim object is used and the effort argument is not used.

Examples

# NOT RUN {
data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)
# Get the fishing mortality when effort is constant
# for all gears and time:
getFMortGear(params, effort = 1)
# Get the fishing mortality when effort is different
# between the four gears but constant in time:
getFMortGear(params, effort = c(0.5,1,1.5,0.75))
# Get the fishing mortality when effort is different
# between the four gears and changes with time:
effort <- array(NA, dim = c(20,4))
effort[,1] <- seq(from=0, to = 1, length=20)
effort[,2] <- seq(from=1, to = 0.5, length=20)
effort[,3] <- seq(from=1, to = 2, length=20)
effort[,4] <- seq(from=2, to = 1, length=20)
getFMortGear(params, effort=effort)
# Get the fishing mortality using the effort already held in a MizerSim object.
sim <- project(params, t_max = 20, effort = 0.5)
getFMortGear(sim)
getFMortGear(sim, time_range=c(10,20))
# }