Get growth curves giving weight as a function of age

If given a MizerSim object, uses the growth rates at the final time of a simulation to calculate the size at age. If given a MizerParams object, uses the initial growth rates instead.

getGrowthCurves(object, max_age = 24)

Arguments

object
MizerSim or MizerParams object

max_age
The age up to which the weight is to be plotted. Default is 20

Value

An array (species x age) containing the weight in grams.

Examples

data(NS_species_params_gears)
data(inter)
params <- suppressMessages(set_multispecies_model(NS_species_params_gears, inter))
getGrowthCurves(params)

#>          Age
#> Species       0 0.489795918367347 0.979591836734694 1.46938775510204
#>   Sprat   0.001         0.3022132         1.6802528         4.743035
#>   Sandeel 0.001         0.8322811         4.4587932         7.823361
#>   N.pout  0.001         1.2903992         7.6145907        21.162345
#>   Herring 0.001         1.5274228         9.0555478        25.983261
#>   Dab     0.001         1.0908586         6.4081541        18.174056
#>   Whiting 0.001         0.9071907         5.2972941        15.138424
#>   Sole    0.001         0.5087551         2.9051276         8.263291
#>   Gurnard 0.001         0.3517496         1.9720635         5.580511
#>   Plaice  0.001         0.1801163         0.9674818         2.702355
#>   Haddock 0.001         1.7255450        10.2594485        29.433381
#>   Cod     0.001         6.4458336        38.9488064       109.298886
#>   Saithe  0.001         3.7320080        22.4787134        63.925297
#>          Age
#> Species   1.95918367346939 2.44897959183673 2.93877551020408 3.42857142857143
#>   Sprat           9.932817         14.83905         17.48829         19.44856
#>   Sandeel        11.218278         14.53660         17.62694         20.41406
#>   N.pout         31.506429         38.86638         45.46661         51.47438
#>   Herring        54.941626         94.36491        122.24836        141.53463
#>   Dab            30.976362         43.57647         57.09078         71.16661
#>   Whiting        31.485263         54.09719         78.68004         98.85515
#>   Sole           17.441846         31.15996         50.00236         73.01013
#>   Gurnard        11.758428         21.00035         33.45092         46.09430
#>   Plaice          5.663479         10.09367         16.20762         24.19176
#>   Haddock        62.097973        110.32774        170.19997        226.30928
#>   Cod           216.358111        346.05137        479.90854        603.62724
#>   Saithe        129.338307        210.78121        295.20684        371.71849
#>          Age
#> Species   3.91836734693878 4.40816326530612 4.89795918367347 5.38775510204082
#>   Sprat           21.06527         22.48612         23.73947         24.84670
#>   Sandeel         22.86799         25.00032         26.81985         28.36919
#>   N.pout          56.90137         61.77798         66.13035         70.01227
#>   Herring        158.00553        173.01144        186.91004        199.82365
#>   Dab             85.46166         99.71009        113.70391        127.30740
#>   Whiting        117.08357        134.57198        151.43786        167.60899
#>   Sole            93.67704        111.83009        129.73861        147.94098
#>   Gurnard         57.55631         69.29325         81.57341         94.34721
#>   Plaice          34.20480         46.37699         60.80349         77.43791
#>   Haddock        282.94675        342.93013        406.02365        471.64532
#>   Cod            709.60330        795.65350        863.56550        915.25554
#>   Saithe         435.03371        484.30668        521.01062        547.61778
#>          Age
#> Species   5.87755102040816 6.36734693877551 6.85714285714286 7.3469387755102
#>   Sprat           25.82717         26.69528         27.46394        28.14454
#>   Sandeel         29.67403         30.76042         31.66495        32.41805
#>   N.pout          73.45782         76.51113         79.21689        81.61457
#>   Herring        211.77193        222.79681        232.96969       242.33674
#>   Dab            140.40535        152.92462        164.82795       176.09919
#>   Whiting        183.02384        197.65821        211.50737       224.56284
#>   Sole           166.44532        185.17017        204.03292       222.91504
#>   Gurnard        107.52509        121.01188        134.70988       148.56071
#>   Plaice          95.52966        113.29183        130.03475       146.43715
#>   Haddock        539.24919        608.33575        678.40353       749.11747
#>   Cod            954.51918        984.61275       1009.83413      1031.45860
#>   Saithe         566.90476        581.37768        593.89427       604.87574
#>          Age
#> Species   7.83673469387755 8.3265306122449 8.81632653061224 9.30612244897959
#>   Sprat           28.74838        29.28797         29.77048         30.20196
#>   Sandeel         33.04505        33.56709         34.00512         34.40581
#>   N.pout          83.73773        85.61563         87.27635         88.74514
#>   Herring        250.89066       258.69593        265.81807        272.31681
#>   Dab            186.70363       196.67997        206.03717        214.76417
#>   Whiting        236.87010       248.45065        259.32730        269.54327
#>   Sole           241.77026       260.49182        279.04057        297.37303
#>   Gurnard        162.47349       176.40957        190.28823        204.09628
#>   Plaice         163.03571       179.96644        197.28021        214.94512
#>   Haddock        820.12436       891.11762        961.85906       1032.18761
#>   Cod           1049.99864      1065.89436       1079.52291       1091.20751
#>   Saithe         614.51032       622.96316        630.37884        636.88480
#>          Age
#> Species   9.79591836734694 10.2857142857143 10.7755102040816 11.265306122449
#>   Sprat           30.58780         30.93283         31.24137        31.51727
#>   Sandeel         34.78137         35.13338         35.46332        35.77257
#>   N.pout          90.04414         91.19297         92.20900        93.10757
#>   Herring        278.24665        283.65653        288.57478       293.03934
#>   Dab            222.90240        230.49157        237.56834       244.13917
#>   Whiting        279.13867        288.15119        296.61625       304.56709
#>   Sole           315.40841        333.14861        350.56518       367.59215
#>   Gurnard        217.75794        231.25779        244.58195       257.66877
#>   Plaice         232.91524        251.15804        269.59879       288.23136
#>   Haddock       1101.81956       1170.75049       1238.76609      1305.74406
#>   Cod           1101.22553       1109.81480       1117.17901      1123.49288
#>   Saithe         642.59265        647.60023        651.99350       655.84785
#>          Age
#> Species   11.7551020408163 12.2448979591837 12.734693877551 13.2244897959184
#>   Sprat           31.76399         31.98461        32.18190         32.35832
#>   Sandeel         36.06243         36.33411        36.58876         36.82744
#>   N.pout          93.90227         94.60509        95.22667         95.77639
#>   Herring        297.09207        300.77096       304.11046        307.14191
#>   Dab            250.22423        255.85955       261.07831        265.91147
#>   Whiting        312.03496        319.04919       325.63735        331.82530
#>   Sole           384.23573        400.50458       416.37342        431.80345
#>   Gurnard        270.51569        283.12686       295.46804        307.51452
#>   Plaice         306.97626        325.80716       344.71553        363.62707
#>   Haddock       1371.70023       1436.56809      1500.19177       1562.58695
#>   Cod           1128.90622       1133.54748      1137.52676       1140.93850
#>   Saithe         659.22937        662.19606       664.79882        667.08228
#>          Age
#> Species   13.7142857142857 14.2040816326531 14.6938775510204 15.1836734693878
#>   Sprat           32.51608         32.65715         32.78330         32.89611
#>   Sandeel         37.05115         37.26084         37.45737         37.64158
#>   N.pout          96.26257         96.69254         97.07281         97.40912
#>   Herring        309.89371        312.39165        314.65917        316.71752
#>   Dab            270.38750        274.53268        278.37147        281.92654
#>   Whiting        337.63830        343.10962        348.26826        353.13157
#>   Sole           446.80666        461.39476        475.57922        489.35733
#>   Gurnard        319.27352        330.75182        341.95086        352.83922
#>   Plaice         382.52303        401.40207        420.22629        438.94987
#>   Haddock       1623.77830       1683.76274       1742.44595       1799.83051
#>   Cod           1143.86363       1146.37155       1148.52178       1150.36533
#>   Saithe         669.08562        670.84320        672.38518        673.73799
#>          Age
#> Species   15.6734693877551 16.1632653061224 16.6530612244898 17.1428571428571
#>   Sprat           32.99699         33.08719         33.16786         33.23999
#>   Sandeel         37.81424         37.97608         38.12777         38.26994
#>   N.pout          97.70695         97.97932         98.24651         98.50989
#>   Herring        318.58599        320.28210        321.82175        323.21937
#>   Dab            285.23711        288.34820        291.27188        294.01994
#>   Whiting        357.71662        362.03933        366.11471        369.95690
#>   Sole           502.69998        515.61775        528.12478        540.23390
#>   Gurnard        363.41782        373.69548        383.68075        393.38197
#>   Plaice         457.57222        476.09373        494.49319        512.72565
#>   Haddock       1855.94516       1910.81798       1964.47579       2016.92616
#>   Cod           1151.94594       1153.30111       1154.46300       1155.45917
#>   Saithe         674.92484        675.96610        676.87963        677.68108
#>          Age
#> Species   17.6326530612245 18.1224489795918 18.6122448979592 19.1020408163265
#>   Sprat            33.3045         33.36218         33.41376         33.45988
#>   Sandeel          38.4032         38.52810         38.64517         38.75491
#>   N.pout           98.7695         99.02540         99.27764         99.52627
#>   Herring         324.4881        325.63973        326.68515        327.63413
#>   Dab             296.6027        299.03026        301.31183        303.45619
#>   Whiting         373.5793        376.99436        380.21406        383.24953
#>   Sole            551.9578        563.30872        574.29732        584.91533
#>   Gurnard         402.8071        411.94504        420.78994        429.35134
#>   Plaice          530.7910        548.69092        566.42648        583.98848
#>   Haddock        2068.1228       2118.10573       2166.90166       2214.53912
#>   Cod            1156.3133       1157.04553       1157.67336       1158.21165
#>   Saithe          678.3842        679.00110        679.54231        680.01712
#>          Age
#> Species   19.5918367346939 20.0816326530612 20.5714285714286 21.0612244897959
#>   Sprat           33.50113         33.53801         33.57100         33.60049
#>   Sandeel         38.85775         38.95416         39.04451         39.12920
#>   N.pout          99.77135        100.01292        100.25105        100.48577
#>   Herring        328.49558        329.27756        329.98740        330.63176
#>   Dab            305.47161        307.36583        309.14615        310.81942
#>   Whiting        386.11133        388.80938        391.35306        393.75120
#>   Sole           595.16239        605.05147        614.59509        623.80529
#>   Gurnard        437.63845        445.66001        453.42451        460.94021
#>   Plaice         601.34305        618.48875        635.42809        652.16354
#>   Haddock       2261.04563       2306.44803       2350.77244       2394.03775
#>   Cod           1158.67316       1159.06885       1159.40810       1159.69897
#>   Saithe         680.43369        680.79915        681.11978        681.40108
#>          Age
#> Species   21.5510204081633 22.0408163265306 22.530612244898 23.0204081632653
#>   Sprat           33.62686         33.65045        33.67154         33.69040
#>   Sandeel         39.20858         39.28298        39.35272         39.41808
#>   N.pout         100.71713        100.94519       101.16998        101.39157
#>   Herring        331.21668        331.74765       332.22963        332.66715
#>   Dab            312.39207        313.87015       315.25935        316.56501
#>   Whiting        396.01212        398.14369       400.15329        402.04791
#>   Sole           632.69373        641.27166       649.54991        657.53897
#>   Gurnard        468.21506        475.25678       482.07194        488.65552
#>   Plaice         668.69758        685.03181       701.15142        717.04519
#>   Haddock       2436.26028       2477.46424      2517.67458       2556.91518
#>   Cod           1159.94835       1160.16216      1160.34548       1160.50265
#>   Saithe         681.64787        681.86439       682.05434        682.22099
#>          Age
#> Species   23.5102040816327         24
#>   Sprat           33.70726   33.72234
#>   Sandeel         39.47935   39.53677
#>   N.pout         101.60998  101.82527
#>   Herring        333.06432  333.42484
#>   Dab            317.79216  318.94552
#>   Whiting        403.86289  405.63552
#>   Sole           665.24893  672.68954
#>   Gurnard        495.01053  501.14557
#>   Plaice         732.71692  748.16953
#>   Haddock       2595.20946 2632.58025
#>   Cod           1160.63740 1160.75294
#>   Saithe         682.36720  682.49547

sim <- project(params, effort=1, t_max=20, t_save = 2, progress_bar = FALSE)
getGrowthCurves(sim, max_age = 24)

#>          Age
#> Species       0 0.489795918367347 0.979591836734694 1.46938775510204
#>   Sprat   0.001         0.2993893         1.6332965         4.515792
#>   Sandeel 0.001         0.8270020         4.4027066         7.706426
#>   N.pout  0.001         0.8669787         4.7089976        15.857002
#>   Herring 0.001         1.5045243         8.7433378        24.591701
#>   Dab     0.001         0.8857705         4.4159471        11.418070
#>   Whiting 0.001         0.6200264         3.0509787         9.215514
#>   Sole    0.001         0.4489029         2.2826839         5.838531
#>   Gurnard 0.001         0.3121939         1.5725175         4.041732
#>   Plaice  0.001         0.1588358         0.7626962         1.921199
#>   Haddock 0.001         1.4569440         7.6251975        20.268913
#>   Cod     0.001         3.8810470        23.1881386        99.051591
#>   Saithe  0.001         2.1297511         9.5580091        26.054034
#>          Age
#> Species   1.95918367346939 2.44897959183673 2.93877551020408 3.42857142857143
#>   Sprat           9.287000        14.238144        16.968269         18.90032
#>   Sandeel        11.016086        14.251343        17.271916         20.00614
#>   N.pout         31.024974        41.618743        51.471617         60.38932
#>   Herring        51.017054        87.736425       116.717390        135.84884
#>   Dab            22.188395        32.477197        43.739046         56.51406
#>   Whiting        24.467547        58.734773       104.175254        151.32081
#>   Sole           11.250204        18.671902        28.350017         40.64279
#>   Gurnard         7.859118        13.189341        20.272820         29.38406
#>   Plaice          3.691813         6.152248         9.428694         13.71200
#>   Haddock        41.681153        75.533931       127.039640        191.37526
#>   Cod           339.641420       888.811067      1782.806200       2684.97065
#>   Saithe         61.052542       131.320348       259.718895        471.97725
#>          Age
#> Species   3.91836734693878 4.40816326530612 4.89795918367347 5.38775510204082
#>   Sprat           20.48712         21.87369         23.10620         24.20218
#>   Sandeel         22.42790         24.54445         26.36416         27.92167
#>   N.pout          68.12594         74.64133         79.98375         84.33284
#>   Herring        151.60178        165.94647        179.18552        191.54104
#>   Dab             70.65341         85.88623        101.89430        118.33426
#>   Whiting        205.27802        264.20389        326.03468        388.95650
#>   Sole            55.95754         73.55296         89.96105        104.92885
#>   Gurnard         39.72877         49.38461         58.93359         69.05629
#>   Plaice          19.27312         26.48034         35.80904         47.86478
#>   Haddock        257.24129        332.73858        419.77291        518.09540
#>   Cod           3696.54813       4821.95282       6035.44819       7312.07340
#>   Saithe         794.57597       1201.39515       1616.37704       2077.84476
#>          Age
#> Species   5.87755102040816 6.36734693877551 6.85714285714286 7.3469387755102
#>   Sprat           25.18015         26.05361         26.83371        27.53045
#>   Sandeel         29.24867         30.36271         31.29699        32.08053
#>   N.pout          87.75874         90.44585         92.55341        94.20646
#>   Herring        203.06648        213.77401        223.71259       232.93748
#>   Dab            134.85947        151.18702        167.02745       182.15912
#>   Whiting        451.49498        512.52303        571.12656       626.76747
#>   Sole           119.89627        135.42267        151.67889       168.68390
#>   Gurnard         79.92848         91.57662        103.98164       117.09977
#>   Plaice          63.38801         83.05307        105.72887       128.60309
#>   Haddock        626.93086        745.14840        871.39616      1004.14998
#>   Cod           8629.76893       9969.27653      11314.02823     12652.10968
#>   Saithe        2596.22782       3168.35965       3789.07380      4452.84771
#>          Age
#> Species   7.83673469387755 8.3265306122449 8.81632653061224 9.30612244897959
#>   Sprat           28.15274        28.70970         29.21179         29.66483
#>   Sandeel         32.73761        33.28866         33.75079         34.15093
#>   N.pout          95.50287        96.51977         97.31740         97.94879
#>   Herring        241.48705       249.35624        256.59319        263.24873
#>   Dab            196.44230       209.79340        222.07081        233.35816
#>   Whiting        679.05556       727.83241        772.86113        814.42623
#>   Sole           186.37902       204.72530        223.60728        242.98722
#>   Gurnard        130.88098       145.27595        160.19206        175.58492
#>   Plaice         152.24412       177.87728        205.93140        236.48494
#>   Haddock       1141.77122      1282.84907       1425.91220       1569.33708
#>   Cod          13972.18869     15265.06354      16525.69921      17748.21604
#>   Saithe        5154.08432      5887.28125       6647.01769       7429.25308
#>          Age
#> Species   9.79591836734694 10.2857142857143 10.7755102040816 11.265306122449
#>   Sprat           30.07363         30.44249         30.77532        31.07564
#>   Sandeel         34.52578         34.87813         35.20932        35.52063
#>   N.pout          98.52943         99.09206         99.63724       100.16552
#>   Herring        269.36953        274.99857        280.17535       284.93394
#>   Dab            243.67233        252.92315        261.21048       268.63451
#>   Whiting        852.50353        887.07945        918.47609       946.98586
#>   Sole           262.72402        282.77399        303.02169       323.35768
#>   Gurnard        191.33624        207.40877        223.66374       240.07994
#>   Plaice         269.50943        304.93639        342.63176       382.44126
#>   Haddock       1712.33615       1853.29080       1991.82828      2126.76864
#>   Cod          18928.24493      20066.54206      21157.56821     22201.39676
#>   Saithe        8229.21753       9042.48670       9866.27079     10696.01334
#>          Age
#> Species   11.7551020408163 12.2448979591837 12.734693877551 13.2244897959184
#>   Sprat           31.34663         31.59114        31.81177         32.01085
#>   Sandeel         35.81326         36.08831        36.34686         36.58988
#>   N.pout         100.67741        101.17342       101.65405        102.11978
#>   Herring        289.29526        293.29090       296.95169        300.30568
#>   Dab            275.28524        281.24345       286.59398        291.43137
#>   Whiting        972.87393        996.21548      1017.09261       1035.76832
#>   Sole           343.76785        364.08055       384.24221        404.25240
#>   Gurnard        256.52512        272.93625       289.29340        305.45046
#>   Plaice         424.26089        467.85148       513.11018        559.78313
#>   Haddock       2257.46283       2383.99502      2505.34336       2621.22283
#>   Cod          23200.07037      24153.54672     25059.04567      25918.79335
#>   Saithe       11529.51644      12363.74177     13195.14489      14023.46992
#>          Age
#> Species   13.7142857142857 14.2040816326531 14.6938775510204 15.1836734693878
#>   Sprat           32.19048         32.35257         32.49882         32.63079
#>   Sandeel         36.81832         37.03304         37.23487         37.42458
#>   N.pout         102.57106        103.00834        103.43206        103.84265
#>   Herring        303.37858        306.19396        308.77339        311.13665
#>   Dab            295.80429        299.75734        303.33080        306.56113
#>   Whiting       1052.47421       1067.41836       1080.78623       1092.74384
#>   Sole           423.97068        443.30741        462.27004        480.86494
#>   Gurnard        321.37522        337.07103        352.44410        367.40850
#>   Plaice         607.77000        656.78667        706.78663        757.44723
#>   Haddock       2731.87984       2837.50379       2937.39737       3031.51830
#>   Cod          26735.08302      27510.12216      28245.67728      28940.41232
#>   Saithe       14844.34757      15656.18816      16459.08556      17250.25371
#>          Age
#> Species   15.6734693877551 16.1632653061224 16.6530612244898 17.1428571428571
#>   Sprat           32.74987         32.85731         32.95426         33.04174
#>   Sandeel         37.60291         37.77053         37.92809         38.07619
#>   N.pout         104.24049        104.62600        104.99955        105.36152
#>   Herring        313.30185        315.28560        317.10310        318.76828
#>   Dab            309.48128        312.12102        314.50729        316.66443
#>   Whiting       1103.44016       1113.00824       1121.56700       1129.22298
#>   Sole           498.97858        516.52615        533.52547        549.99359
#>   Gurnard        381.97506        396.15434        409.92864        423.18722
#>   Plaice         808.71176        860.42097        912.30782        964.36521
#>   Haddock       3120.19966       3203.75554       3282.48308       3356.62747
#>   Cod          29595.69943      30213.78249      30796.77216      31346.66754
#>   Saithe       18026.96095      18789.45332      19537.98938      20270.90390
#>          Age
#> Species   17.6326530612245 18.1224489795918 18.6122448979592 19.1020408163265
#>   Sprat           33.12068         33.19190         33.25617         33.31416
#>   Sandeel         38.21540         38.34625         38.46924         38.58486
#>   N.pout         105.71226        106.05213        106.38145        106.70056
#>   Herring        320.29391        321.69168        322.97231        324.14561
#>   Dab            318.61444        320.37721        321.97071        323.41121
#>   Whiting       1136.07142       1142.19748       1147.67736       1152.57922
#>   Sole           565.94712        581.37204        596.15054        610.29376
#>   Gurnard        435.93416        448.18921        459.97135        471.29882
#>   Plaice        1016.45533       1068.31426       1119.93426       1171.30870
#>   Haddock       3426.05353       3490.64618       3550.73899       3606.64545
#>   Cod          31865.34725      32354.57729      32816.03063      33251.27752
#>   Saithe       20985.34741      21681.69673      22360.41036      23021.93491
#>          Age
#> Species   19.5918367346939 20.0816326530612 20.5714285714286 21.0612244897959
#>   Sprat           33.36649         33.41370         33.45630         33.49474
#>   Sandeel         38.69353         38.79568         38.89169         38.98195
#>   N.pout         107.00977        107.30939        107.59972        107.88105
#>   Herring        325.22059        326.20547        327.10782        327.93455
#>   Dab            324.71339        325.89053        326.95464        327.91657
#>   Whiting       1156.96403       1160.88632       1164.51568       1168.08207
#>   Sole           623.82886        636.78198        649.17820        661.04139
#>   Gurnard        482.18650        492.56591        502.42344        511.78533
#>   Plaice        1222.19122       1272.46313       1322.13186       1371.20471
#>   Haddock       3658.65706       3707.04522       3752.06236       3793.94330
#>   Cod          33661.19503      34045.98336      34406.71953      34745.09448
#>   Saithe       23666.42952      24291.24085      24896.02416      25481.42206
#>          Age
#> Species   21.5510204081633 22.0408163265306 22.530612244898 23.0204081632653
#>   Sprat           33.52943         33.56073        33.58897         33.61445
#>   Sandeel         39.06678         39.14652        39.22148         39.29193
#>   N.pout         108.15365        108.41780       108.67375        108.92177
#>   Herring        328.69199        329.38595       330.02175        330.60427
#>   Dab            328.78614        329.57221       330.28280        330.92516
#>   Whiting       1171.59101       1175.04348      1178.44035       1181.78254
#>   Sole           672.39448        683.25741       693.59483        703.36934
#>   Gurnard        520.67651        529.12063       537.14018        544.75650
#>   Plaice        1419.60652       1467.11350      1513.72604       1559.46089
#>   Haddock       3832.90657       3869.15543      3902.87898       3934.25315
#>   Cod          35062.43146      35360.04762     35639.16767      35900.94118
#>   Saithe       26048.06180      26596.54156     27127.44416      27641.33232
#>          Age
#> Species   23.5102040816327          24
#>   Sprat           33.63744    33.65819
#>   Sandeel         39.35816    39.42041
#>   N.pout         109.16209   109.39496
#>   Herring        331.13797   331.62695
#>   Dab            331.50583   332.03076
#>   Whiting       1185.07092  1188.30635
#>   Sole           712.61138   721.34997
#>   Gurnard        551.98987   558.85954
#>   Plaice        1604.33457  1648.36332
#>   Haddock       3963.44161  3990.59992
#>   Cod          36146.44612 36376.69354
#>   Saithe       28138.57836 28617.50223