Get growth curves giving weight as a function of age

If given a MizerSim object, uses the growth rates at the final time of a simulation to calculate the size at age. If given a MizerParams object, uses the initial growth rates instead.

getGrowthCurves(object, max_age = 24)

Arguments

object
MizerSim or MizerParams object

max_age
The age up to which the weight is to be plotted. Default is 20

Value

An array (species x age) containing the weight in grams.

Examples

data(NS_species_params_gears)
data(inter)
params <- suppressMessages(set_multispecies_model(NS_species_params_gears, inter))
getGrowthCurves(params)

#>          Age
#> Species       0 0.489795918367347 0.979591836734694 1.46938775510204
#>   Sprat   0.001         0.3022132         1.6802528         4.743035
#>   Sandeel 0.001         0.8322811         4.4587948         7.823354
#>   N.pout  0.001         1.2903992         7.6145909        21.162349
#>   Herring 0.001         1.5274228         9.0555478        25.983261
#>   Dab     0.001         1.0908586         6.4081541        18.174056
#>   Whiting 0.001         0.9071907         5.2972941        15.138424
#>   Sole    0.001         0.5087551         2.9051276         8.263291
#>   Gurnard 0.001         0.3517496         1.9720635         5.580511
#>   Plaice  0.001         0.1801163         0.9674818         2.702355
#>   Haddock 0.001         1.7255450        10.2594485        29.433381
#>   Cod     0.001         6.4458336        38.9488064       109.298971
#>   Saithe  0.001         3.7320080        22.4787134        63.925297
#>          Age
#> Species   1.95918367346939 2.44897959183673 2.93877551020408 3.42857142857143
#>   Sprat           9.932817         14.83905         17.48829         19.44856
#>   Sandeel        11.218248         14.53658         17.62694         20.41406
#>   N.pout         31.506472         38.86652         45.46705         51.47487
#>   Herring        54.941626         94.36491        122.24836        141.53462
#>   Dab            30.976362         43.57647         57.09078         71.16661
#>   Whiting        31.485263         54.09719         78.68002         98.85516
#>   Sole           17.441846         31.15996         50.00236         73.01013
#>   Gurnard        11.758428         21.00035         33.45092         46.09430
#>   Plaice          5.663479         10.09367         16.20762         24.19176
#>   Haddock        62.097973        110.32774        170.19997        226.30928
#>   Cod           216.359206        346.04973        479.90597        603.62332
#>   Saithe        129.338276        210.78100        295.20584        371.71767
#>          Age
#> Species   3.91836734693878 4.40816326530612 4.89795918367347 5.38775510204082
#>   Sprat           21.06527         22.48612         23.73947         24.84670
#>   Sandeel         22.86799         25.00027         26.81983         28.36918
#>   N.pout          56.90179         61.77835         66.13063         70.01297
#>   Herring        158.00553        173.01144        186.91004        199.82365
#>   Dab             85.46166         99.71009        113.70391        127.30740
#>   Whiting        117.08320        134.57160        151.43754        167.60914
#>   Sole            93.67704        111.83009        129.73861        147.94098
#>   Gurnard         57.55631         69.29325         81.57341         94.34721
#>   Plaice          34.20480         46.37699         60.80349         77.43791
#>   Haddock        282.94675        342.93013        406.02365        471.64532
#>   Cod            709.59646        795.64817        863.56260        915.25392
#>   Saithe         435.03221        484.30571        521.01003        547.61729
#>          Age
#> Species   5.87755102040816 6.36734693877551 6.85714285714286 7.3469387755102
#>   Sprat           25.82717         26.69528         27.46394        28.14454
#>   Sandeel         29.67405         30.76046         31.66498        32.41806
#>   N.pout          73.45843         76.51170         79.21738        81.61506
#>   Herring        211.77193        222.79681        232.96969       242.33674
#>   Dab            140.40535        152.92462        164.82795       176.09919
#>   Whiting        183.02398        197.65834        211.50784       224.56321
#>   Sole           166.44532        185.17017        204.03292       222.91504
#>   Gurnard        107.52509        121.01188        134.70988       148.56071
#>   Plaice          95.52966        113.29183        130.03474       146.43714
#>   Haddock        539.24919        608.33575        678.40353       749.11747
#>   Cod            954.51766        984.61347       1009.83473      1031.45888
#>   Saithe         566.90439        581.37749        593.89457       604.87620
#>          Age
#> Species   7.83673469387755 8.3265306122449 8.81632653061224 9.30612244897959
#>   Sprat           28.74837        29.28796         29.77047         30.20196
#>   Sandeel         33.04507        33.56709         34.00511         34.40581
#>   N.pout          83.73838        85.61609         87.27680         88.74554
#>   Herring        250.89065       258.69593        265.81807        272.31681
#>   Dab            186.70363       196.67997        206.03717        214.76417
#>   Whiting        236.87036       248.45046        259.32689        269.54280
#>   Sole           241.77026       260.49182        279.04057        297.37303
#>   Gurnard        162.47349       176.40957        190.28823        204.09628
#>   Plaice         163.03571       179.96645        197.28018        214.94508
#>   Haddock        820.12436       891.11762        961.85906       1032.18761
#>   Cod           1049.99927      1065.89495       1079.52347       1091.20816
#>   Saithe         614.51089       622.96345        630.37910        636.88503
#>          Age
#> Species   9.79591836734694 10.2857142857143 10.7755102040816 11.265306122449
#>   Sprat           30.58780         30.93283         31.24136        31.51726
#>   Sandeel         34.78134         35.13335         35.46329        35.77254
#>   N.pout          90.04449         91.19328         92.20927        93.10781
#>   Herring        278.24665        283.65653        288.57478       293.03934
#>   Dab            222.90240        230.49157        237.56834       244.13917
#>   Whiting        279.13816        288.15073        296.61583       304.56671
#>   Sole           315.40841        333.14861        350.56518       367.59215
#>   Gurnard        217.75794        231.25779        244.58196       257.66877
#>   Plaice         232.91522        251.15778        269.59853       288.23076
#>   Haddock       1101.81956       1170.75049       1238.76609      1305.74406
#>   Cod           1101.22609       1109.81521       1117.17937      1123.49319
#>   Saithe         642.59279        647.60034        651.99362       655.84795
#>          Age
#> Species   11.7551020408163 12.2448979591837 12.734693877551 13.2244897959184
#>   Sprat           31.76398         31.98461        32.18189         32.35831
#>   Sandeel         36.06240         36.33409        36.58873         36.82742
#>   N.pout          93.90248         94.60528        95.22683         95.77654
#>   Herring        297.09206        300.77096       304.11046        307.14191
#>   Dab            250.22424        255.85955       261.07832        265.91148
#>   Whiting        312.03460        319.04886       325.63703        331.82501
#>   Sole           384.23573        400.50458       416.37342        431.80345
#>   Gurnard        270.51569        283.12687       295.46805        307.51453
#>   Plaice         306.97565        325.80655       344.71488        363.62612
#>   Haddock       1371.70023       1436.56809      1500.19177       1562.58695
#>   Cod           1128.90648       1133.54770      1137.52696       1140.93866
#>   Saithe         659.22946        662.19614       664.79889        667.08234
#>          Age
#> Species   13.7142857142857 14.2040816326531 14.6938775510204 15.1836734693878
#>   Sprat           32.51607         32.65715         32.78330         32.89611
#>   Sandeel         37.05113         37.26081         37.45735         37.64157
#>   N.pout          96.26270         96.69265         97.07291         97.40921
#>   Herring        309.89371        312.39165        314.65917        316.71752
#>   Dab            270.38751        274.53268        278.37147        281.92654
#>   Whiting        337.63832        343.10838        348.26683        353.13022
#>   Sole           446.80666        461.39476        475.57922        489.35733
#>   Gurnard        319.27352        330.75183        341.95086        352.83923
#>   Plaice         382.52209        401.40268        420.22764        438.95121
#>   Haddock       1623.77830       1683.76274       1742.44595       1799.83051
#>   Cod           1143.86377       1146.37167       1148.52188       1150.36542
#>   Saithe         669.08568        670.84325        672.38522        673.73802
#>          Age
#> Species   15.6734693877551 16.1632653061224 16.6530612244898 17.1428571428571
#>   Sprat           32.99699         33.08719         33.16786         33.23999
#>   Sandeel         37.81423         37.97606         38.12775         38.26993
#>   N.pout          97.70686         97.97884         98.24604         98.50942
#>   Herring        318.58599        320.28210        321.82175        323.21937
#>   Dab            285.23711        288.34820        291.27188        294.01994
#>   Whiting        357.71534        362.03813        366.11357        369.95584
#>   Sole           502.69998        515.61775        528.12478        540.23390
#>   Gurnard        363.41783        373.69548        383.68076        393.38197
#>   Plaice         457.57354        476.09521        494.49446        512.72689
#>   Haddock       1855.94516       1910.81798       1964.47579       2016.92616
#>   Cod           1151.94601       1153.30117       1154.46305       1155.45922
#>   Saithe         674.92488        675.96613        676.87965        677.68110
#>          Age
#> Species   17.6326530612245 18.1224489795918 18.6122448979592 19.1020408163265
#>   Sprat           33.30450         33.36218         33.41376         33.45988
#>   Sandeel         38.40319         38.52809         38.64516         38.75489
#>   N.pout          98.76903         99.02494         99.27719         99.52583
#>   Herring        324.48807        325.63973        326.68515        327.63413
#>   Dab            296.60274        299.03026        301.31183        303.45619
#>   Whiting        373.57826        376.99341        380.21316        383.24869
#>   Sole           551.95780        563.30872        574.29732        584.91533
#>   Gurnard        402.80713        411.94503        420.78994        429.35133
#>   Plaice         530.79227        548.69213        566.42798        583.99050
#>   Haddock       2068.12279       2118.10573       2166.90166       2214.53912
#>   Cod           1156.31330       1157.04557       1157.67339       1158.21168
#>   Saithe         678.38424        679.00112        679.54232        680.01714
#>          Age
#> Species   19.5918367346939 20.0816326530612 20.5714285714286 21.0612244897959
#>   Sprat           33.50113         33.53801          33.5710         33.60049
#>   Sandeel         38.85774         38.95415          39.0445         39.12919
#>   N.pout          99.77091        100.01249         100.2506        100.48535
#>   Herring        328.49558        329.27756         329.9874        330.63176
#>   Dab            305.47161        307.36583         309.1462        310.81942
#>   Whiting        386.11053        388.80863         391.3524        393.75053
#>   Sole           595.16239        605.05147         614.5951        623.80529
#>   Gurnard        437.63845        445.66000         453.4245        460.94021
#>   Plaice         601.34506        618.49070         635.4300        652.16544
#>   Haddock       2261.04563       2306.44803        2350.7724       2394.03775
#>   Cod           1158.67318       1159.06887        1159.4081       1159.69898
#>   Saithe         680.43370        680.79916         681.1198        681.40109
#>          Age
#> Species   21.5510204081633 22.0408163265306 22.530612244898 23.0204081632653
#>   Sprat           33.62686         33.65045        33.67154         33.69040
#>   Sandeel         39.20857         39.28297        39.35271         39.41808
#>   N.pout         100.71672        100.94478       101.16958        101.39117
#>   Herring        331.21668        331.74765       332.22963        332.66715
#>   Dab            312.39207        313.87015       315.25935        316.56501
#>   Whiting        396.01150        398.14309       400.15273        402.04740
#>   Sole           632.69373        641.27166       649.54991        657.53897
#>   Gurnard        468.21505        475.25678       482.07193        488.65550
#>   Plaice         668.69945        685.03425       701.15364        717.04759
#>   Haddock       2436.26028       2477.46424      2517.67458       2556.91518
#>   Cod           1159.94836       1160.16217      1160.34549       1160.50266
#>   Saithe         681.64788        681.86439       682.05435        682.22100
#>          Age
#> Species   23.5102040816327         24
#>   Sprat           33.70726   33.72234
#>   Sandeel         39.47934   39.53676
#>   N.pout         101.60959  101.82489
#>   Herring        333.06432  333.42484
#>   Dab            317.79216  318.94553
#>   Whiting        403.86237  405.63501
#>   Sole           665.24893  672.68954
#>   Gurnard        495.01051  501.14556
#>   Plaice         732.71927  748.17185
#>   Haddock       2595.20946 2632.58025
#>   Cod           1160.63741 1160.75294
#>   Saithe         682.36720  682.49548

sim <- project(params, effort=1, t_max=20, t_save = 2, progress_bar = FALSE)
getGrowthCurves(sim, max_age = 24)

#>          Age
#> Species       0 0.489795918367347 0.979591836734694 1.46938775510204
#>   Sprat   0.001         0.2993893         1.6332965         4.515792
#>   Sandeel 0.001         0.8270020         4.4027066         7.706426
#>   N.pout  0.001         0.8669787         4.7089976        15.857009
#>   Herring 0.001         1.5045243         8.7433378        24.591701
#>   Dab     0.001         0.8857705         4.4159471        11.418070
#>   Whiting 0.001         0.6200264         3.0509787         9.215514
#>   Sole    0.001         0.4489029         2.2826839         5.838531
#>   Gurnard 0.001         0.3121939         1.5725175         4.041732
#>   Plaice  0.001         0.1588358         0.7626962         1.921199
#>   Haddock 0.001         1.4569440         7.6251975        20.268913
#>   Cod     0.001         3.8810467        23.1881165        99.051534
#>   Saithe  0.001         2.1297511         9.5580091        26.054034
#>          Age
#> Species   1.95918367346939 2.44897959183673 2.93877551020408 3.42857142857143
#>   Sprat           9.287000        14.238144        16.968269         18.90032
#>   Sandeel        11.016086        14.251343        17.271916         20.00614
#>   N.pout         31.024997        41.618740        51.471484         60.38942
#>   Herring        51.017054        87.736425       116.717390        135.84884
#>   Dab            22.188395        32.477285        43.738944         56.51395
#>   Whiting        24.467547        58.734770       104.175267        151.32085
#>   Sole           11.250204        18.671902        28.350018         40.64279
#>   Gurnard         7.859118        13.189341        20.272820         29.38406
#>   Plaice          3.691813         6.152248         9.428694         13.71200
#>   Haddock        41.681153        75.533931       127.039640        191.37526
#>   Cod           339.636846       888.803958      1782.798461       2684.96215
#>   Saithe         61.052641       131.321351       259.718206        471.97618
#>          Age
#> Species   3.91836734693878 4.40816326530612 4.89795918367347 5.38775510204082
#>   Sprat           20.48712         21.87369         23.10620         24.20218
#>   Sandeel         22.42790         24.54445         26.36416         27.92167
#>   N.pout          68.12599         74.64139         79.98383         84.33318
#>   Herring        151.60178        165.94647        179.18552        191.54104
#>   Dab             70.65319         85.88596        101.89409        118.33417
#>   Whiting        205.27745        264.20243        326.03303        388.95539
#>   Sole            55.95754         73.55296         89.96108        104.92887
#>   Gurnard         39.72877         49.38461         58.93359         69.05629
#>   Plaice          19.27312         26.48034         35.80904         47.86478
#>   Haddock        257.24129        332.73857        419.77289        518.09537
#>   Cod           3696.51408       4821.93650       6035.45949       7312.07166
#>   Saithe         794.57500       1201.39382       1616.37342       2077.83461
#>          Age
#> Species   5.87755102040816 6.36734693877551 6.85714285714286 7.3469387755102
#>   Sprat           25.18015         26.05361         26.83371        27.53045
#>   Sandeel         29.24867         30.36271         31.29699        32.08053
#>   N.pout          87.75897         90.44604         92.55355        94.20656
#>   Herring        203.06648        213.77401        223.71259       232.93748
#>   Dab            134.85937        151.18693        167.02739       182.15907
#>   Whiting        451.49297        512.52167        571.12528       626.76575
#>   Sole           119.89639        135.42278        151.67885       168.68390
#>   Gurnard         79.92848         91.57662        103.98164       117.09977
#>   Plaice          63.38802         83.05301        105.72901       128.60332
#>   Haddock        626.93059        745.14693        871.39439      1004.14896
#>   Cod           8629.76619       9969.33992      11314.09158     12652.16288
#>   Saithe        2596.22152       3168.35366       3789.08700      4452.84382
#>          Age
#> Species   7.83673469387755 8.3265306122449 8.81632653061224 9.30612244897959
#>   Sprat           28.15274        28.70970         29.21179         29.66483
#>   Sandeel         32.73761        33.28866         33.75079         34.15093
#>   N.pout          95.50303        96.51985         97.31745         97.94887
#>   Herring        241.48705       249.35624        256.59319        263.24873
#>   Dab            196.44220       209.79303        222.07049        233.35786
#>   Whiting        679.05410       727.83152        772.86010        814.42555
#>   Sole           186.37894       204.72395        223.60594        242.98621
#>   Gurnard        130.88098       145.27597        160.19208        175.58486
#>   Plaice         152.24409       177.87676        205.93081        236.48270
#>   Haddock       1141.76761      1282.84413       1425.90626       1569.33115
#>   Cod          13972.26537     15265.13811      16525.76018      17748.24865
#>   Saithe        5154.06474      5887.25281       6646.99017       7429.22335
#>          Age
#> Species   9.79591836734694 10.2857142857143 10.7755102040816 11.265306122449
#>   Sprat           30.07363         30.44249         30.77532        31.07564
#>   Sandeel         34.52578         34.87813         35.20932        35.52063
#>   N.pout          98.52951         99.09213         99.63731       100.16558
#>   Herring        269.36953        274.99857        280.17535       284.93394
#>   Dab            243.67169        252.92291        261.21037       268.63446
#>   Whiting        852.50176        887.07757        918.47405       946.98401
#>   Sole           262.72295        282.77179        303.01972       323.35570
#>   Gurnard        191.33618        207.40870        223.66366       240.08016
#>   Plaice         269.50703        304.93385        342.62697       382.43640
#>   Haddock       1712.33353       1853.28847       1991.82777      2126.76816
#>   Cod          18928.27640      20066.60377      21157.61456     22201.43471
#>   Saithe        8229.17992       9042.45612       9866.24266     10695.97252
#>          Age
#> Species   11.7551020408163 12.2448979591837 12.734693877551 13.2244897959184
#>   Sprat           31.34663         31.59114        31.81177         32.01085
#>   Sandeel         35.81326         36.08831        36.34686         36.58988
#>   N.pout         100.67747        101.17348       101.65411        102.11983
#>   Herring        289.29526        293.29090       296.95169        300.30568
#>   Dab            275.28513        281.24295       286.59511        291.43240
#>   Whiting        972.87245        996.21275      1017.09117       1035.76740
#>   Sole           343.76668        364.07938       384.24105        404.25129
#>   Gurnard        256.52534        272.93647       289.29444        305.45149
#>   Plaice         424.25543        467.84573       513.10739        559.78037
#>   Haddock       2257.46237       2383.99661      2505.34488       2621.22426
#>   Cod          23200.04651      24153.46005     25058.98375      25918.73986
#>   Saithe       11529.48626      12363.72712     13195.12656      14023.30521
#>          Age
#> Species   13.7142857142857 14.2040816326531 14.6938775510204 15.1836734693878
#>   Sprat           32.19048         32.35257         32.49882         32.63079
#>   Sandeel         36.81832         37.03304         37.23487         37.42458
#>   N.pout         102.57111        103.00840        103.43212        103.84270
#>   Herring        303.37858        306.19396        308.77339        311.13665
#>   Dab            295.80523        299.75819        303.33157        306.56183
#>   Whiting       1052.47378       1067.41784       1080.78558       1092.74327
#>   Sole           423.96963        443.30634        462.26898        480.86322
#>   Gurnard        321.37624        337.07193        352.44514        367.40952
#>   Plaice         607.76928        656.78592        706.78602        757.44662
#>   Haddock       2731.88121       2837.50600       2937.40174       3031.52336
#>   Cod          26735.04083      27510.08307      28245.63606      28940.39078
#>   Saithe       14844.18470      15656.02709      16458.93007      17250.11768
#>          Age
#> Species   15.6734693877551 16.1632653061224 16.6530612244898 17.1428571428571
#>   Sprat           32.74987         32.85731         32.95426         33.04174
#>   Sandeel         37.60291         37.77053         37.92809         38.07619
#>   N.pout         104.24054        104.62605        104.99960        105.36157
#>   Herring        313.30185        315.28560        317.10310        318.76828
#>   Dab            309.48191        312.12159        314.50780        316.66490
#>   Whiting       1103.43960       1113.00767       1121.56651       1129.22256
#>   Sole           498.97685        516.52457        533.52396        549.99215
#>   Gurnard        381.97606        396.15531        409.93006        423.18857
#>   Plaice         808.71050        860.41862        912.30546        964.36658
#>   Haddock       3120.20506       3203.76139       3282.48829       3356.63087
#>   Cod          29595.66687      30213.74455      30796.73876      31346.63355
#>   Saithe       18026.82602      18789.32067      19537.86058      20270.82481
#>          Age
#> Species   17.6326530612245 18.1224489795918 18.6122448979592 19.1020408163265
#>   Sprat           33.12068         33.19190         33.25617         33.31416
#>   Sandeel         38.21540         38.34625         38.46924         38.58486
#>   N.pout         105.71231        106.05217        106.38149        106.70060
#>   Herring        320.29391        321.69168        322.97231        324.14561
#>   Dab            318.61486        320.37759        321.97106        323.41152
#>   Whiting       1136.07104       1142.19714       1147.67706       1152.57895
#>   Sole           565.94574        581.37057        596.14934        610.29266
#>   Gurnard        435.93543        448.19040        459.97246        471.29988
#>   Plaice        1016.45941       1068.31832       1119.93824       1171.30958
#>   Haddock       3426.05615       3490.64861       3550.74125       3606.64756
#>   Cod          31865.31356      32354.54602      32816.00130      33251.25738
#>   Saithe       20985.27021      21681.62147      22360.33701      23021.86346
#>          Age
#> Species   19.5918367346939 20.0816326530612 20.5714285714286 21.0612244897959
#>   Sprat           33.36649         33.41370         33.45630         33.49474
#>   Sandeel         38.69353         38.79568         38.89169         38.98195
#>   N.pout         107.00981        107.30943        107.59976        107.88108
#>   Herring        325.22059        326.20547        327.10782        327.93455
#>   Dab            324.71367        325.89078        326.95487        327.91678
#>   Whiting       1156.96378       1160.88610       1164.51509       1168.08148
#>   Sole           623.82786        636.78107        649.17731        661.04053
#>   Gurnard        482.18662        492.56628        502.42374        511.78557
#>   Plaice        1222.19201       1272.46395       1322.13271       1371.20556
#>   Haddock       3658.65901       3707.04704       3752.06405       3793.94488
#>   Cod          33661.30057      34045.99318      34406.77809      34745.14156
#>   Saithe       23666.36780      24291.17835      24895.96440      25481.36797
#>          Age
#> Species   21.5510204081633 22.0408163265306 22.530612244898 23.0204081632653
#>   Sprat           33.52943         33.56073        33.58897         33.61445
#>   Sandeel         39.06678         39.14652        39.22148         39.29193
#>   N.pout         108.15368        108.41783       108.67378        108.92180
#>   Herring        328.69199        329.38595       330.02175        330.60427
#>   Dab            328.78632        329.57238       330.28295        330.92529
#>   Whiting       1171.59045       1175.04292      1178.43980       1181.78200
#>   Sole           672.39365        683.25396       693.59814        703.37440
#>   Gurnard        520.67668        529.12074       537.14030        544.75663
#>   Plaice        1419.60848       1467.11540      1513.72789       1559.46268
#>   Haddock       3832.90804       3869.15680      3902.88025       3934.25433
#>   Cod          35062.47702      35360.09040     35639.20785      35900.97884
#>   Saithe       26048.01037      26596.49258     27127.39674      27641.28619
#>          Age
#> Species   23.5102040816327          24
#>   Sprat           33.63744    33.65819
#>   Sandeel         39.35816    39.42041
#>   N.pout         109.16212   109.39499
#>   Herring        331.13797   331.62695
#>   Dab            331.50596   332.03087
#>   Whiting       1185.07039  1188.30583
#>   Sole           712.61617   721.35449
#>   Gurnard        551.98999   558.85966
#>   Plaice        1604.33633  1648.36504
#>   Haddock       3963.44271  3990.60089
#>   Cod          36146.48128 36376.72636
#>   Saithe       28138.36459 28617.35431