Alias for getAvailEnergy

An alias provided for backward compatibility with mizer version <= 1.0

getPhiPrey(object, n, n_pp)

Arguments

object
An MizerParams object

n
A matrix of species abundances (species x size)

n_pp
A vector of the plankton abundance by size

Value

A two dimensional array (predator species x predator size)

Details

This function is used by the project method for performing simulations.

The function returns values also for sizes outside the size-range of the species. These values should not be trusted, as they are meaningless.

See also

Examples

data(NS_species_params_gears)
data(inter)
params <- MizerParams(NS_species_params_gears, inter)

#> Note: No sel_func column in species data frame. Setting selectivity to be 'knife_edge' for all species.

#> Note: No knife_edge_size column in species data frame. Setting knife edge selectivity equal to w_mat.

#> Note: No h provided for some species, so using f0 and k_vb to calculate it.

#> Note: No gamma provided for some species, so using f0, h, beta, sigma, lambda and kappa to calculate it.

#> Note: Using z0 = z0pre * w_inf ^ z0exp for missing z0 values.

#> Note: No ks column in species data frame so using ks = h * 0.2.

# With constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
n <- sim@n[21,,]
n_pp <- sim@n_pp[21,]
getAvailEnergy(params,n,n_pp)

#>          w
#> sp               0.001      0.00119      0.00142       0.0017      0.00203
#>   Sprat   2.345556e+12 2.290874e+12 2.237458e+12 2.185276e+12 2.134300e+12
#>   Sandeel 7.521293e+12 7.346038e+12 7.174856e+12 7.007653e+12 6.844333e+12
#>   N.pout  1.496466e+12 1.457053e+12 1.418548e+12 1.380937e+12 1.344207e+12
#>   Herring 1.279054e+13 1.249210e+13 1.220057e+13 1.191578e+13 1.163756e+13
#>   Dab     2.662750e+12 2.596483e+12 2.531597e+12 2.468067e+12 2.405866e+12
#>   Whiting 1.496466e+12 1.457052e+12 1.418545e+12 1.380932e+12 1.344198e+12
#>   Sole    2.938475e+12 2.866457e+12 2.795937e+12 2.726882e+12 2.659261e+12
#>   Gurnard 2.662832e+12 2.597118e+12 2.532762e+12 2.469738e+12 2.408019e+12
#>   Plaice  2.063133e+12 2.010971e+12 1.959893e+12 1.909881e+12 1.860922e+12
#>   Haddock 3.448426e+12 3.364585e+12 3.282510e+12 3.202160e+12 3.123495e+12
#>   Cod     1.539627e+12 1.499987e+12 1.461180e+12 1.423195e+12 1.386026e+12
#>   Saithe  1.204958e+12 1.173417e+12 1.142554e+12 1.112365e+12 1.082847e+12
#>          w
#> sp             0.00242      0.00289      0.00345      0.00411      0.00491
#>   Sprat   2.084499e+12 2.035844e+12 1.988306e+12 1.941856e+12 1.896467e+12
#>   Sandeel 6.684804e+12 6.528977e+12 6.376763e+12 6.228075e+12 6.082829e+12
#>   N.pout  1.308341e+12 1.273318e+12 1.239118e+12 1.205716e+12 1.173086e+12
#>   Herring 1.136576e+13 1.110022e+13 1.084080e+13 1.058734e+13 1.033970e+13
#>   Dab     2.344972e+12 2.285362e+12 2.227014e+12 2.169908e+12 2.114023e+12
#>   Whiting 1.308326e+12 1.273297e+12 1.239087e+12 1.205673e+12 1.173027e+12
#>   Sole    2.593046e+12 2.528208e+12 2.464722e+12 2.402564e+12 2.341709e+12
#>   Gurnard 2.347581e+12 2.288401e+12 2.230458e+12 2.173731e+12 2.118201e+12
#>   Plaice  1.813001e+12 1.766104e+12 1.720219e+12 1.675334e+12 1.631436e+12
#>   Haddock 3.046480e+12 2.971078e+12 2.897257e+12 2.824983e+12 2.754226e+12
#>   Cod     1.349666e+12 1.314108e+12 1.279347e+12 1.245377e+12 1.212193e+12
#>   Saithe  1.053998e+12 1.025816e+12 9.983018e+11 9.714527e+11 9.452661e+11
#>          w
#> sp             0.00586      0.00699      0.00834      0.00995       0.0119
#>   Sprat   1.852110e+12 1.808757e+12 1.766382e+12 1.724958e+12 1.684457e+12
#>   Sandeel 5.940941e+12 5.802330e+12 5.666917e+12 5.534622e+12 5.405369e+12
#>   N.pout  1.141198e+12 1.110021e+12 1.079522e+12 1.049667e+12 1.020421e+12
#>   Herring 1.009774e+13 9.861320e+12 9.630308e+12 9.404571e+12 9.183980e+12
#>   Dab     2.059340e+12 2.005837e+12 1.953496e+12 1.902296e+12 1.852216e+12
#>   Whiting 1.141119e+12 1.109917e+12 1.079386e+12 1.049491e+12 1.020195e+12
#>   Sole    2.282137e+12 2.223824e+12 2.166752e+12 2.110899e+12 2.056246e+12
#>   Gurnard 2.063849e+12 2.010659e+12 1.958613e+12 1.907693e+12 1.857882e+12
#>   Plaice  1.588512e+12 1.546548e+12 1.505532e+12 1.465446e+12 1.426273e+12
#>   Haddock 2.684957e+12 2.617149e+12 2.550773e+12 2.485805e+12 2.422220e+12
#>   Cod     1.179787e+12 1.148152e+12 1.117277e+12 1.087148e+12 1.057750e+12
#>   Saithe  9.197369e+11 8.948562e+11 8.706105e+11 8.469807e+11 8.239420e+11
#>          w
#> sp              0.0142       0.0169       0.0202       0.0241       0.0288
#>   Sprat   1.644853e+12 1.606121e+12 1.568234e+12 1.531166e+12 1.494893e+12
#>   Sandeel 5.279083e+12 5.155689e+12 5.035115e+12 4.917290e+12 4.802142e+12
#>   N.pout  9.917511e+11 9.636246e+11 9.360128e+11 9.088909e+11 8.822394e+11
#>   Herring 8.968409e+12 8.757734e+12 8.551836e+12 8.350597e+12 8.153901e+12
#>   Dab     1.803234e+12 1.755328e+12 1.708473e+12 1.662646e+12 1.617819e+12
#>   Whiting 9.914638e+11 9.632621e+11 9.355587e+11 9.083259e+11 8.815407e+11
#>   Sole    2.002773e+12 1.950460e+12 1.899286e+12 1.849230e+12 1.800270e+12
#>   Gurnard 1.809163e+12 1.761518e+12 1.714927e+12 1.669371e+12 1.624827e+12
#>   Plaice  1.387995e+12 1.350590e+12 1.314034e+12 1.278302e+12 1.243365e+12
#>   Haddock 2.359991e+12 2.299096e+12 2.239511e+12 2.181210e+12 2.124172e+12
#>   Cod     1.029061e+12 1.001057e+12 9.737093e+11 9.469857e+11 9.208505e+11
#>   Saithe  8.014630e+11 7.795069e+11 7.580318e+11 7.369923e+11 7.163410e+11
#>          w
#> sp              0.0343       0.0409       0.0489       0.0583       0.0696
#>   Sprat   1.459391e+12 1.424634e+12 1.390602e+12 1.357270e+12 1.324619e+12
#>   Sandeel 4.689605e+12 4.579609e+12 4.472089e+12 4.366981e+12 4.264220e+12
#>   N.pout  8.560450e+11 8.303016e+11 8.050108e+11 7.801825e+11 7.558349e+11
#>   Herring 7.961638e+12 7.773698e+12 7.589974e+12 7.410362e+12 7.234761e+12
#>   Dab     1.573967e+12 1.531062e+12 1.489076e+12 1.447979e+12 1.407744e+12
#>   Whiting 8.551860e+11 8.292516e+11 8.037345e+11 7.786394e+11 7.539793e+11
#>   Sole    1.752382e+12 1.705543e+12 1.659726e+12 1.614905e+12 1.571053e+12
#>   Gurnard 1.581274e+12 1.538688e+12 1.497042e+12 1.456310e+12 1.416465e+12
#>   Plaice  1.209193e+12 1.175756e+12 1.143021e+12 1.110955e+12 1.079526e+12
#>   Haddock 2.068371e+12 2.013783e+12 1.960383e+12 1.908148e+12 1.857051e+12
#>   Cod     8.952658e+11 8.701927e+11 8.455924e+11 8.214279e+11 7.976656e+11
#>   Saithe  6.960313e+11 6.760185e+11 6.562628e+11 6.367308e+11 6.173973e+11
#>          w
#> sp               0.083       0.0991        0.118        0.141        0.168
#>   Sprat   1.292628e+12 1.261279e+12 1.230554e+12 1.200438e+12 1.170916e+12
#>   Sandeel 4.163744e+12 4.065494e+12 3.969408e+12 3.875431e+12 3.783504e+12
#>   N.pout  7.319942e+11 7.086948e+11 6.859780e+11 6.638913e+11 6.424875e+11
#>   Herring 7.063073e+12 6.895201e+12 6.731052e+12 6.570534e+12 6.413558e+12
#>   Dab     1.368343e+12 1.329749e+12 1.291937e+12 1.254884e+12 1.218569e+12
#>   Whiting 7.297749e+11 7.060548e+11 6.828547e+11 6.602170e+11 6.381898e+11
#>   Sole    1.528139e+12 1.486135e+12 1.445010e+12 1.404734e+12 1.365277e+12
#>   Gurnard 1.377477e+12 1.339318e+12 1.301957e+12 1.265368e+12 1.229523e+12
#>   Plaice  1.048705e+12 1.018463e+12 9.887761e+11 9.596229e+11 9.309884e+11
#>   Haddock 1.807067e+12 1.758172e+12 1.710339e+12 1.663544e+12 1.617764e+12
#>   Cod     7.742767e+11 7.512386e+11 7.285364e+11 7.061633e+11 6.841216e+11
#>   Saithe  5.982460e+11 5.792706e+11 5.604745e+11 5.418706e+11 5.234803e+11
#>          w
#> sp               0.201         0.24        0.286        0.342        0.408
#>   Sprat   1.141975e+12 1.113604e+12 1.085793e+12 1.058533e+12 1.031816e+12
#>   Sandeel 3.693574e+12 3.605587e+12 3.519492e+12 3.435238e+12 3.352779e+12
#>   N.pout  6.218232e+11 6.019574e+11 5.829498e+11 5.648599e+11 5.477449e+11
#>   Herring 6.260040e+12 6.109894e+12 5.963039e+12 5.819397e+12 5.678890e+12
#>   Dab     1.182976e+12 1.148091e+12 1.113907e+12 1.080419e+12 1.047630e+12
#>   Whiting 6.168259e+11 5.961817e+11 5.763165e+11 5.572909e+11 5.391663e+11
#>   Sole    1.326609e+12 1.288703e+12 1.251533e+12 1.215074e+12 1.179307e+12
#>   Gurnard 1.194396e+12 1.159964e+12 1.126207e+12 1.093111e+12 1.060663e+12
#>   Plaice  9.028630e+11 8.752433e+11 8.481331e+11 8.215427e+11 7.954901e+11
#>   Haddock 1.572974e+12 1.529153e+12 1.486282e+12 1.444343e+12 1.403321e+12
#>   Cod     6.624227e+11 6.410872e+11 6.201446e+11 5.996321e+11 5.795943e+11
#>   Saithe  5.053326e+11 4.874626e+11 4.699102e+11 4.527184e+11 4.359327e+11
#>          w
#> sp               0.486         0.58        0.693        0.827        0.987
#>   Sprat   1.005634e+12 9.799800e+11 9.548476e+11 9.302298e+11 9.061193e+11
#>   Sandeel 3.272067e+12 3.193058e+12 3.115710e+12 3.039982e+12 2.965836e+12
#>   N.pout  5.316590e+11 5.166518e+11 5.027679e+11 4.900461e+11 4.785196e+11
#>   Herring 5.541444e+12 5.406987e+12 5.275449e+12 5.146761e+12 5.020858e+12
#>   Dab     1.015546e+12 9.841798e+11 9.535511e+11 9.236843e+11 8.946102e+11
#>   Whiting 5.220038e+11 5.058634e+11 4.908040e+11 4.768828e+11 4.641553e+11
#>   Sole    1.144214e+12 1.109782e+12 1.076004e+12 1.042875e+12 1.010397e+12
#>   Gurnard 1.028856e+12 9.976906e+11 9.671701e+11 9.373054e+11 9.081126e+11
#>   Plaice  7.699998e+11 7.451031e+11 7.208379e+11 6.972472e+11 6.743796e+11
#>   Haddock 1.363204e+12 1.323983e+12 1.285655e+12 1.248218e+12 1.211678e+12
#>   Cod     5.600819e+11 5.411505e+11 5.228598e+11 5.052719e+11 4.884506e+11
#>   Saithe  4.195988e+11 4.037623e+11 3.884674e+11 3.737561e+11 3.596674e+11
#>          w
#> sp                1.18          1.4         1.68            2         2.39
#>   Sprat   8.825089e+11 8.593913e+11 8.367597e+11 8.146078e+11 7.929310e+11
#>   Sandeel 2.893234e+12 2.822142e+12 2.752526e+12 2.684355e+12 2.617600e+12
#>   N.pout  4.682157e+11 4.591567e+11 4.513608e+11 4.448432e+11 4.396187e+11
#>   Herring 4.897676e+12 4.777153e+12 4.659230e+12 4.543849e+12 4.430954e+12
#>   Dab     8.663654e+11 8.389922e+11 8.125385e+11 7.870577e+11 7.626087e+11
#>   Whiting 4.526761e+11 4.424993e+11 4.336795e+11 4.262736e+11 4.203426e+11
#>   Sole    9.785778e+11 9.474300e+11 9.169719e+11 8.872274e+11 8.582262e+11
#>   Gurnard 8.796144e+11 8.518385e+11 8.248188e+11 7.985936e+11 7.732066e+11
#>   Plaice  6.522878e+11 6.310283e+11 6.106606e+11 5.912466e+11 5.728502e+11
#>   Haddock 1.176042e+12 1.141327e+12 1.107551e+12 1.074739e+12 1.042921e+12
#>   Cod     4.724607e+11 4.573667e+11 4.432324e+11 4.301204e+11 4.180923e+11
#>   Saithe  3.462368e+11 3.334959e+11 3.214721e+11 3.101881e+11 2.996620e+11
#>          w
#> sp                2.85          3.4         4.06         4.84         5.78
#>   Sprat   7.717269e+11 7.509953e+11 7.307391e+11 7.109641e+11 6.916779e+11
#>   Sandeel 2.552233e+12 2.488227e+12 2.425558e+12 2.364201e+12 2.304134e+12
#>   N.pout  4.357032e+11 4.331186e+11 4.319076e+11 4.321639e+11 4.340254e+11
#>   Herring 4.320491e+12 4.212407e+12 4.106654e+12 4.003183e+12 3.901948e+12
#>   Dab     7.392559e+11 7.170700e+11 6.961332e+11 6.765493e+11 6.584416e+11
#>   Whiting 4.159544e+11 4.131894e+11 4.121625e+11 4.130684e+11 4.161693e+11
#>   Sole    8.300032e+11 8.025987e+11 7.760597e+11 7.504414e+11 7.258068e+11
#>   Gurnard 7.487052e+11 7.251409e+11 7.025699e+11 6.810544e+11 6.606624e+11
#>   Plaice  5.555373e+11 5.393755e+11 5.244374e+11 5.108056e+11 4.985730e+11
#>   Haddock 1.012134e+12 9.824194e+11 9.538265e+11 9.264182e+11 9.002681e+11
#>   Cod     4.072080e+11 3.975266e+11 3.891091e+11 3.820219e+11 3.763390e+11
#>   Saithe  2.899077e+11 2.809346e+11 2.727487e+11 2.653534e+11 2.587503e+11
#>          w
#> sp                 6.9         8.23         9.82         11.7           14
#>   Sprat   6.728890e+11 6.546052e+11 6.368307e+11 6.195647e+11 6.027982e+11
#>   Sandeel 2.245336e+12 2.187786e+12 2.131463e+12 2.076347e+12 2.022419e+12
#>   N.pout  4.375494e+11 4.428000e+11 4.498827e+11 4.589356e+11 4.691529e+11
#>   Herring 3.802905e+12 3.706011e+12 3.611226e+12 3.518513e+12 3.427832e+12
#>   Dab     6.419068e+11 6.270468e+11 6.139828e+11 6.028577e+11 5.936925e+11
#>   Whiting 4.216035e+11 4.295178e+11 4.401233e+11 4.537006e+11 4.697570e+11
#>   Sole    7.022170e+11 6.797373e+11 6.584409e+11 6.384094e+11 6.196802e+11
#>   Gurnard 6.414574e+11 6.235047e+11 6.068740e+11 5.916402e+11 5.778243e+11
#>   Plaice  4.878213e+11 4.786363e+11 4.711167e+11 4.653770e+11 4.614618e+11
#>   Haddock 8.754394e+11 8.519989e+11 8.300234e+11 8.095967e+11 7.907224e+11
#>   Cod     3.721311e+11 3.694756e+11 3.684634e+11 3.692033e+11 3.717814e+11
#>   Saithe  2.529390e+11 2.479191e+11 2.436919e+11 2.402615e+11 2.375964e+11
#>          w
#> sp                16.7         19.9         23.8         28.4         33.8
#>   Sprat   5.865132e+11 5.706811e+11 5.552635e+11 5.402130e+11 5.254762e+11
#>   Sandeel 1.969658e+12 1.918043e+12 1.867554e+12 1.818167e+12 1.769861e+12
#>   N.pout  4.816054e+11 4.964350e+11 5.135015e+11 5.329741e+11 5.550632e+11
#>   Herring 3.339155e+12 3.252451e+12 3.167684e+12 3.084825e+12 3.003850e+12
#>   Dab     5.869228e+11 5.827163e+11 5.808587e+11 5.812636e+11 5.840059e+11
#>   Whiting 4.898969e+11 5.143982e+11 5.420141e+11 5.720518e+11 6.044216e+11
#>   Sole    6.024333e+11 5.867681e+11 5.726772e+11 5.601900e+11 5.493861e+11
#>   Gurnard 5.655756e+11 5.549761e+11 5.460473e+11 5.388352e+11 5.334116e+11
#>   Plaice  4.596741e+11 4.601699e+11 4.628732e+11 4.677820e+11 4.749930e+11
#>   Haddock 7.735767e+11 7.582514e+11 7.448017e+11 7.333072e+11 7.238568e+11
#>   Cod     3.764167e+11 3.832867e+11 3.924901e+11 4.041695e+11 4.185171e+11
#>   Saithe  2.357281e+11 2.346652e+11 2.344163e+11 2.349952e+11 2.364157e+11
#>          w
#> sp                40.4         48.2         57.5         68.7         81.9
#>   Sprat   5.109965e+11 4.967186e+11 4.825911e+11 4.685712e+11 4.546264e+11
#>   Sandeel 1.722610e+12 1.676389e+12 1.631171e+12 1.586927e+12 1.543628e+12
#>   N.pout  5.792247e+11 6.049903e+11 6.325395e+11 6.618228e+11 6.927769e+11
#>   Herring 2.924735e+12 2.847464e+12 2.772021e+12 2.698394e+12 2.626576e+12
#>   Dab     5.892002e+11 5.969613e+11 6.075259e+11 6.210016e+11 6.374706e+11
#>   Whiting 6.389481e+11 6.754668e+11 7.143989e+11 7.556649e+11 7.991704e+11
#>   Sole    5.403635e+11 5.332227e+11 5.280876e+11 5.250542e+11 5.242179e+11
#>   Gurnard 5.298395e+11 5.281824e+11 5.285371e+11 5.309754e+11 5.355670e+11
#>   Plaice  4.846286e+11 4.968084e+11 5.116968e+11 5.293830e+11 5.499306e+11
#>   Haddock 7.164947e+11 7.112708e+11 7.083293e+11 7.077738e+11 7.097010e+11
#>   Cod     4.357289e+11 4.559988e+11 4.795481e+11 5.065412e+11 5.370960e+11
#>   Saithe  2.386801e+11 2.417848e+11 2.457264e+11 2.504906e+11 2.560532e+11
#>          w
#> sp                97.8          117          139          166          198
#>   Sprat   4.407364e+11 4.268938e+11 4.131037e+11 3.993823e+11 3.857552e+11
#>   Sandeel 1.501245e+12 1.459745e+12 1.419097e+12 1.379271e+12 1.340235e+12
#>   N.pout  7.256244e+11 7.617895e+11 8.015560e+11 8.419883e+11 8.803418e+11
#>   Herring 2.556563e+12 2.488366e+12 2.421991e+12 2.357418e+12 2.294639e+12
#>   Dab     6.570232e+11 6.799330e+11 7.061957e+11 7.352104e+11 7.664234e+11
#>   Whiting 8.451083e+11 8.950921e+11 9.491336e+11 1.003219e+12 1.053834e+12
#>   Sole    5.256651e+11 5.295205e+11 5.358301e+11 5.444989e+11 5.554363e+11
#>   Gurnard 5.423922e+11 5.515965e+11 5.632504e+11 5.772228e+11 5.933777e+11
#>   Plaice  5.733890e+11 5.998983e+11 6.294142e+11 6.615144e+11 6.957536e+11
#>   Haddock 7.142369e+11 7.216320e+11 7.320141e+11 7.451652e+11 7.608834e+11
#>   Cod     5.712867e+11 6.092041e+11 6.507821e+11 6.956269e+11 7.432336e+11
#>   Saithe  2.623855e+11 2.694730e+11 2.772874e+11 2.857515e+11 2.947832e+11
#>          w
#> sp                 237          282          337          402          480
#>   Sprat   3.722558e+11 3.589228e+11 3.457986e+11 3.329280e+11 3.203566e+11
#>   Sandeel 1.301960e+12 1.264418e+12 1.227583e+12 1.191435e+12 1.155953e+12
#>   N.pout  9.159114e+11 9.484389e+11 9.777772e+11 1.003679e+12 1.025724e+12
#>   Herring 2.233666e+12 2.174520e+12 2.117228e+12 2.061821e+12 2.008331e+12
#>   Dab     7.996143e+11 8.346208e+11 8.712406e+11 9.091947e+11 9.481194e+11
#>   Whiting 1.100091e+12 1.141782e+12 1.178821e+12 1.210985e+12 1.237841e+12
#>   Sole    5.686183e+11 5.840194e+11 6.015849e+11 6.212188e+11 6.427798e+11
#>   Gurnard 6.116728e+11 6.320686e+11 6.545015e+11 6.788685e+11 7.050191e+11
#>   Plaice  7.318309e+11 7.694394e+11 8.082120e+11 8.477053e+11 8.873989e+11
#>   Haddock 7.791689e+11 8.000468e+11 8.235360e+11 8.496216e+11 8.782401e+11
#>   Cod     7.931070e+11 8.446750e+11 8.972720e+11 9.501426e+11 1.002457e+12
#>   Saithe  3.043239e+11 3.143177e+11 3.247029e+11 3.354025e+11 3.463132e+11
#>          w
#> sp                 573          683          816          973         1160
#>   Sprat   3.081299e+11 2.962934e+11 2.848910e+11 2.739660e+11 2.635597e+11
#>   Sandeel 1.121124e+12 1.086939e+12 1.053395e+12 1.020496e+12 9.882504e+11
#>   N.pout  1.043346e+12 1.055941e+12 1.062982e+12 1.064056e+12 1.058909e+12
#>   Herring 1.956795e+12 1.907252e+12 1.859745e+12 1.814325e+12 1.771042e+12
#>   Dab     9.875728e+11 1.027055e+12 1.066032e+12 1.103944e+12 1.140214e+12
#>   Whiting 1.258781e+12 1.273153e+12 1.280392e+12 1.280058e+12 1.271873e+12
#>   Sole    6.660798e+11 6.908876e+11 7.169321e+11 7.439038e+11 7.714573e+11
#>   Gurnard 7.327539e+11 7.618276e+11 7.919535e+11 8.228049e+11 8.540176e+11
#>   Plaice  9.267055e+11 9.649899e+11 1.001591e+12 1.035837e+12 1.067058e+12
#>   Haddock 9.092764e+11 9.425702e+11 9.779230e+11 1.015097e+12 1.053817e+12
#>   Cod     1.053335e+12 1.101871e+12 1.147162e+12 1.188322e+12 1.224501e+12
#>   Saithe  3.572942e+11 3.681571e+11 3.786589e+11 3.884982e+11 3.973198e+11
#>          w
#> sp                1390         1650         1970         2360         2810
#>   Sprat   2.537126e+11 2.444631e+11 2.358486e+11 2.279050e+11 2.206668e+11
#>   Sandeel 9.566780e+11 9.258046e+11 8.956657e+11 8.663064e+11 8.377822e+11
#>   N.pout  1.047425e+12 1.029659e+12 1.005837e+12 9.763764e+11 9.418040e+11
#>   Herring 1.729952e+12 1.691115e+12 1.654593e+12 1.620451e+12 1.588755e+12
#>   Dab     1.174262e+12 1.205520e+12 1.233442e+12 1.257515e+12 1.277252e+12
#>   Whiting 1.255723e+12 1.231710e+12 1.200181e+12 1.161654e+12 1.116756e+12
#>   Sole    7.992144e+11 8.267702e+11 8.536992e+11 8.795609e+11 9.039045e+11
#>   Gurnard 8.851922e+11 9.158996e+11 9.456879e+11 9.740871e+11 1.000614e+12
#>   Plaice  1.094609e+12 1.117879e+12 1.136314e+12 1.149426e+12 1.156800e+12
#>   Haddock 1.093765e+12 1.134587e+12 1.175891e+12 1.217250e+12 1.258202e+12
#>   Cod     1.254889e+12 1.278732e+12 1.295343e+12 1.304114e+12 1.304535e+12
#>   Saithe  4.047227e+11 4.102812e+11 4.135733e+11 4.142228e+11 4.119398e+11
#>          w
#> sp                3350         4000         4780         5700         6800
#>   Sprat   2.141675e+11 2.084402e+11 2.035173e+11 1.994323e+11 1.962195e+11
#>   Sandeel 8.101601e+11 7.835189e+11 7.579509e+11 7.335626e+11 7.104757e+11
#>   N.pout  9.028266e+11 8.603134e+11 8.152436e+11 7.686411e+11 7.215452e+11
#>   Herring 1.559572e+12 1.532971e+12 1.509018e+12 1.487779e+12 1.469320e+12
#>   Dab     1.292228e+12 1.302090e+12 1.306565e+12 1.305468e+12 1.298708e+12
#>   Whiting 1.066320e+12 1.011380e+12 9.531070e+11 8.927554e+11 8.316187e+11
#>   Sole    9.262816e+11 9.462564e+11 9.634145e+11 9.773728e+11 9.877888e+11
#>   Gurnard 1.024787e+12 1.046138e+12 1.064222e+12 1.078628e+12 1.088991e+12
#>   Plaice  1.158119e+12 1.153173e+12 1.141870e+12 1.124242e+12 1.100451e+12
#>   Haddock 1.298263e+12 1.336935e+12 1.373715e+12 1.408095e+12 1.439578e+12
#>   Cod     1.296232e+12 1.279000e+12 1.252847e+12 1.218027e+12 1.175071e+12
#>   Saithe  4.065632e+11 3.980974e+11 3.867346e+11 3.728621e+11 3.570517e+11
#>          w
#> sp                8120         9690        11600        13800        16500
#>   Sprat   1.939161e+11 1.925631e+11 1.922072e+11 1.929032e+11 1.947167e+11
#>   Sandeel 6.888287e+11 6.687775e+11 6.504966e+11 6.341797e+11 6.200403e+11
#>   N.pout  6.749351e+11 6.296976e+11 5.866212e+11 5.463706e+11 5.094501e+11
#>   Herring 1.453698e+12 1.440971e+12 1.431187e+12 1.424389e+12 1.420610e+12
#>   Dab     1.286292e+12 1.268324e+12 1.245006e+12 1.216634e+12 1.183588e+12
#>   Whiting 7.709552e+11 7.119399e+11 6.556362e+11 6.029388e+11 5.545238e+11
#>   Sole    9.943698e+11 9.968812e+11 9.951536e+11 9.890884e+11 9.786619e+11
#>   Gurnard 1.095006e+12 1.096431e+12 1.093105e+12 1.084953e+12 1.071990e+12
#>   Plaice  1.070792e+12 1.035684e+12 9.956718e+11 9.514094e+11 9.036441e+11
#>   Haddock 1.467686e+12 1.491965e+12 1.512002e+12 1.527434e+12 1.537959e+12
#>   Cod     1.124797e+12 1.068313e+12 1.006984e+12 9.423871e+11 8.762369e+11
#>   Saithe  3.400217e+11 3.225866e+11 3.055960e+11 2.899268e+11 2.762166e+11
#>          w
#> sp               19600        23400        28000        33400        39900
#>   Sprat   1.977273e+11 2.020331e+11 2.077550e+11 2.150428e+11 2.240810e+11
#>   Sandeel 6.083107e+11 5.992416e+11 5.931000e+11 5.901657e+11 5.907274e+11
#>   N.pout  4.761393e+11 4.465254e+11 4.203286e+11 3.973216e+11 3.772731e+11
#>   Herring 1.419874e+12 1.422194e+12 1.427568e+12 1.435982e+12 1.447408e+12
#>   Dab     1.146322e+12 1.105346e+12 1.061230e+12 1.014595e+12 9.661093e+11
#>   Whiting 5.107976e+11 4.719347e+11 4.376997e+11 4.079064e+11 3.824049e+11
#>   Sole    9.639268e+11 9.450119e+11 9.221227e+11 8.955403e+11 8.656164e+11
#>   Gurnard 1.054323e+12 1.032152e+12 1.005765e+12 9.755429e+11 9.419486e+11
#>   Plaice  8.531948e+11 8.009253e+11 7.477222e+11 6.944729e+11 6.420388e+11
#>   Haddock 1.543335e+12 1.543387e+12 1.538019e+12 1.527224e+12 1.511093e+12
#>   Cod     8.103218e+11 7.463842e+11 6.860266e+11 6.306387e+11 5.813853e+11
#>   Saithe  2.650448e+11 2.567931e+11 2.516145e+11 2.494472e+11 2.500467e+11