Get available energy

This is deprecated and is no longer used by the mizer project() method. Calculates the amount \(E_{a,i}(w)\) of food exposed to each predator as a function of predator size.

Usage

getPhiPrey(object, n, n_pp, ...)

Arguments

object

An MizerParams object

n

A matrix of species abundances (species x size)

n_pp

A vector of the background abundance by size

...

Other arguments (currently unused)

Value

A two dimensional array (predator species x predator size)

See also

Examples

# \donttest{
params <-  NS_params
sim <- project(params, t_max = 20, effort = 0.5)
n <- sim@n[21,,]
n_pp <- sim@n_pp[21,]
getPhiPrey(params,n,n_pp)
#> Warning: `getPhiPrey()` was deprecated in mizer 2.0.0.
#> ℹ Please use `newMultispeciesParams()` instead.
#>          w
#> sp               0.001      0.00119      0.00142       0.0017      0.00203
#>   Sprat   2.351401e+12 2.296587e+12 2.243044e+12 2.190740e+12 2.139645e+12
#>   Sandeel 7.542075e+12 7.366344e+12 7.194698e+12 7.027041e+12 6.863280e+12
#>   N.pout  1.507852e+12 1.468295e+12 1.429562e+12 1.391632e+12 1.354483e+12
#>   Herring 1.280071e+13 1.250988e+13 1.222455e+13 1.194476e+13 1.167054e+13
#>   Dab     2.677115e+12 2.610901e+12 2.546062e+12 2.482566e+12 2.420385e+12
#>   Whiting 1.507852e+12 1.468298e+12 1.429571e+12 1.391648e+12 1.354509e+12
#>   Sole    2.952370e+12 2.880395e+12 2.809923e+12 2.740921e+12 2.673354e+12
#>   Gurnard 2.674783e+12 2.609164e+12 2.544908e+12 2.481982e+12 2.420359e+12
#>   Plaice  2.074646e+12 2.022596e+12 1.971619e+12 1.921695e+12 1.872802e+12
#>   Haddock 3.463425e+12 3.379590e+12 3.297529e+12 3.217197e+12 3.138554e+12
#>   Cod     1.548712e+12 1.509212e+12 1.470530e+12 1.432651e+12 1.395560e+12
#>   Saithe  1.212534e+12 1.181125e+12 1.150375e+12 1.120270e+12 1.090799e+12
#>          w
#> sp             0.00242      0.00289      0.00345      0.00411      0.00491
#>   Sprat   2.089731e+12 2.040966e+12 1.993323e+12 1.946774e+12 1.901290e+12
#>   Sandeel 6.703321e+12 6.547076e+12 6.394455e+12 6.245372e+12 6.099743e+12
#>   N.pout  1.318093e+12 1.282438e+12 1.247495e+12 1.213241e+12 1.179651e+12
#>   Herring 1.140189e+13 1.113879e+13 1.088120e+13 1.062909e+13 1.038238e+13
#>   Dab     2.359489e+12 2.299853e+12 2.241449e+12 2.184250e+12 2.128232e+12
#>   Whiting 1.318132e+12 1.282494e+12 1.247574e+12 1.213348e+12 1.179795e+12
#>   Sole    2.607190e+12 2.542397e+12 2.478946e+12 2.416806e+12 2.355950e+12
#>   Gurnard 2.360008e+12 2.300904e+12 2.243019e+12 2.186327e+12 2.130804e+12
#>   Plaice  1.824921e+12 1.778031e+12 1.732115e+12 1.687152e+12 1.643125e+12
#>   Haddock 3.061559e+12 2.986173e+12 2.912358e+12 2.840079e+12 2.769300e+12
#>   Cod     1.359244e+12 1.323687e+12 1.288874e+12 1.254791e+12 1.221420e+12
#>   Saithe  1.061949e+12 1.033706e+12 1.006057e+12 9.789881e+11 9.524826e+11
#>          w
#> sp             0.00586      0.00699      0.00834      0.00995       0.0119
#>   Sprat   1.856845e+12 1.813409e+12 1.770958e+12 1.729463e+12 1.688899e+12
#>   Sandeel 5.957485e+12 5.818516e+12 5.682756e+12 5.550128e+12 5.420555e+12
#>   N.pout  1.146704e+12 1.114376e+12 1.082646e+12 1.051496e+12 1.020909e+12
#>   Herring 1.014103e+13 9.904939e+12 9.674038e+12 9.448234e+12 9.227435e+12
#>   Dab     2.073371e+12 2.019640e+12 1.967018e+12 1.915481e+12 1.865006e+12
#>   Whiting 1.146892e+12 1.114619e+12 1.082956e+12 1.051885e+12 1.021393e+12
#>   Sole    2.296350e+12 2.237979e+12 2.180812e+12 2.124822e+12 2.069985e+12
#>   Gurnard 2.076425e+12 2.023165e+12 1.971001e+12 1.919909e+12 1.869865e+12
#>   Plaice  1.600013e+12 1.557799e+12 1.516463e+12 1.475988e+12 1.436355e+12
#>   Haddock 2.699986e+12 2.632106e+12 2.565627e+12 2.500517e+12 2.436746e+12
#>   Cod     1.188747e+12 1.156753e+12 1.125421e+12 1.094734e+12 1.064672e+12
#>   Saithe  9.265242e+11 9.010950e+11 8.761753e+11 8.517439e+11 8.277784e+11
#>          w
#> sp              0.0142       0.0169       0.0202       0.0241       0.0288
#>   Sprat   1.649239e+12 1.610458e+12 1.572530e+12 1.535430e+12 1.499134e+12
#>   Sandeel 5.293961e+12 5.170273e+12 5.049418e+12 4.931326e+12 4.815926e+12
#>   N.pout  9.908698e+11 9.613695e+11 9.324012e+11 9.039634e+11 8.760596e+11
#>   Herring 9.011547e+12 8.800471e+12 8.594108e+12 8.392357e+12 8.195118e+12
#>   Dab     1.815570e+12 1.767154e+12 1.719735e+12 1.673295e+12 1.627815e+12
#>   Whiting 9.914665e+11 9.620976e+11 9.332820e+11 9.050204e+11 8.773183e+11
#>   Sole    2.016275e+12 1.963668e+12 1.912141e+12 1.861668e+12 1.812226e+12
#>   Gurnard 1.820846e+12 1.772827e+12 1.725787e+12 1.679700e+12 1.634544e+12
#>   Plaice  1.397545e+12 1.359542e+12 1.322328e+12 1.285888e+12 1.250207e+12
#>   Haddock 2.374285e+12 2.313103e+12 2.253173e+12 2.194467e+12 2.136957e+12
#>   Cod     1.035216e+12 1.006346e+12 9.780433e+11 9.502885e+11 9.230637e+11
#>   Saithe  8.042553e+11 7.811502e+11 7.584390e+11 7.360984e+11 7.141067e+11
#>          w
#> sp              0.0343       0.0409       0.0489       0.0583       0.0696
#>   Sprat   1.463617e+12 1.428857e+12 1.394830e+12 1.361515e+12 1.328890e+12
#>   Sandeel 4.703150e+12 4.592932e+12 4.485205e+12 4.379905e+12 4.276969e+12
#>   N.pout  8.486987e+11 8.218954e+11 7.956700e+11 7.700486e+11 7.450629e+11
#>   Herring 8.002289e+12 7.813771e+12 7.629463e+12 7.449267e+12 7.273087e+12
#>   Dab     1.583279e+12 1.539670e+12 1.496976e+12 1.455186e+12 1.414290e+12
#>   Whiting 8.501867e+11 8.236422e+11 7.977069e+11 7.724084e+11 7.477796e+11
#>   Sole    1.763793e+12 1.716346e+12 1.669862e+12 1.624321e+12 1.579703e+12
#>   Gurnard 1.590295e+12 1.546931e+12 1.504430e+12 1.462770e+12 1.421930e+12
#>   Plaice  1.215273e+12 1.181074e+12 1.147602e+12 1.114853e+12 1.082826e+12
#>   Haddock 2.080617e+12 2.025420e+12 1.971342e+12 1.918358e+12 1.866446e+12
#>   Cod     8.963528e+11 8.701419e+11 8.444204e+11 8.191815e+11 7.944229e+11
#>   Saithe  6.924451e+11 6.710980e+11 6.500546e+11 6.293089e+11 6.088604e+11
#>          w
#> sp               0.083       0.0991        0.118        0.141        0.168
#>   Sprat   1.296937e+12 1.265635e+12 1.234968e+12 1.204918e+12 1.175473e+12
#>   Sandeel 4.176334e+12 4.077942e+12 3.981732e+12 3.887647e+12 3.795631e+12
#>   N.pout  7.207495e+11 6.971495e+11 6.743079e+11 6.522726e+11 6.310932e+11
#>   Herring 7.100826e+12 6.932390e+12 6.767687e+12 6.606627e+12 6.449121e+12
#>   Dab     1.374283e+12 1.335163e+12 1.296930e+12 1.259588e+12 1.223145e+12
#>   Whiting 7.238578e+11 7.006846e+11 6.783047e+11 6.567652e+11 6.361146e+11
#>   Sole    1.535987e+12 1.493157e+12 1.451198e+12 1.410094e+12 1.369835e+12
#>   Gurnard 1.381893e+12 1.342640e+12 1.304156e+12 1.266429e+12 1.229449e+12
#>   Plaice  1.051521e+12 1.020948e+12 9.911168e+11 9.620446e+11 9.337531e+11
#>   Haddock 1.815583e+12 1.765748e+12 1.716924e+12 1.669094e+12 1.622241e+12
#>   Cod     7.701476e+11 7.463644e+11 7.230879e+11 7.003391e+11 6.781451e+11
#>   Saithe  5.887145e+11 5.688819e+11 5.493788e+11 5.302263e+11 5.114496e+11
#>          w
#> sp               0.201         0.24        0.286        0.342        0.408
#>   Sprat   1.146617e+12 1.118338e+12 1.090626e+12 1.063470e+12 1.036861e+12
#>   Sandeel 3.705628e+12 3.617587e+12 3.531455e+12 3.447181e+12 3.364719e+12
#>   N.pout  6.108202e+11 5.915035e+11 5.731915e+11 5.559294e+11 5.397588e+11
#>   Herring 6.295082e+12 6.144427e+12 5.997072e+12 5.852938e+12 5.711946e+12
#>   Dab     1.187612e+12 1.153006e+12 1.119344e+12 1.086650e+12 1.054949e+12
#>   Whiting 6.164016e+11 5.976740e+11 5.799777e+11 5.633557e+11 5.478472e+11
#>   Sole    1.330412e+12 1.291817e+12 1.254050e+12 1.217108e+12 1.180995e+12
#>   Gurnard 1.193209e+12 1.157704e+12 1.122936e+12 1.088908e+12 1.055627e+12
#>   Plaice  9.062692e+11 8.796243e+11 8.538547e+11 8.290007e+11 8.051058e+11
#>   Haddock 1.576354e+12 1.531424e+12 1.487442e+12 1.444406e+12 1.402313e+12
#>   Cod     6.565386e+11 6.355579e+11 6.152459e+11 5.956493e+11 5.768177e+11
#>   Saithe  4.930772e+11 4.751402e+11 4.576712e+11 4.407036e+11 4.242708e+11
#>          w
#> sp               0.486         0.58        0.693        0.827        0.987
#>   Sprat   1.010792e+12 9.852540e+11 9.602399e+11 9.357423e+11 9.117535e+11
#>   Sandeel 3.284020e+12 3.205040e+12 3.127735e+12 3.052062e+12 2.977983e+12
#>   N.pout  5.247160e+11 5.108316e+11 4.981300e+11 4.866288e+11 4.763393e+11
#>   Herring 5.574021e+12 5.439088e+12 5.307076e+12 5.177915e+12 5.051538e+12
#>   Dab     1.024272e+12 9.946487e+11 9.661146e+11 9.387056e+11 9.124589e+11
#>   Whiting 5.334866e+11 5.203034e+11 5.083216e+11 4.975601e+11 4.880328e+11
#>   Sole    1.145719e+12 1.111289e+12 1.077719e+12 1.045025e+12 1.013228e+12
#>   Gurnard 1.023106e+12 9.913603e+11 9.604099e+11 9.302782e+11 9.009922e+11
#>   Plaice  7.822161e+11 7.603797e+11 7.396450e+11 7.200601e+11 7.016715e+11
#>   Haddock 1.361168e+12 1.320975e+12 1.281745e+12 1.243491e+12 1.206230e+12
#>   Cod     5.588027e+11 5.416561e+11 5.254297e+11 5.101735e+11 4.959349e+11
#>   Saithe  4.084053e+11 3.931381e+11 3.784982e+11 3.645120e+11 3.512029e+11
#>          w
#> sp                1.18          1.4         1.68            2         2.39
#>   Sprat   8.882657e+11 8.652706e+11 8.427595e+11 8.207235e+11 7.991533e+11
#>   Sandeel 2.905457e+12 2.834448e+12 2.764920e+12 2.696838e+12 2.630169e+12
#>   N.pout  4.672667e+11 4.594108e+11 4.527674e+11 4.473293e+11 4.430888e+11
#>   Herring 4.927879e+12 4.806875e+12 4.688465e+12 4.572590e+12 4.459193e+12
#>   Dab     8.874129e+11 8.636061e+11 8.410772e+11 8.198642e+11 8.000044e+11
#>   Whiting 4.797492e+11 4.727161e+11 4.669382e+11 4.624208e+11 4.591714e+11
#>   Sole    9.823498e+11 9.524156e+11 9.234529e+11 8.954904e+11 8.685587e+11
#>   Gurnard 8.725817e+11 8.450797e+11 8.185210e+11 7.929425e+11 7.683822e+11
#>   Plaice  6.845234e+11 6.686560e+11 6.541051e+11 6.409013e+11 6.290692e+11
#>   Haddock 1.169984e+12 1.134776e+12 1.100635e+12 1.067593e+12 1.035683e+12
#>   Cod     4.827575e+11 4.706808e+11 4.597388e+11 4.499603e+11 4.413681e+11
#>   Saithe  3.385911e+11 3.266932e+11 3.155220e+11 3.050866e+11 2.953922e+11
#>          w
#> sp                2.85          3.4         4.06         4.84         5.78
#>   Sprat   7.780400e+11 7.573744e+11 7.371479e+11 7.173522e+11 6.979791e+11
#>   Sandeel 2.564882e+12 2.500946e+12 2.438330e+12 2.377007e+12 2.316947e+12
#>   N.pout  4.400394e+11 4.381804e+11 4.375302e+11 4.381544e+11 4.401573e+11
#>   Herring 4.348220e+12 4.239618e+12 4.133337e+12 4.029330e+12 3.927550e+12
#>   Dab     7.815343e+11 7.644901e+11 7.489124e+11 7.348556e+11 7.223849e+11
#>   Whiting 4.572030e+11 4.565392e+11 4.572349e+11 4.594172e+11 4.632723e+11
#>   Sole    8.426890e+11 8.179131e+11 7.942643e+11 7.717786e+11 7.504947e+11
#>   Gurnard 7.448789e+11 7.224720e+11 7.012017e+11 6.811112e+11 6.622456e+11
#>   Plaice  6.186276e+11 6.095894e+11 6.019644e+11 5.957646e+11 5.910047e+11
#>   Haddock 1.004945e+12 9.754200e+11 9.471535e+11 9.202015e+11 8.946276e+11
#>   Cod     4.339798e+11 4.278078e+11 4.228619e+11 4.191540e+11 4.166994e+11
#>   Saithe  2.864407e+11 2.782303e+11 2.707569e+11 2.640145e+11 2.579962e+11
#>          w
#> sp                 6.9         8.23         9.82         11.7           14
#>   Sprat   6.790203e+11 6.604670e+11 6.423092e+11 6.245352e+11 6.071314e+11
#>   Sandeel 2.258123e+12 2.200509e+12 2.144077e+12 2.088801e+12 2.034654e+12
#>   N.pout  4.435682e+11 4.484216e+11 4.547879e+11 4.627602e+11 4.714898e+11
#>   Herring 3.827955e+12 3.730503e+12 3.635155e+12 3.541876e+12 3.450629e+12
#>   Dab     7.115350e+11 7.023392e+11 6.948416e+11 6.890957e+11 6.850002e+11
#>   Whiting 4.688696e+11 4.762837e+11 4.856418e+11 4.971115e+11 5.099884e+11
#>   Sole    7.304438e+11 7.116565e+11 6.941646e+11 6.780019e+11 6.631446e+11
#>   Gurnard 6.446427e+11 6.283394e+11 6.133736e+11 5.997844e+11 5.875512e+11
#>   Plaice  5.876824e+11 5.857940e+11 5.853419e+11 5.863363e+11 5.886962e+11
#>   Haddock 8.704826e+11 8.478184e+11 8.266929e+11 8.071678e+11 7.892205e+11
#>   Cod     4.155072e+11 4.155910e+11 4.169741e+11 4.196937e+11 4.237535e+11
#>   Saithe  2.526939e+11 2.481003e+11 2.442101e+11 2.410215e+11 2.384982e+11
#>          w
#> sp                16.7         19.9         23.8         28.4         33.8
#>   Sprat   5.900816e+11 5.733673e+11 5.569677e+11 5.408608e+11 5.250242e+11
#>   Sandeel 1.981612e+12 1.929646e+12 1.878733e+12 1.828844e+12 1.779954e+12
#>   N.pout  4.818908e+11 4.940350e+11 5.079383e+11 5.238708e+11 5.420408e+11
#>   Herring 3.361385e+12 3.274117e+12 3.188795e+12 3.105396e+12 3.023900e+12
#>   Dab     6.828012e+11 6.825358e+11 6.840918e+11 6.874279e+11 6.925608e+11
#>   Whiting 5.254045e+11 5.434339e+11 5.635694e+11 5.856332e+11 6.096342e+11
#>   Sole    6.496875e+11 6.376572e+11 6.270366e+11 6.178271e+11 6.100477e+11
#>   Gurnard 5.767658e+11 5.674610e+11 5.596463e+11 5.533457e+11 5.485907e+11
#>   Plaice  5.925551e+11 5.979323e+11 6.047615e+11 6.130242e+11 6.227448e+11
#>   Haddock 7.729873e+11 7.585265e+11 7.458835e+11 7.351208e+11 7.262972e+11
#>   Cod     4.292679e+11 4.363178e+11 4.449634e+11 4.552986e+11 4.674432e+11
#>   Saithe  2.366638e+11 2.355201e+11 2.350707e+11 2.353228e+11 2.362827e+11
#>          w
#> sp                40.4         48.2         57.5         68.7         81.9
#>   Sprat   5.094370e+11 4.940806e+11 4.789413e+11 4.640111e+11 4.492895e+11
#>   Sandeel 1.732038e+12 1.685069e+12 1.639021e+12 1.593869e+12 1.549590e+12
#>   N.pout  5.618677e+11 5.828510e+11 6.051059e+11 6.285856e+11 6.532363e+11
#>   Herring 2.944287e+12 2.866542e+12 2.790655e+12 2.716617e+12 2.644423e+12
#>   Dab     6.994986e+11 7.082426e+11 7.189323e+11 7.316248e+11 7.463635e+11
#>   Whiting 6.353057e+11 6.623647e+11 6.911637e+11 7.217237e+11 7.540611e+11
#>   Sole    6.037191e+11 5.988596e+11 5.955159e+11 5.937171e+11 5.934922e+11
#>   Gurnard 5.453954e+11 5.437727e+11 5.437690e+11 5.454144e+11 5.487378e+11
#>   Plaice  6.339508e+11 6.466693e+11 6.609948e+11 6.769838e+11 6.946868e+11
#>   Haddock 7.194211e+11 7.145042e+11 7.116429e+11 7.109011e+11 7.123369e+11
#>   Cod     4.815106e+11 4.976128e+11 5.158982e+11 5.364802e+11 5.594442e+11
#>   Saithe  2.379432e+11 2.402913e+11 2.433128e+11 2.469831e+11 2.512682e+11
#>          w
#> sp                97.8          117          139          166          198
#>   Sprat   4.347834e+11 4.205079e+11 4.064852e+11 3.927441e+11 3.793186e+11
#>   Sandeel 1.506158e+12 1.463552e+12 1.421751e+12 1.380735e+12 1.340487e+12
#>   N.pout  6.792370e+11 7.077025e+11 7.388601e+11 7.705068e+11 8.005095e+11
#>   Herring 2.574074e+12 2.505581e+12 2.438949e+12 2.374169e+12 2.311235e+12
#>   Dab     7.632265e+11 7.824740e+11 8.040962e+11 8.276144e+11 8.525831e+11
#>   Whiting 7.884743e+11 8.264244e+11 8.679749e+11 9.100790e+11 9.499811e+11
#>   Sole    5.948761e+11 5.979465e+11 6.027109e+11 6.090649e+11 6.169093e+11
#>   Gurnard 5.537809e+11 5.606398e+11 5.693511e+11 5.798005e+11 5.918696e+11
#>   Plaice  7.141731e+11 7.356241e+11 7.590464e+11 7.841567e+11 8.106607e+11
#>   Haddock 7.160313e+11 7.221628e+11 7.308121e+11 7.417986e+11 7.549462e+11
#>   Cod     5.848560e+11 6.128133e+11 6.433001e+11 6.760657e+11 7.107888e+11
#>   Saithe  2.561301e+11 2.615420e+11 2.674680e+11 2.738382e+11 2.805818e+11
#>          w
#> sp                 237          282          337          402          480
#>   Sprat   3.662464e+11 3.535678e+11 3.413245e+11 3.295580e+11 3.183092e+11
#>   Sandeel 1.300991e+12 1.262237e+12 1.224214e+12 1.186918e+12 1.150344e+12
#>   N.pout  8.282768e+11 8.535252e+11 8.760886e+11 8.957597e+11 9.122098e+11
#>   Herring 2.250160e+12 2.190962e+12 2.133664e+12 2.078292e+12 2.024873e+12
#>   Dab     8.788432e+11 9.062968e+11 9.348293e+11 9.642794e+11 9.944268e+11
#>   Whiting 9.869498e+11 1.020753e+12 1.051268e+12 1.078280e+12 1.101407e+12
#>   Sole    6.262079e+11 6.369342e+11 6.490497e+11 6.624951e+11 6.771856e+11
#>   Gurnard 6.055159e+11 6.207024e+11 6.373796e+11 6.554735e+11 6.748790e+11
#>   Plaice  8.384066e+11 8.672551e+11 8.970341e+11 9.275200e+11 9.584312e+11
#>   Haddock 7.702321e+11 7.876476e+11 8.071832e+11 8.288073e+11 8.524518e+11
#>   Cod     7.471770e+11 7.848967e+11 8.235591e+11 8.627180e+11 9.018726e+11
#>   Saithe  2.876507e+11 2.950024e+11 3.025949e+11 3.103776e+11 3.182817e+11
#>          w
#> sp                 573          683          816          973         1160
#>   Sprat   3.076174e+11 2.975195e+11 2.880495e+11 2.792375e+11 2.711097e+11
#>   Sandeel 1.114497e+12 1.079381e+12 1.045008e+12 1.011392e+12 9.785555e+11
#>   N.pout  9.249938e+11 9.336384e+11 9.377331e+11 9.369582e+11 9.311291e+11
#>   Herring 1.973436e+12 1.924013e+12 1.876635e+12 1.831341e+12 1.788167e+12
#>   Dab     1.024991e+12 1.055646e+12 1.086031e+12 1.115760e+12 1.144424e+12
#>   Whiting 1.120097e+12 1.133739e+12 1.141774e+12 1.143722e+12 1.139224e+12
#>   Sole    6.930092e+11 7.098282e+11 7.274803e+11 7.457793e+11 7.645147e+11
#>   Gurnard 6.954563e+11 7.170335e+11 7.394094e+11 7.623533e+11 7.856065e+11
#>   Plaice  9.894268e+11 1.020115e+12 1.050062e+12 1.078798e+12 1.105824e+12
#>   Haddock 8.780089e+11 9.053355e+11 9.342591e+11 9.645770e+11 9.960579e+11
#>   Cod     9.404738e+11 9.779342e+11 1.013638e+12 1.046943e+12 1.077185e+12
#>   Saithe  3.262095e+11 3.340258e+11 3.415512e+11 3.485594e+11 3.547814e+11
#>          w
#> sp                1390         1650         1970         2360         2810
#>   Sprat   2.636875e+11 2.569872e+11 2.510193e+11 2.457886e+11 2.412939e+11
#>   Sandeel 9.465232e+11 9.153275e+11 8.850065e+11 8.556053e+11 8.271760e+11
#>   N.pout  9.201641e+11 9.041229e+11 8.831993e+11 8.577460e+11 8.281935e+11
#>   Herring 1.747155e+12 1.708345e+12 1.671780e+12 1.637504e+12 1.605560e+12
#>   Dab     1.171595e+12 1.196842e+12 1.219737e+12 1.239856e+12 1.256773e+12
#>   Whiting 1.128055e+12 1.110180e+12 1.085769e+12 1.055162e+12 1.018789e+12
#>   Sole    7.834528e+11 8.023388e+11 8.208996e+11 8.388453e+11 8.558694e+11
#>   Gurnard 8.088820e+11 8.318686e+11 8.542342e+11 8.756284e+11 8.956841e+11
#>   Plaice  1.130614e+12 1.152634e+12 1.171346e+12 1.186225e+12 1.196757e+12
#>   Haddock 1.028438e+12 1.061425e+12 1.094698e+12 1.127911e+12 1.160687e+12
#>   Cod     1.103681e+12 1.125734e+12 1.142652e+12 1.153761e+12 1.158432e+12
#>   Saithe  3.599109e+11 3.636217e+11 3.655907e+11 3.655312e+11 3.632240e+11
#>          w
#> sp                3350         4000         4780         5700         6800
#>   Sprat   2.375283e+11 2.344795e+11 2.321306e+11 2.304612e+11 2.294488e+11
#>   Sandeel 7.997791e+11 7.734838e+11 7.483690e+11 7.245247e+11 7.020525e+11
#>   N.pout  7.951138e+11 7.592039e+11 7.212476e+11 6.820792e+11 6.425488e+11
#>   Herring 1.575991e+12 1.548840e+12 1.524148e+12 1.501953e+12 1.482291e+12
#>   Dab     1.270092e+12 1.279459e+12 1.284564e+12 1.285157e+12 1.281052e+12
#>   Whiting 9.772768e+11 9.314406e+11 8.822384e+11 8.307289e+11 7.780284e+11
#>   Sole    8.716573e+11 8.858926e+11 8.982618e+11 9.084610e+11 9.162028e+11
#>   Gurnard 9.140291e+11 9.302940e+11 9.441203e+11 9.551689e+11 9.631300e+11
#>   Plaice  1.202473e+12 1.202965e+12 1.197900e+12 1.187043e+12 1.170269e+12
#>   Haddock 1.192636e+12 1.223354e+12 1.252434e+12 1.279468e+12 1.304059e+12
#>   Cod     1.156122e+12 1.146421e+12 1.129101e+12 1.104157e+12 1.071843e+12
#>   Saithe  3.585510e+11 3.515231e+11 3.422980e+11 3.311854e+11 3.186363e+11
#>          w
#> sp                8120         9690        11600        13800        16500
#>   Sprat   2.290708e+11 2.293064e+11 2.301399e+11 2.315626e+11 2.335766e+11
#>   Sandeel 6.810670e+11 6.616971e+11 6.440867e+11 6.283952e+11 6.147983e+11
#>   N.pout  6.034628e+11 5.655556e+11 5.294583e+11 4.956715e+11 4.645475e+11
#>   Herring 1.465196e+12 1.450694e+12 1.438807e+12 1.429553e+12 1.422939e+12
#>   Dab     1.272134e+12 1.258369e+12 1.239800e+12 1.216556e+12 1.188847e+12
#>   Whiting 7.252457e+11 6.734368e+11 6.235562e+11 5.764122e+11 5.326300e+11
#>   Sole    9.212238e+11 9.232912e+11 9.222107e+11 9.178323e+11 9.100569e+11
#>   Gurnard 9.677320e+11 9.687513e+11 9.660210e+11 9.594390e+11 9.489736e+11
#>   Plaice  1.147583e+12 1.119121e+12 1.085165e+12 1.046134e+12 1.002582e+12
#>   Haddock 1.325821e+12 1.344393e+12 1.359441e+12 1.370674e+12 1.377842e+12
#>   Cod     1.032697e+12 9.875383e+11 9.374543e+11 8.837555e+11 8.279149e+11
#>   Saithe  3.052168e+11 2.915675e+11 2.783542e+11 2.662137e+11 2.557018e+11
#>          w
#> sp               19600        23400        28000        33400        39900
#>   Sprat   2.361982e+11 2.394611e+11 2.434196e+11 2.481525e+11 2.537644e+11
#>   Sandeel 6.034877e+11 5.946702e+11 5.885658e+11 5.854055e+11 5.854268e+11
#>   N.pout  4.362816e+11 4.109164e+11 3.883584e+11 3.684096e+11 3.508099e+11
#>   Herring 1.418966e+12 1.417626e+12 1.418900e+12 1.422760e+12 1.429163e+12
#>   Dab     1.156957e+12 1.121247e+12 1.082143e+12 1.040133e+12 9.957591e+11
#>   Whiting 4.926283e+11 4.566134e+11 4.245956e+11 3.964283e+11 3.718669e+11
#>   Sole    8.988406e+11 8.841990e+11 8.662097e+11 8.450138e+11 8.208159e+11
#>   Gurnard 9.346682e+11 9.166434e+11 8.950973e+11 8.703031e+11 8.426049e+11
#>   Plaice  9.551774e+11 9.046900e+11 8.519607e+11 7.978767e+11 7.433407e+11
#>   Haddock 1.380750e+12 1.379261e+12 1.373300e+12 1.362865e+12 1.348021e+12
#>   Cod     7.714905e+11 7.160396e+11 6.630314e+11 6.137684e+11 5.693207e+11
#>   Saithe  2.472494e+11 2.411299e+11 2.374426e+11 2.361103e+11 2.368920e+11
# ->
getEncounter(params) / getSearchVolume(params)
#>          w
#> sp               0.001      0.00119      0.00142       0.0017      0.00203
#>   Sprat   2.351411e+12 2.296597e+12 2.243053e+12 2.190748e+12 2.139650e+12
#>   Sandeel 7.542068e+12 7.366333e+12 7.194680e+12 7.027015e+12 6.863242e+12
#>   N.pout  1.461295e+12 1.420432e+12 1.380615e+12 1.341840e+12 1.304096e+12
#>   Herring 1.279931e+13 1.250827e+13 1.222269e+13 1.194264e+13 1.166813e+13
#>   Dab     2.646976e+12 2.578483e+12 2.511343e+12 2.445543e+12 2.381075e+12
#>   Whiting 1.461295e+12 1.420436e+12 1.380626e+12 1.341861e+12 1.304131e+12
#>   Sole    2.930500e+12 2.856486e+12 2.783898e+12 2.712714e+12 2.642913e+12
#>   Gurnard 2.651205e+12 2.583464e+12 2.517021e+12 2.451858e+12 2.387964e+12
#>   Plaice  2.043473e+12 1.989174e+12 1.935963e+12 1.883841e+12 1.832812e+12
#>   Haddock 3.442482e+12 3.356698e+12 3.272603e+12 3.190165e+12 3.109350e+12
#>   Cod     1.517225e+12 1.475680e+12 1.435022e+12 1.395260e+12 1.356406e+12
#>   Saithe  1.180392e+12 1.147297e+12 1.114987e+12 1.083472e+12 1.052763e+12
#>          w
#> sp             0.00242      0.00289      0.00345      0.00411      0.00491
#>   Sprat   2.089730e+12 2.040958e+12 1.993304e+12 1.946738e+12 1.901230e+12
#>   Sandeel 6.703269e+12 6.547005e+12 6.394361e+12 6.245249e+12 6.099584e+12
#>   N.pout  1.267372e+12 1.231648e+12 1.196904e+12 1.163112e+12 1.130243e+12
#>   Herring 1.139915e+13 1.113569e+13 1.087771e+13 1.062516e+13 1.037799e+13
#>   Dab     2.317929e+12 2.256099e+12 2.195582e+12 2.136371e+12 2.078465e+12
#>   Whiting 1.267424e+12 1.231724e+12 1.197009e+12 1.163255e+12 1.130434e+12
#>   Sole    2.574478e+12 2.507394e+12 2.441648e+12 2.377232e+12 2.314137e+12
#>   Gurnard 2.325327e+12 2.263940e+12 2.203796e+12 2.144891e+12 2.087223e+12
#>   Plaice  1.782881e+12 1.734051e+12 1.686329e+12 1.639718e+12 1.594222e+12
#>   Haddock 3.030132e+12 2.952483e+12 2.876382e+12 2.801808e+12 2.728744e+12
#>   Cod     1.318470e+12 1.281460e+12 1.245385e+12 1.210247e+12 1.176047e+12
#>   Saithe  1.022866e+12 9.937864e+11 9.655236e+11 9.380744e+11 9.114307e+11
#>          w
#> sp             0.00586      0.00699      0.00834      0.00995       0.0119
#>   Sprat   1.856753e+12 1.813275e+12 1.770768e+12 1.729202e+12 1.688548e+12
#>   Sandeel 5.957280e+12 5.818256e+12 5.682429e+12 5.549718e+12 5.420045e+12
#>   N.pout  1.098265e+12 1.067142e+12 1.036837e+12 1.007313e+12 9.785304e+11
#>   Herring 1.013611e+13 9.899453e+12 9.667932e+12 9.441457e+12 9.219930e+12
#>   Dab     2.021859e+12 1.966549e+12 1.912531e+12 1.859798e+12 1.808344e+12
#>   Whiting 1.098516e+12 1.067466e+12 1.037249e+12 1.007830e+12 9.791740e+11
#>   Sole    2.252357e+12 2.191887e+12 2.132723e+12 2.074861e+12 2.018296e+12
#>   Gurnard 2.030789e+12 1.975588e+12 1.921617e+12 1.868875e+12 1.817358e+12
#>   Plaice  1.549841e+12 1.506577e+12 1.464425e+12 1.423380e+12 1.383435e+12
#>   Haddock 2.657173e+12 2.587081e+12 2.518456e+12 2.451285e+12 2.385559e+12
#>   Cod     1.142785e+12 1.110452e+12 1.079038e+12 1.048530e+12 1.018906e+12
#>   Saithe  8.855797e+11 8.605034e+11 8.361784e+11 8.125759e+11 7.896621e+11
#>          w
#> sp              0.0142       0.0169       0.0202       0.0241       0.0288
#>   Sprat   1.648776e+12 1.609856e+12 1.571758e+12 1.534451e+12 1.497905e+12
#>   Sandeel 5.293332e+12 5.169501e+12 5.048477e+12 4.930184e+12 4.814549e+12
#>   N.pout  9.504530e+11 9.230460e+11 8.962776e+11 8.701206e+11 8.445532e+11
#>   Herring 9.003255e+12 8.791332e+12 8.584057e+12 8.381329e+12 8.183046e+12
#>   Dab     1.758160e+12 1.709235e+12 1.661559e+12 1.615116e+12 1.569891e+12
#>   Whiting 9.512454e+11 9.240126e+11 8.974467e+11 8.715235e+11 8.462239e+11
#>   Sole    1.963024e+12 1.909038e+12 1.856333e+12 1.804899e+12 1.754725e+12
#>   Gurnard 1.767061e+12 1.717975e+12 1.670094e+12 1.623403e+12 1.577890e+12
#>   Plaice  1.344578e+12 1.306795e+12 1.270070e+12 1.234384e+12 1.199715e+12
#>   Haddock 2.321267e+12 2.258397e+12 2.196941e+12 2.136885e+12 2.078218e+12
#>   Cod     9.901445e+11 9.622164e+11 9.350908e+11 9.087336e+11 8.831084e+11
#>   Saithe  7.673983e+11 7.457416e+11 7.246465e+11 7.040652e+11 6.839499e+11
#>          w
#> sp              0.0343       0.0409       0.0489       0.0583       0.0696
#>   Sprat   1.462090e+12 1.426975e+12 1.392531e+12 1.358727e+12 1.325537e+12
#>   Sandeel 4.701499e+12 4.590962e+12 4.482866e+12 4.377143e+12 4.273722e+12
#>   N.pout  8.195596e+11 7.951309e+11 7.712658e+11 7.479703e+11 7.252584e+11
#>   Herring 7.989104e+12 7.799403e+12 7.613842e+12 7.432323e+12 7.254748e+12
#>   Dab     1.525868e+12 1.483029e+12 1.441354e+12 1.400824e+12 1.361420e+12
#>   Whiting 8.215350e+11 7.974509e+11 7.739725e+11 7.511083e+11 7.288738e+11
#>   Sole    1.705799e+12 1.658107e+12 1.611631e+12 1.566352e+12 1.522250e+12
#>   Gurnard 1.533536e+12 1.490322e+12 1.448226e+12 1.407222e+12 1.367284e+12
#>   Plaice  1.166041e+12 1.133340e+12 1.101586e+12 1.070759e+12 1.040838e+12
#>   Haddock 2.020926e+12 1.964995e+12 1.910409e+12 1.857151e+12 1.805202e+12
#>   Cod     8.581782e+11 8.339058e+11 8.102557e+11 7.871954e+11 7.646965e+11
#>   Saithe  6.642540e+11 6.449341e+11 6.259511e+11 6.072720e+11 5.888712e+11
#>          w
#> sp               0.083       0.0991        0.118        0.141        0.168
#>   Sprat   1.292934e+12 1.260894e+12 1.229393e+12 1.198413e+12 1.167936e+12
#>   Sandeel 4.172536e+12 4.073519e+12 3.976606e+12 3.881734e+12 3.788840e+12
#>   N.pout  7.031513e+11 6.816772e+11 6.608707e+11 6.407715e+11 6.214234e+11
#>   Herring 7.081022e+12 6.911053e+12 6.744751e+12 6.582027e+12 6.422795e+12
#>   Dab     1.323123e+12 1.285916e+12 1.249781e+12 1.214705e+12 1.180676e+12
#>   Whiting 7.072917e+11 6.863907e+11 6.662052e+11 6.467737e+11 6.281383e+11
#>   Sole    1.479302e+12 1.437484e+12 1.396771e+12 1.357138e+12 1.318559e+12
#>   Gurnard 1.328383e+12 1.290490e+12 1.253575e+12 1.217607e+12 1.182559e+12
#>   Plaice  1.011803e+12 9.836404e+11 9.563395e+11 9.298945e+11 9.043048e+11
#>   Haddock 1.754541e+12 1.705148e+12 1.657001e+12 1.610076e+12 1.564351e+12
#>   Cod     7.427360e+11 7.212976e+11 7.003725e+11 6.799601e+11 6.600681e+11
#>   Saithe  5.707310e+11 5.528422e+11 5.352042e+11 5.178245e+11 5.007185e+11
#>          w
#> sp               0.201         0.24        0.286        0.342        0.408
#>   Sprat   1.137951e+12 1.108448e+12 1.079421e+12 1.050869e+12 1.022793e+12
#>   Sandeel 3.697865e+12 3.608751e+12 3.521443e+12 3.435888e+12 3.352034e+12
#>   N.pout  6.028730e+11 5.851685e+11 5.683577e+11 5.524868e+11 5.375989e+11
#>   Herring 6.266974e+12 6.114483e+12 5.965245e+12 5.819185e+12 5.676231e+12
#>   Dab     1.147686e+12 1.115730e+12 1.084805e+12 1.054913e+12 1.026061e+12
#>   Whiting 6.103427e+11 5.934313e+11 5.774476e+11 5.624329e+11 5.484252e+11
#>   Sole    1.281009e+12 1.244464e+12 1.208902e+12 1.174304e+12 1.140650e+12
#>   Gurnard 1.148403e+12 1.115113e+12 1.082668e+12 1.051048e+12 1.020239e+12
#>   Plaice  8.795759e+11 8.557189e+11 8.327510e+11 8.106946e+11 7.895772e+11
#>   Haddock 1.519803e+12 1.476410e+12 1.434149e+12 1.393002e+12 1.352950e+12
#>   Cod     6.407129e+11 6.219186e+11 6.037171e+11 5.861466e+11 5.692509e+11
#>   Saithe  4.839079e+11 4.674202e+11 4.512870e+11 4.355431e+11 4.202248e+11
#>          w
#> sp               0.486         0.58        0.693        0.827        0.987
#>   Sprat   9.952004e+11 9.680989e+11 9.415002e+11 9.154176e+11 8.898659e+11
#>   Sandeel 3.269835e+12 3.189246e+12 3.110226e+12 3.032735e+12 2.956740e+12
#>   N.pout  5.237328e+11 5.109218e+11 4.991927e+11 4.885657e+11 4.790541e+11
#>   Herring 5.536315e+12 5.399371e+12 5.265334e+12 5.134144e+12 5.005744e+12
#>   Dab     9.982569e+11 9.715145e+11 9.458496e+11 9.212815e+11 8.978316e+11
#>   Whiting 5.354583e+11 5.235609e+11 5.127560e+11 5.030612e+11 4.944891e+11
#>   Sole    1.107928e+12 1.076125e+12 1.045234e+12 1.015253e+12 9.861794e+11
#>   Gurnard 9.902308e+11 9.610175e+11 9.325985e+11 9.049782e+11 8.781660e+11
#>   Plaice  7.694305e+11 7.502894e+11 7.321908e+11 7.151726e+11 6.992723e+11
#>   Haddock 1.313976e+12 1.276069e+12 1.239218e+12 1.203417e+12 1.168663e+12
#>   Cod     5.530774e+11 5.376762e+11 5.230983e+11 5.093939e+11 4.966114e+11
#>   Saithe  4.053690e+11 3.910120e+11 3.771886e+11 3.639312e+11 3.512690e+11
#>          w
#> sp                1.18          1.4         1.68            2         2.39
#>   Sprat   8.648604e+11 8.404162e+11 8.165474e+11 7.932670e+11 7.705861e+11
#>   Sandeel 2.882208e+12 2.809112e+12 2.737426e+12 2.667130e+12 2.598203e+12
#>   N.pout  4.706644e+11 4.633969e+11 4.572473e+11 4.522074e+11 4.482674e+11
#>   Herring 4.880076e+12 4.757089e+12 4.636731e+12 4.518954e+12 4.403712e+12
#>   Dab     8.755235e+11 8.543821e+11 8.344336e+11 8.157043e+11 7.982209e+11
#>   Whiting 4.870472e+11 4.807399e+11 4.755694e+11 4.715378e+11 4.686496e+11
#>   Sole    9.580188e+11 9.307785e+11 9.044693e+11 8.791052e+11 8.547029e+11
#>   Gurnard 8.521757e+11 8.270259e+11 8.027388e+11 7.793405e+11 7.568597e+11
#>   Plaice  6.845254e+11 6.709646e+11 6.586183e+11 6.475098e+11 6.376564e+11
#>   Haddock 1.134957e+12 1.102305e+12 1.070718e+12 1.040208e+12 1.010795e+12
#>   Cod     4.847956e+11 4.739865e+11 4.642189e+11 4.555211e+11 4.479151e+11
#>   Saithe  3.392279e+11 3.278294e+11 3.170910e+11 3.070257e+11 2.976421e+11
#>          w
#> sp                2.85          3.4         4.06         4.84         5.78
#>   Sprat   7.485138e+11 7.270567e+11 7.062193e+11 6.860036e+11 6.664090e+11
#>   Sandeel 2.530632e+12 2.464402e+12 2.399503e+12 2.335926e+12 2.273664e+12
#>   N.pout  4.454181e+11 4.436551e+11 4.429916e+11 4.435017e+11 4.453118e+11
#>   Herring 4.290962e+12 4.180661e+12 4.072770e+12 3.967253e+12 3.864072e+12
#>   Dab     7.820102e+11 7.670991e+11 7.535195e+11 7.413227e+11 7.305772e+11
#>   Whiting 4.669142e+11 4.663520e+11 4.670131e+11 4.690424e+11 4.726658e+11
#>   Sole    8.312812e+11 8.088612e+11 7.874661e+11 7.671245e+11 7.478697e+11
#>   Gurnard 7.353276e+11 7.147772e+11 6.952439e+11 6.767678e+11 6.593935e+11
#>   Plaice  6.290694e+11 6.217539e+11 6.157113e+11 6.109481e+11 6.074755e+11
#>   Haddock 9.825007e+11 9.553532e+11 9.293857e+11 9.046444e+11 8.811874e+11
#>   Cod     4.414163e+11 4.360343e+11 4.317753e+11 4.286481e+11 4.266666e+11
#>   Saithe  2.889446e+11 2.809337e+11 2.736066e+11 2.669588e+11 2.609844e+11
#>          w
#> sp                 6.9         8.23         9.82         11.7           14
#>   Sprat   6.474320e+11 6.290660e+11 6.113004e+11 5.941205e+11 5.775062e+11
#>   Sandeel 2.212710e+12 2.153060e+12 2.094708e+12 2.037648e+12 1.981876e+12
#>   N.pout  4.484916e+11 4.531284e+11 4.593704e+11 4.674259e+11 4.766121e+11
#>   Herring 3.763195e+12 3.664591e+12 3.568228e+12 3.474079e+12 3.382116e+12
#>   Dab     7.213281e+11 7.136255e+11 7.075412e+11 7.031723e+11 7.004831e+11
#>   Whiting 4.780215e+11 4.852731e+11 4.946759e+11 5.065830e+11 5.205507e+11
#>   Sole    7.297305e+11 7.127369e+11 6.969241e+11 6.823335e+11 6.689550e+11
#>   Gurnard 6.431606e+11 6.281099e+11 6.142864e+11 6.017406e+11 5.904691e+11
#>   Plaice  6.052925e+11 6.044005e+11 6.048138e+11 6.065654e+11 6.096118e+11
#>   Haddock 8.590644e+11 8.383299e+11 8.190503e+11 8.013038e+11 7.850956e+11
#>   Cod     4.258411e+11 4.261894e+11 4.277454e+11 4.305656e+11 4.346871e+11
#>   Saithe  2.556765e+11 2.510291e+11 2.470394e+11 2.437089e+11 2.410071e+11
#>          w
#> sp                16.7         19.9         23.8         28.4         33.8
#>   Sprat   5.614321e+11 5.458669e+11 5.307737e+11 5.161108e+11 5.018333e+11
#>   Sandeel 1.927382e+12 1.874159e+12 1.822196e+12 1.771479e+12 1.721995e+12
#>   N.pout  4.881120e+11 5.023860e+11 5.198802e+11 5.414439e+11 5.680282e+11
#>   Herring 3.292318e+12 3.204664e+12 3.119132e+12 3.035704e+12 2.954363e+12
#>   Dab     6.998584e+11 7.015492e+11 7.056532e+11 7.123267e+11 7.218350e+11
#>   Whiting 5.382433e+11 5.605352e+11 5.876555e+11 6.200826e+11 6.586250e+11
#>   Sole    6.569178e+11 6.463047e+11 6.371542e+11 6.295197e+11 6.234875e+11
#>   Gurnard 5.805965e+11 5.722056e+11 5.653614e+11 5.601518e+11 5.566919e+11
#>   Plaice  6.141721e+11 6.204037e+11 6.283847e+11 6.382426e+11 6.501905e+11
#>   Haddock 7.706089e+11 7.579717e+11 7.473169e+11 7.388211e+11 7.326894e+11
#>   Cod     4.402892e+11 4.475578e+11 4.566862e+11 4.679360e+11 4.816545e+11
#>   Saithe  2.389672e+11 2.376061e+11 2.369487e+11 2.370342e+11 2.379143e+11
#>          w
#> sp                40.4         48.2         57.5         68.7         81.9
#>   Sprat   4.878945e+11 4.742494e+11 4.608566e+11 4.476815e+11 4.346985e+11
#>   Sandeel 1.673724e+12 1.626647e+12 1.580740e+12 1.535977e+12 1.492331e+12
#>   N.pout  5.997974e+11 6.313967e+11 6.634961e+11 6.982089e+11 7.353068e+11
#>   Herring 2.875094e+12 2.797862e+12 2.722660e+12 2.649487e+12 2.578339e+12
#>   Dab     7.344681e+11 7.485462e+11 7.645452e+11 7.834527e+11 8.054326e+11
#>   Whiting 7.038372e+11 7.479444e+11 7.922063e+11 8.399299e+11 8.909013e+11
#>   Sole    6.191613e+11 6.162257e+11 6.148288e+11 6.152340e+11 6.175394e+11
#>   Gurnard 5.550945e+11 5.550534e+11 5.567148e+11 5.603452e+11 5.660490e+11
#>   Plaice  6.644866e+11 6.804479e+11 6.984115e+11 7.189749e+11 7.423572e+11
#>   Haddock 7.290876e+11 7.272324e+11 7.273702e+11 7.300150e+11 7.352941e+11
#>   Cod     4.982469e+11 5.175729e+11 5.400832e+11 5.664048e+11 5.969463e+11
#>   Saithe  2.396426e+11 2.422104e+11 2.456730e+11 2.501057e+11 2.555590e+11
#>          w
#> sp                97.8          117          139          166          198
#>   Sprat   4.218928e+11 4.092611e+11 3.968118e+11 3.845640e+11 3.725461e+11
#>   Sandeel 1.449772e+12 1.408267e+12 1.367784e+12 1.328289e+12 1.289749e+12
#>   N.pout  7.746584e+11 8.168983e+11 8.603878e+11 9.029173e+11 9.428085e+11
#>   Herring 2.509213e+12 2.442117e+12 2.377043e+12 2.313976e+12 2.252913e+12
#>   Dab     8.306787e+11 8.595563e+11 8.919168e+11 9.273207e+11 9.654858e+11
#>   Whiting 9.450757e+11 1.003435e+12 1.063903e+12 1.122882e+12 1.178019e+12
#>   Sole    6.218547e+11 6.283365e+11 6.370179e+11 6.478624e+11 6.608729e+11
#>   Gurnard 5.739464e+11 5.842129e+11 5.968882e+11 6.119368e+11 6.293554e+11
#>   Plaice  7.687936e+11 7.986207e+11 8.318412e+11 8.682405e+11 9.076348e+11
#>   Haddock 7.433573e+11 7.544467e+11 7.685707e+11 7.856370e+11 8.056243e+11
#>   Cod     6.320979e+11 6.722583e+11 7.175664e+11 7.679187e+11 8.230738e+11
#>   Saithe  2.620794e+11 2.697175e+11 2.784755e+11 2.883157e+11 2.991680e+11
#>          w
#> sp                 237          282          337          402          480
#>   Sprat   3.607942e+11 3.493499e+11 3.382583e+11 3.275666e+11 3.173217e+11
#>   Sandeel 1.252130e+12 1.215401e+12 1.179532e+12 1.144494e+12 1.110266e+12
#>   N.pout  9.796362e+11 1.012900e+12 1.041979e+12 1.066160e+12 1.084712e+12
#>   Herring 2.193866e+12 2.136852e+12 2.081886e+12 2.028986e+12 1.978175e+12
#>   Dab     1.006344e+12 1.049778e+12 1.095560e+12 1.143352e+12 1.192706e+12
#>   Whiting 1.228684e+12 1.274282e+12 1.314049e+12 1.347104e+12 1.372550e+12
#>   Sole    6.760999e+11 6.935757e+11 7.132938e+11 7.352033e+11 7.592051e+11
#>   Gurnard 6.491934e+11 6.714676e+11 6.961458e+11 7.231397e+11 7.522995e+11
#>   Plaice  9.499075e+11 9.948489e+11 1.042115e+12 1.091218e+12 1.141535e+12
#>   Haddock 8.286297e+11 8.547128e+11 8.838892e+11 9.161230e+11 9.513230e+11
#>   Cod     8.826528e+11 9.460278e+11 1.012295e+12 1.080278e+12 1.148551e+12
#>   Saithe  3.109309e+11 3.234447e+11 3.364839e+11 3.497529e+11 3.628892e+11
#>          w
#> sp                 573          683          816          973         1160
#>   Sprat   3.075696e+11 2.983535e+11 2.897130e+11 2.816830e+11 2.742930e+11
#>   Sandeel 1.076825e+12 1.044158e+12 1.012254e+12 9.811101e+11 9.507288e+11
#>   N.pout  1.096971e+12 1.102413e+12 1.100688e+12 1.091653e+12 1.075401e+12
#>   Herring 1.929476e+12 1.882914e+12 1.838520e+12 1.796326e+12 1.756367e+12
#>   Dab     1.243085e+12 1.293868e+12 1.344361e+12 1.393808e+12 1.441409e+12
#>   Whiting 1.389589e+12 1.397599e+12 1.396176e+12 1.385160e+12 1.364653e+12
#>   Sole    7.851500e+11 8.128355e+11 8.420038e+11 8.723406e+11 9.034763e+11
#>   Gurnard 7.834121e+11 8.161990e+11 8.503144e+11 8.853461e+11 9.208195e+11
#>   Plaice  1.192317e+12 1.242705e+12 1.291747e+12 1.338430e+12 1.381701e+12
#>   Haddock 9.893429e+11 1.029980e+12 1.072974e+12 1.118004e+12 1.164689e+12
#>   Cod     1.215485e+12 1.279308e+12 1.338184e+12 1.390300e+12 1.433953e+12
#>   Saithe  3.754724e+11 3.870409e+11 3.971143e+11 4.052217e+11 4.109342e+11
#>          w
#> sp                1390         1650         1970         2360         2810
#>   Sprat   2.675661e+11 2.615190e+11 2.561610e+11 2.514939e+11 2.475125e+11
#>   Sandeel 9.211212e+11 8.923065e+11 8.643135e+11 8.371815e+11 8.109621e+11
#>   N.pout  1.052228e+12 1.022636e+12 9.873112e+11 9.471088e+11 9.029457e+11
#>   Herring 1.718682e+12 1.683310e+12 1.650293e+12 1.619676e+12 1.591501e+12
#>   Dab     1.486337e+12 1.527761e+12 1.564875e+12 1.596904e+12 1.623120e+12
#>   Whiting 1.335014e+12 1.296873e+12 1.251128e+12 1.198833e+12 1.141097e+12
#>   Sole    9.349888e+11 9.664100e+11 9.972330e+11 1.026921e+12 1.054913e+12
#>   Gurnard 9.562027e+11 9.909167e+11 1.024348e+12 1.055860e+12 1.084806e+12
#>   Plaice  1.420508e+12 1.453835e+12 1.480742e+12 1.500393e+12 1.512088e+12
#>   Haddock 1.212589e+12 1.261207e+12 1.309995e+12 1.358363e+12 1.405674e+12
#>   Cod     1.467644e+12 1.490152e+12 1.500606e+12 1.498519e+12 1.483810e+12
#>   Saithe  4.138921e+11 4.138331e+11 4.106146e+11 4.042338e+11 3.948341e+11
#>          w
#> sp                3350         4000         4780         5700         6800
#>   Sprat   2.442045e+11 2.415515e+11 2.395299e+11 2.381126e+11 2.372709e+11
#>   Sandeel 7.857206e+11 7.615380e+11 7.385138e+11 7.167686e+11 6.964479e+11
#>   N.pout  8.558425e+11 8.068866e+11 7.571781e+11 7.077836e+11 6.597026e+11
#>   Herring 1.565815e+12 1.542660e+12 1.522082e+12 1.504120e+12 1.488814e+12
#>   Dab     1.642882e+12 1.655656e+12 1.661034e+12 1.658745e+12 1.648667e+12
#>   Whiting 1.079155e+12 1.014336e+12 9.480004e+11 8.814879e+11 8.160784e+11
#>   Sole    1.080647e+12 1.103565e+12 1.123139e+12 1.138878e+12 1.150345e+12
#>   Gurnard 1.110556e+12 1.132511e+12 1.150124e+12 1.162920e+12 1.170512e+12
#>   Plaice  1.515294e+12 1.509671e+12 1.495087e+12 1.471626e+12 1.439583e+12
#>   Haddock 1.451273e+12 1.494494e+12 1.534674e+12 1.571173e+12 1.603381e+12
#>   Cod     1.456810e+12 1.418244e+12 1.369192e+12 1.311038e+12 1.245406e+12
#>   Saithe  3.827061e+11 3.682788e+11 3.521008e+11 3.348143e+11 3.171210e+11
#>          w
#> sp                8120         9690        11600        13800        16500
#>   Sprat   2.369764e+11 2.372041e+11 2.379352e+11 2.391600e+11 2.408820e+11
#>   Sandeel 6.777253e+11 6.608078e+11 6.459398e+11 6.334083e+11 6.235484e+11
#>   N.pout  6.138002e+11 5.707876e+11 5.312037e+11 4.954014e+11 4.635453e+11
#>   Herring 1.476200e+12 1.466307e+12 1.459161e+12 1.454778e+12 1.453168e+12
#>   Dab     1.630830e+12 1.605419e+12 1.572761e+12 1.533325e+12 1.487700e+12
#>   Whiting 7.529269e+11 6.930320e+11 6.372060e+11 5.860496e+11 5.399347e+11
#>   Sole    1.157171e+12 1.159069e+12 1.155839e+12 1.147379e+12 1.133688e+12
#>   Gurnard 1.172615e+12 1.169054e+12 1.159777e+12 1.144854e+12 1.124478e+12
#>   Plaice  1.399461e+12 1.351946e+12 1.297894e+12 1.238292e+12 1.174230e+12
#>   Haddock 1.630740e+12 1.652749e+12 1.668985e+12 1.679113e+12 1.682894e+12
#>   Cod     1.174096e+12 1.099006e+12 1.022066e+12 9.451635e+11 8.700828e+11
#>   Saithe  2.997431e+11 2.833815e+11 2.686750e+11 2.561623e+11 2.462513e+11
#>          w
#> sp               19600        23400        28000        33400        39900
#>   Sprat   2.431210e+11 2.459176e+11 2.493375e+11 2.534767e+11 2.584665e+11
#>   Sandeel 6.167476e+11 6.134504e+11 6.141616e+11 6.194485e+11 6.299399e+11
#>   N.pout  4.356178e+11 4.114370e+11 3.906852e+11 3.729522e+11 3.577903e+11
#>   Herring 1.454331e+12 1.458257e+12 1.464922e+12 1.474292e+12 1.486318e+12
#>   Dab     1.436576e+12 1.380731e+12 1.321002e+12 1.258274e+12 1.193457e+12
#>   Whiting 4.990002e+11 4.631627e+11 4.321483e+11 4.055474e+11 3.828895e+11
#>   Sole    1.114866e+12 1.091114e+12 1.062726e+12 1.030087e+12 9.936603e+11
#>   Gurnard 1.098961e+12 1.068720e+12 1.034269e+12 9.962018e+11 9.551771e+11
#>   Plaice  1.106863e+12 1.037373e+12 9.669326e+11 8.966738e+11 8.276575e+11
#>   Haddock 1.680191e+12 1.670977e+12 1.655336e+12 1.633460e+12 1.605653e+12
#>   Cod     7.984432e+11 7.316491e+11 6.708506e+11 6.169156e+11 5.704166e+11
#>   Saithe  2.391969e+11 2.350898e+11 2.338548e+11 2.352600e+11 2.389343e+11
# }