Get density dependent reproduction rate

Calculates the density dependent rate of egg production \(R_i\) (units 1/year) for each species. This is the flux entering the smallest size class of each species. The density dependent rate is the density independent rate obtained with getRDI() after it has been put through the density dependence function. This is the Beverton-Holt function BevertonHoltRDD() by default, but this can be changed. See setReproduction() for more details.

getRDD(
  params,
  n = initialN(params),
  n_pp = initialNResource(params),
  n_other = initialNOther(params),
  rdi = getRDI(params, n = n, n_pp = n_pp, n_other = n_other)
)

Arguments

params
A MizerParams object

n
A matrix of species abundances (species x size).

n_pp
A vector of the resource abundance by size

n_other
A list of abundances for other dynamical components of the ecosystem

rdi
A vector of density-independent reproduction rates for each species. If not specified, rdi is calculated internally using getRDI().

Value

A numeric vector the length of the number of species.

See also

Examples

if (FALSE) {
params <- newMultispeciesParams(NS_species_params_gears, inter)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project(params, t_max = 20, effort = 0.5)
# Get the rate at a particular time step
getRDD(params,sim@n[21,,],sim@n_pp[21,])
}