Get mean trophic level of each species

Calculates the consumption-rate-weighted mean trophic level of each species, defined as $$ T_i = \frac{\int r_i(w)\,N_i(w)\,T_i(w)\,dw} {\int r_i(w)\,N_i(w)\,dw}, $$ where \(r_i(w) = (1 - f_i(w))\,E_i(w)\) is the consumption rate of an individual of species \(i\) at weight \(w\), \(N_i(w)\) is the abundance density, and \(T_i(w)\) is the size-resolved trophic level from getTrophicLevel().

Usage

getTrophicLevelBySpecies(
  params,
  n = initialN(params),
  n_pp = initialNResource(params),
  n_other = initialNOther(params),
  ...
)

Arguments

params

A MizerParams object

n

A matrix of species abundances (species x size).

n_pp

A vector of the resource abundance by size

n_other

A list of abundances for other dynamical components of the ecosystem

...

Unused

Value

A named vector with the mean trophic level for each species.

See also

Examples

getTrophicLevelBySpecies(NS_params)
#>    Sprat  Sandeel   N.pout  Herring      Dab  Whiting     Sole  Gurnard 
#> 1.000097 1.000046 1.215283 1.005033 1.278265 1.525713 1.217122 1.044212 
#>   Plaice  Haddock      Cod   Saithe 
#> 1.146182 1.151806 2.074701 1.602743