Calculate the rate at which biomass of each species is fished by each gear

This yield rate is given in grams per year. It is calculated at each time step saved in the MizerSim object.

Usage

getYieldGear(object)

Arguments

object

An object of class MizerParams or MizerSim.

Value

If called with a MizerParams object, an array (gear x species) with the yield rate in grams per year from each gear for each species in the model. If called with a MizerSim object, an array (time x gear x species) containing the yield rate at each time step.

Details

For details of how the yield rate is defined see the help page of getYield().

See also

Examples

yield <- getYieldGear(NS_sim)
yield["1972", "Herring", "Herring"]
#> [1] 80002050975
# (In this example MizerSim object each species was set up with its own gear)