Resource parameters

Functions for working with the resource parameters.

Usage

resource_params(params)

resource_params(params) <- value

getResourceLevel(params)

getResourceRate(params)

getResourceCapacity(params)

Arguments

params

A MizerParams object

value

A named list of resource parameters.

Value

A vector of the same length as w_full(params) with the value of the resource dynamics parameter in each size class.

Details

Both the resource_semichemostat() and the resource_logistic() dynamics are parametrised in terms of a size-dependent rate \(r_R(w)\) and a size-dependent capacity \(c_R\). You can see their current values with getResourceRate() and getResourceCapacity().

Due to the predation mortality, the actual resource abundance \(N_R\) is always lower than the capacity \(c_R\). The ratio \(N_R / c_R\) is denoted as the resource level. This can be NaN when both \(N_R\) and \(c_R\) are zero. You can see the current values with getResourceLevel().

The recommended way to change these paramters is with the functions setResourceSemichemostat() or setResourceLogistic() (dependent on your choice of dynamics). The resource_params list only contains values that are helpful in setting up the actual size-dependent parameters. Also If you have specified a different resource dynamics function that requires additional parameters, then these should also be added to the resource_params list.

The resource_params list will at least contain the slots kappa, lambda, w_pp_cutoff and n.

The resource parameter n is the exponent for the power-law form for the replenishment rate \(r_R(w)\): $$r_R(w) = r_R\, w^{n-1}.$$

The resource parameter lambda (\(\lambda\)) is the exponent for the power-law form for the carrying capacity \(c_R(w)\) and w_pp_cutoff is its cutoff value: $$c_R(w) = c_R w^{-\lambda}$$ for all \(w\) less than w_pp_cutoff and zero for larger sizes.

The resource parameter kappa (\(\kappa\)) determines the initial resource abundance: $$N_R(w) = \kappa\, w^{-\lambda}$$ for all \(w\) less than w_pp_cutoff and zero for larger sizes. Of course you can overrule this with initialNResource().

See also

Examples

resource_params(NS_params)
#> $kappa
#> [1] 1e+11
#> 
#> $lambda
#> [1] 2.133333
#> 
#> $r_pp
#> [1] 10
#> 
#> $n
#> [1] 0.6666667
#> 
#> $w_pp_cutoff
#> [1] 9.820907
#>