Set metabolic rate

Sets the rate at which energy is used for metabolism and activity

setMetabolicRate(params, metab = NULL, p = NULL, ...)

getMetabolicRate(params)

Arguments

params
MizerParams

metab
Optional. An array (species x size) holding the metabolic rate for each species at size. If not supplied, a default is set as described in the section "Setting metabolic rate".

p
The allometric metabolic exponent. This is only used if metab is not given explicitly and if the exponent is not specified in a p column in the species_params.

...
Unused

Value

MizerParams object with updated metabolic rate. Because of the way the R language works, setMetabolicRate() does not make the changes to the params object that you pass to it but instead returns a new params object. So to affect the change you call the function in the form params <- setMetabolicRate(params, ...).

Setting metabolic rate

The metabolic rate is subtracted from the energy income rate to calculate the rate at which energy is available for growth and reproduction, see getEReproAndGrowth(). It is measured in grams/year.

If the metab argument is not supplied, then for each species the metabolic rate \(k(w)\) for an individual of size \(w\) is set to $$k(w) = ks w^p + k w,$$ where \(ks w^p\) represents the rate of standard metabolism and \(k w\) is the rate at which energy is expended on activity and movement. The values of \(ks\), \(p\) and \(k\) are taken from the ks, p and k columns in the species parameter dataframe. If any of these parameters are not supplied, the defaults are \(k = 0\), \(p = n\) and $$ks = f_c h \alpha w_{mat}^{n-p},$$ where \(f_c\) is the critical feeding level taken from the fc column in the species parameter data frame. If the critical feeding level is not specified, a default of \(f_c = 0.2\) is used.

See also

Examples

if (FALSE) {
params <- NS_params
# Change activity coefficient for species 3
params@species_params$k[3] <- 8
params <- setMetabolicRate(params)
}