Set predation kernel

The predation kernel determines the distribution of prey sizes that a predator feeds on. It is used in getEncounter when calculating the rate at which food is encountered and in getPredRate when calculating the rate at which a prey is predated upon. The predation kernel can be a function of the predator/prey size ratio or it can be a function of the predator size and the prey size separately. Both types can be set up with this function.

setPredKernel(params, pred_kernel = NULL,
  pred_kernel_type = "lognormal", store_kernel = FALSE)

Arguments

params
A MizerParams object

pred_kernel
Optional. An array (species x predator size x prey size) that holds the predation coefficient of each predator at size on each prey size. The dimensions are thus no_sp, no_w, no_w_full.

pred_kernel_type
Only used if pred_kernel is not supplied. A string or vector of strings that determines which predation kernel function is used for each species. If not supplied, is taken from species parameter dataframe if possible, or defaults to "lognormal".

store_kernel
Only used if pred_kernel is not supplied. A boolean flag that determines whether the full three dimensional predation kernel array is calculated and stored, even though it is not needed by mizer. Only useful if you have your own code that relies on this array. Default is FALSE.

Value

A MizerParams object

Setting predation kernel

Kernel dependent on predator to prey size ratio
If the pred_kernel argument is not supplied, then this function sets a predation kernel that depends only on the ratio of predator mass to prey mass, not on the two masses independently. The shape of that kernel is then determined by the pred_kernel_type argument. The default pred_kernel_type is "lognormal". This will call the function lognormal_pred_kernel to calculate the predation kernel and the function lognormal_max_ppmr to return the maximal predator/prey mass ratio for each species. An alternative pred_kernel type is "box", implemented by the functions box_pred_kernel and box_max_ppmr. You can use any other string as the type. If for example you choose "my" then you need to define a function my_pred_kernel that you can model on the existing functions like lognormal_pred_kernel. You can also define a function my_max_ppmr if you want the smallest plankton size to be calculated automatically by set_multispecies_model, otherwise you need to specify the min_w_pp argument explicitly. When using a kernel dependent on the predator/prey size ratio only, mizer does not need to store the entire three dimensional array in the MizerParams object. Such an array can be very big when there is a large number of size bins. Instead mizer only needs to store two two-dimensional arrays that hold Fourier transforms of the feeding kernel function that allow the encounter rate and the predation rate to be calculated very efficiently. However mizer gives you the option of storing the full three-dimensional array anyway by setting store_kernel = TRUE. This might be useful if you have code for analysing the results of a mizer simulation that relies on the full array.
Kernel dependent on both predator and prey size
If you want to work with a feeding kernel that depends on predator mass and prey mass independently, you can specify the full feeding kernel as a three-dimensional array (predator species x predator size x prey size). You should use this option only if a kernel dependent only on the predator/prey mass ratio is not appropriate. Using a kernel dependent on predator/prey mass ratio only allows mizer to use fast Fourier transform methods to significantly reduce the running time of simulations. The order of the predator species in pred_kernel should be the same as the order in the species params dataframe. If you supply a named array then the function will check the order and warn if it is different.
Examples

# NOT RUN {
## Set up a MizerParams object
data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter)

## If you change predation kernel parameters after setting up a model, you
# need to call setPredKernel
params@species_params$beta["Cod"] <- 200
params <- setPredKernel(params)

## You can change to a different predation kernel type
params <- setPredKernel(params, pred_kernel_type = "box")
# }