Set search volume

Set search volume

setSearchVolume(params, search_vol = NULL, ...)

getSearchVolume(params)

Arguments

params
MizerParams

search_vol
Optional. An array (species x size) holding the search volume for each species at size. If not supplied, a default is set as described in the section "Setting search volume".

...
Unused

Value

MizerParams with updated search volume. Because of the way the R language works, setSearchVolume() does not make the changes to the params object that you pass to it but instead returns a new params object. So to affect the change you call the function in the form params <- setSearchVolume(params, ...).

Setting search volume

The search volume \(\gamma_i(w)\) of an individual of species \(i\) and weight \(w\) multiplies the predation kernel when calculating the encounter rate in getEncounter() and the predation rate in getPredRate().

The name "search volume" is a bit misleading, because \(\gamma_i(w)\) does not have units of volume. It is simply a parameter that determines the rate of predation. Its units depend on your choice, see section "Units in mizer". If you have chose to work with total abundances, then it is a rate with units 1/year. If you have chosen to work with abundances per m^2 then it has units of m^2/year. If you have chosen to work with abundances per m^3 then it has units of m^3/year.

If the search_vol argument is not supplied, then the search volume is set to $$\gamma_i(w) = \gamma_i w^q_i.$$ The values of \(\gamma_i\) (the search volume at 1g) and \(q_i\) (the allometric exponent of the search volume) are taken from the gamma and q columns in the species parameter dataframe. If the gamma column is not supplied in the species parameter dataframe, a default is calculated by the get_gamma_default() function. Note that only for predators of size \(w = 1\) gram is the value of the species parameter \(\gamma_i\) the same as the value of the search volume \(\gamma_i(w)\).

See also

Examples

if (FALSE) {
params <- newTraitParams()
params@species_params$gamma[3] <- 1000
params <- setSearchVolume(params)
}