Construct MizerParams object for multispecies model

Sets up a multi-species size spectrum model by filling all slots in the MizerParams object based on user-provided or default parameters. It does this by creating an empty MizerParams object with emptyParams and then filling the slots by passing its arguments to setParams.

set_multispecies_model(species_params, interaction = matrix(1, nrow =
  nrow(species_params), ncol = nrow(species_params)), no_w = 100,
  min_w = 0.001, max_w = NA, min_w_pp = NA, no_w_pp = NA,
  n = 2/3, p = 0.7, q = 0.8, kappa = 1e+11, lambda = (2 + q - n),
  f0 = 0.6, pred_kernel = NULL, pred_kernel_type = "lognormal",
  store_kernel = FALSE, search_vol = NULL, intake_max = NULL,
  metab = NULL, z0pre = 0.6, z0exp = n - 1, psi = NULL,
  r_pp = 10, w_pp_cutoff = 10,
  plankton_dynamics = plankton_semichemostat,
  resource_dynamics = list(), resource_params = list(), rho = NULL,
  srr = srrBevertonHolt)

Arguments

species_params
A data frame of species-specific parameter values.

interaction
Interaction matrix of the species (predator by prey). Entries should be numbers between 0 and 1. See "Setting interactions" section below.

no_w
The number of size bins in the consumer spectrum.

min_w
Sets the size of the eggs of all species for which this is not given in the w_min column of the species_params dataframe.

max_w
The largest size of the consumer spectrum. By default this is set to the largest w_inf specified in the species_params data frame.

min_w_pp
The smallest size of the plankton spectrum. By default this is set to the smallest value at which any of the consumers can feed.

no_w_pp
Obsolete argument that is no longer used because the number of plankton size bins is determined because all size bins have to be logarithmically equally spaced.

n
Scaling exponent of the intake rate.

p
Scaling exponent of the standard metabolic rate.

q
Exponent of the allometric search volume. Only needed if gamma is not an array.

kappa
Carrying capacity of the plankton spectrum.

lambda
Exponent of the plankton spectrum.

f0
Average feeding level. Used to calculated h and gamma if those are not columns in the species data frame. Also requires k_vb (the von Bertalanffy K parameter) to be a column in the species data frame. If h and gamma are supplied then this argument is ignored. Default is 0.6.

pred_kernel
Optional. An array (species x predator size x prey size) that holds the predation coefficient of each predator at size on each prey size. The dimensions are thus no_sp, no_w, no_w_full.

pred_kernel_type
Only used if pred_kernel is not supplied. A string or vector of strings that determines which predation kernel function is used for each species. If not supplied, is taken from species parameter dataframe if possible, or defaults to "lognormal".

store_kernel
Only used if pred_kernel is not supplied. A boolean flag that determines whether the full three dimensional predation kernel array is calculated and stored, even though it is not needed by mizer. Only useful if you have your own code that relies on this array. Default is FALSE.

metab
Optional. An array (species x size) holding the metabolic rate for each species at size. If not supplied, a default is set as described in the section "Setting metabolic rate".

z0pre
If z0, the mortality from other sources, is not a column in the species data frame, it is calculated as z0pre * w_inf ^ z0exp. Default value is 0.6.

z0exp
If z0, the mortality from other sources, is not a column in the species data frame, it is calculated as z0pre * w_inf ^ z0exp. Default value is n-1.

psi
Optional. An array (species x size) that holds the allocation to reproduction for each species at size. If not supplied, a default is set as described in the section "Setting reproduction".

r_pp
Growth rate of the primary productivity. Default is 10 g/year.

w_pp_cutoff
The upper cut off size of the plankton spectrum. Default is 10 g.

plankton_dynamics
Function that determines plankton dynamics by calculating the plankton spectrum at the next time step from the current state. The default is "plankton_semichemostat".

resource_dynamics
A named list of functions that determine the dynamics of the unstructured resources by calculating their biomasses at the next time step from the current state. See resource_dynamics for details. An empty list if the model does not have unstructured resources.

resource_params
A list of parameters needed by the resource_dynamics functions. An empty list if no parameters are needed.

rho
Either a 3-dim array (predator species x resource x predator size) or a 2-dim array (predator species x resource). In the latter case the size dependence is given by allometry.

srr
The stock recruitment function. Default is srrBevertonHolt.

Value

An object of type MizerParams

Species parameters

The only essential argument is a data frame which contains the species data. The data frame is arranged species by parameter, so each column of the parameter data frame is a parameter and each row has the parameters for one of the species in the model.

There are some essential columns that must be included in the species parameter data.frame and that do not have default values. Other columns do have default values, so that if they are not included in the species parameter data frame, they will be automatically added when the MizerParams object is created.

The following table gives different components of a species parameters data.frame.

Column name	Description

species	Name of the species	Compulsory (no default)

w_inf	The asymptotic mass of the species	Compulsory

w_mat	The maturation mass of the species	Compulsory

beta	Preferred predator prey mass ratio	Compulsory

sigma	Width of prey size preference	Compulsory

h	Maximum food intake rate. If this is not provided, it is calculated using the k_vb column. Therefore, either h or k_vb must be provided.	Optional

k_vb	The von Bertalanffy K parameter. Only used to calculate h if that column is not provided	Optional

gamma	Volumetric search rate. If this is not provided, it is calculated using the h column and other parameters.	Optional

ks	Standard metabolism coefficient	h*0.2

z0	Background mortality (constant for all sizes). If this is not provided then z0 is calculated as z0pre ∗ w_inf^z0exp. Also z0pre and z0exp have default values of 0.6 and -1/3 respectively.	Optional (no default)

k	Activity coefficient	0

alpha	Assimilation efficiency	0.6

erepro	Reproductive efficiency	1

w_min	The size class that recruits are placed in.	smallest size class of the species size spectrum

Setting interactions

The species interaction matrix \(\theta_{ij}\) is used when calculating the food encounter rate in getEncounter and the predation rate in getPredRate. This function checks that the supplied interaction matrix is valid and then stores it in the interaction slot of the params object before returning that object.

Any dimnames of the interaction matrix argument are ignored by the constructor. The dimnames of the interaction matrix in the returned MizerParams object are taken from the species names in the species_params slot. This means that the order of the columns and rows of the interaction matrix argument should be the same as the species name in the species_params slot.

Setting predation kernel

Kernel dependent on predator to prey size ratio
If the pred_kernel argument is not supplied, then this function sets a predation kernel that depends only on the ratio of predator mass to prey mass, not on the two masses independently. The shape of that kernel is then determined by the pred_kernel_type argument. The default pred_kernel_type is "lognormal". This will call the function lognormal_pred_kernel to calculate the predation kernel and the function lognormal_max_ppmr to return the maximal predator/prey mass ratio for each species. An alternative pred_kernel type is "box", implemented by the functions box_pred_kernel and box_max_ppmr. You can use any other string as the type. If for example you choose "my" then you need to define a function my_pred_kernel that you can model on the existing functions like lognormal_pred_kernel. You can also define a function my_max_ppmr if you want the smallest plankton size to be calculated automatically by set_multispecies_model, otherwise you need to specify the min_w_pp argument explicitly. When using a kernel dependent on the predator/prey size ratio only, mizer does not need to store the entire three dimensional array in the MizerParams object. Such an array can be very big when there is a large number of size bins. Instead mizer only needs to store two two-dimensional arrays that hold Fourier transforms of the feeding kernel function that allow the encounter rate and the predation rate to be calculated very efficiently. However mizer gives you the option of storing the full three-dimensional array anyway by setting store_kernel = TRUE. This might be useful if you have code for analysing the results of a mizer simulation that relies on the full array.
Kernel dependent on both predator and prey size
If you want to work with a feeding kernel that depends on predator mass and prey mass independently, you can specify the full feeding kernel as a three-dimensional array (predator species x predator size x prey size). You should use this option only if a kernel dependent only on the predator/prey mass ratio is not appropriate. Using a kernel dependent on predator/prey mass ratio only allows mizer to use fast Fourier transform methods to significantly reduce the running time of simulations. The order of the predator species in pred_kernel should be the same as the order in the species params dataframe. If you supply a named array then the function will check the order and warn if it is different.
Setting search volume

The search volume \(\gamma_i(w)\) of an individual of species \(i\) and weight \(w\) multiplies the predation kernel when calculating the encounter rate and the predation rate.

If the gamma argument is not supplied, then it is set to $$\gamma_i(w) = \gamma_i w^q.$$ The values of \(gamma_i\) are taken from the gamma column in the species parameter dataframe. If the gamma column is not supplied in the species parameter dataframe, it is calculated from f0, h, beta, sigma, lambda and kappa.

Setting maximum intake rate

The maximum intake rate \(h_i(w)\) of an individual of species \(i\) and weight \(w\) determines the feeding level, calculated with getFeedingLevel.

If the h argument is not supplied, then it is set to $$h_i(w) = h_i w^n.$$ The values of \(h_i\) are taken from the h column in the species parameter dataframe. If the h column is not supplied in the species parameter dataframe, it is calculated from f0 and the k_vb column, if they are supplied.

If \(h_i\) is set to Inf, fish will consume all encountered food.

Setting metabolic rate

To be documented

Setting background mortality rate

Still need to document

Setting reproduction

If the psi argument is not supplied, the proportion is set to the product of a sigmoidal maturity ogive that gives the proportion of individuals of a given species and size that are mature, and a factor that describes how investment into reproduction by mature individuals scales with size. In formulas: $$\psi(w) = \left[1+\left(\frac{w}{w_{mat}}\right)^{-U}\right]^{-1} \left(\frac{w}{w_{inf}}\right)^{m-n}.$$ Here \(n\) is the scaling exponent of the energy income rate. The exponent \(m\) determines the scaling of the investment into reproduction for mature individuals. By default it is chosen to be \(m = 1\) so that the rate at which energy is invested into reproduction scales linearly with the size. This default can be overridden by including a column m in the species parameter dataframe.

The exponent \(U\) determines the steepness of the maturity ogive. By default it is chosen as \(U = 10\), however this can be overridden by including a column w_mat25 in the species parameter dataframe that specifies the weight at which 25% of individuals are mature, which sets $$U = \frac{\log(3)}{\log(w_{mat} / w_{25})}$$

Setting plankton dynamics

Still need to document

Setting resource dynamics

Still need to document

Setting resource encounter rate

The resource encounter rate \(\rho_{id}(w)\) determines the rate at which an individual of species \(i\) encounters biomass of resource \(d\), $$\sum_d\rho_{id}(w) B_d,$$ where \(B_d\) is the biomass of the d-th unstructured resource component. See resource_dynamics for more details.

If only a two-dimensional array (species x resource) is given for the rho argument, then the size dependence is set to a power-law: $$\rho_{id}(w) = \rho_{id} w^n.$$

The ordering of the entries in the array rho is important. The order of the species in the first array dimension needs to be the same as that in the species parameter dataframe. The order of the resources in the second array dimension must be the same as in the list of resource dynamics. The third dimension, if given, is the size dimension.

Setting fishing

Needs to be documented

See also

Examples

data(NS_species_params_gears)
data(inter)
params <- set_multispecies_model(NS_species_params_gears, inter)

#> Note: No interaction_p column in species data frame so assuming all species feed on plankton.

#> Note: No sel_func column in species data frame. Setting selectivity to be 'knife_edge' for all species.

#> Note: No knife_edge_size column in species data frame. Setting knife edge selectivity equal to w_mat.

#> Note: No h provided for some species, so using f0 and k_vb to calculate it.

#> Note: No ks column in species data frame so using ks = h * 0.2.

#> Note: Using z0 = z0pre * w_inf ^ z0exp for missing z0 values.

#> Note: No gamma provided for some species, so using f0, h, beta, sigma, lambda and kappa to calculate it.