Length based sigmoid selectivity function

A sigmoid shaped selectivity function. Based on two parameters l25 and l50 which determine the length at which 25% and 50% of the stock is selected respectively. As the size-based model is weight based, and this selectivity function is length based, it uses the length-weight parameters a and b to convert between length and weight.

Usage

sigmoid_length(w, l25, l50, species_params, ...)

Arguments

w

Vector of sizes.

l25

the length which gives a selectivity of 25%.

l50

the length which gives a selectivity of 50%.

species_params

A list with the species params for the current species. Used to get at the length-weight parameters a and b.

...

Unused

Value

Vector of selectivities at the given sizes.

See also