Species parameters

These functions allow you to get or set the species-specific parameters stored in a MizerParams object.

Usage

species_params(params)

species_params(params) <- value

given_species_params(params)

given_species_params(params) <- value

calculated_species_params(params)

Arguments

params

A MizerParams object

value

A data frame with the species parameters

Value

Data frame of species parameters

Details

There are a lot of species parameters and we will list them all below, but most of them have sensible default values. The only required columns are species for the species name and w_max for its maximum size. However if you have information about the values of other parameters then you should provide them.

Mizer distinguishes between the species parameters that you have given explicitly and the species parameters that have been calculated by mizer or set to default values. You can retrieve the given species parameters with given_species_params() and the calculated ones with calculated_species_params(). You get all species_params with species_params().

If you change given species parameters with given_species_params<-() this will trigger a re-calculation of the calculated species parameters, where necessary. However if you change species parameters with species_params<-() no recalculation will take place and furthermore your values could be overwritten by a future recalculation triggered by a call to given_species_params<-() . So in most use cases you will only want to use given_species_params<-().

There are some species parameters that are used to set up the size-dependent parameters that are used in the mizer model:

gamma and q are used to set the search volume, see setSearchVolume().

h and n are used to set the maximum intake rate, see setMaxIntakeRate().

k, ks and p are used to set activity and basic metabolic rate, see setMetabolicRate().

z0 is used to set the external mortality rate, see setExtMort().

w_mat, w_mat25, w_repro_max and m are used to set the allocation to reproduction, see setReproduction().

pred_kernel_type specifies the shape of the predation kernel. The default is a "lognormal", for other options see the "Setting predation kernel" section in the help for setPredKernel().

beta and sigma are parameters of the lognormal predation kernel, see lognormal_pred_kernel(). There will be other parameters if you are using other predation kernel functions.

When you change one of the above species parameters using given_species_params<-() or species_params<-(), the new value will be used to update the corresponding size-dependent rates automatically, unless you have set those size-dependent rates manually, in which case the corresponding species parameters will be ignored.

There are some species parameters that are used directly in the model rather than being used for setting up size-dependent parameters:

alpha is the assimilation efficiency, the proportion of the consumed biomass that can be used for growth, metabolism and reproduction, see the help for getEReproAndGrowth().

w_min is the egg size.

interaction_resource sets the interaction strength with the resource, see "Predation encounter" section in the help for getEncounter().

erepro is the reproductive efficiency, the proportion of the energy invested into reproduction that is converted to egg biomass, see getRDI().

Rmax is the parameter in the Beverton-Holt density dependence added to the reproduction, see setBevertonHolt(). There will be other such parameters if you use other density dependence functions, see the "Density dependence" section in the help for setReproduction().

Two parameters are used only by functions that need to convert between weight and length:

a and b are the parameters in the allometric weight-length relationship \(w = a l ^ b\).

If you have supplied the a and b parameters, then you can replace weight parameters like w_max, w_mat, w_mat25, w_repro_max and w_min by their corresponding length parameters l_max, l_mat, l_mat25, l_repro_max and l_min.

The parameters that are only used to calculate default values for other parameters are:

f0 is the feeding level and is used to get a default value for the coefficient of the search volume gamma, see get_gamma_default().

fc is the critical feeding level below which the species can not maintain itself. This is used to get a default value for the coefficient ks of the metabolic rate, see get_ks_default().

age_mat is the age at maturity and is used to get a default value for the coefficient h of the maximum intake rate, see get_h_default().

Note that setting these parameters with species_params<-() will have no effect. You need to set them with given_species_params<-() in order to trigger a re-calculation of the other species parameters.

In the past, mizer also used the von Bertalanffy parameters k_vb, w_inf and t0 to determine a default for h. This is unreliable and is therefore now deprecated.

There are other species parameters that are used in tuning the model to observations:

biomass_observed and biomass_cutoff allow you to specify for each species the total observed biomass above some cutoff size. This is used by calibrateBiomass() and matchBiomasses().

yield_observed allows you to specify for each species the total annual fisheries yield. This is used by calibrateYield() and matchYields().

Finally there are two species parameters that control the way the species are represented in plots:

linecolour specifies the colour and can be any valid R colour value.

linetype specifies the line type ("solid", "dashed", "dotted", "dotdash", "longdash", "twodash" or "blank")

Other species-specific information that is related to how the species is fished is specified in a gear parameter data frame, see gear_params(). However in the case where each species is caught by only a single gear, this information can also optionally be provided as species parameters and newMultispeciesParams() will transfer them to the gear_params data frame. However changing these parameters later in the species parameter data frames will have no effect.

You are allowed to include additional columns in the species parameter data frames. They will simply be ignored by mizer but will be stored in the MizerParams object, in case your own code makes use of them.

See also