Display the structure of mizer objects

Mizer provides str() methods for MizerParams() and MizerSim() objects, as well as ArraySpeciesBySize(), ArrayTimeBySpecies() and ArrayTimeBySpeciesBySize() objects. These methods produce a clean, compact overview of the object's structure without polluting the console with large amounts of internal data.

Arguments

object

The object to display the structure of.

max.level

Maximum level of nesting to print. Defaults to NA (no limit).

...

Further arguments. They are passed to the default str() method.

Value

NULL, invisibly.

See also

Examples

# \donttest{
str(NS_params)
#> Formal class 'MizerParams' [package "mizer"] with 48 slots
#>  @ metadata            :  list()
#>  @ mizer_version       : Classes 'package_version', 'numeric_version'  hidden list of 1
#>     $ : int [1:4] 3 0 0 9003 
#>  @ extensions          :  chr(0) 
#>  @ time_created        :  POSIXct[1:1], format: "2021-09-03 20:29:38"
#>  @ time_modified       :  POSIXct[1:1], format: "2026-06-24 13:50:03"
#>  @ w                   :  num [1:100] 0.001 0.00119 0.00142 0.0017 0.00203 ...
#>  @ dw                  :  num [1:100] 0.000193 0.000231 0.000275 0.000329 0.000392 ...
#>  @ w_full              :  num [1:226] 2.12e-13 2.53e-13 3.02e-13 3.61e-13 4.30e-13 ...
#>  @ dw_full             :  num [1:226] 4.10e-14 4.90e-14 5.84e-14 6.97e-14 8.32e-14 ...
#>  @ w_min_idx           :  Named num [1:12] 1 1 1 1 1 1 1 1 1 1 ...
#>     - attr(*, "names")= chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ... 
#>  @ maturity            :  num [1:12, 1:100] 0 0 0 0 0 0 0 0 0 0 ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>  @ psi                 :  num [1:12, 1:100] 0 0 0 0 0 0 0 0 0 0 ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>  @ initial_n           :  num [1:12, 1:100] 1.84e+13 6.86e+12 1.44e+14 1.45e+13 1.71e+11 ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>  @ intake_max          :  num [1:12, 1:100] 0.182 0.275 0.328 0.35 0.307 ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>  @ search_vol          :  num [1:12, 1:100] 1.27e-13 6.00e-14 3.44e-13 4.49e-14 1.86e-13 ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>  @ metab               :  num [1:12, 1:100] 0.0289 0.0437 0.0522 0.0557 0.0487 ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>  @ pred_kernel         :  logi [1(1d)] NA
#>  @ ft_pred_kernel_e    :  cplx [1:12, 1:226] 11.3+0i 26.9+0i 20.8+0i ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ k : chr [1:226] "1" "2" "3" "4" ... 
#>  @ ft_pred_kernel_p    :  cplx [1:12, 1:226] 11.3+0i 26.9+0i 20.8+0i ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ k : chr [1:226] "1" "2" "3" "4" ... 
#>  @ ft_pred_kernel_d    :  cplx [1:12, 1:226] 11.3+0i 26.9+0i 20.8+0i ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ k : chr [1:226] "1" "2" "3" "4" ... 
#>  @ mu_b                :  num [1:12, 1:100] 0.1871 0.1817 0.1293 0.0865 0.0874 ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>  @ ext_encounter       :  num [1:12, 1:100] 0 0 0 0 0 0 0 0 0 0 ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>  @ ext_diffusion       :  num [1:12, 1:100] 0 0 0 0 0 0 0 0 0 0 ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>  @ rr_pp               :  Named num [1:226] 167660 158066 149021 140494 132455 ...
#>     - attr(*, "names")= chr [1:226] "2.12e-13" "2.53e-13" "3.02e-13" "3.61e-13" ... 
#>  @ cc_pp               :  Named num [1:226] 1.09e+38 7.46e+37 5.11e+37 3.51e+37 2.41e+37 ...
#>     - attr(*, "names")= chr [1:226] "2.12e-13" "2.53e-13" "3.02e-13" "3.61e-13" ... 
#>  @ resource_dynamics   :  chr "resource_semichemostat"
#>  @ resource_params     : List of 5
#>     $ kappa      : num 1e+11
#>     $ lambda     : num 2.13
#>     $ r_pp       : num 10
#>     $ n          : num 0.667
#>     $ w_pp_cutoff: num 9.82 
#>  @ other_dynamics      :  list()
#>  @ other_params        :  list()
#>  @ other_encounter     :  list()
#>  @ other_mort          :  list()
#>  @ rates_funcs         : List of 14
#>     $ Rates          : chr "mizerRates"
#>     $ Encounter      : chr "mizerEncounter"
#>     $ FeedingLevel   : chr "mizerFeedingLevel"
#>     $ EReproAndGrowth: chr "mizerEReproAndGrowth"
#>     $ PredRate       : chr "mizerPredRate"
#>     $ PredMort       : chr "mizerPredMort"
#>     $ FMort          : chr "mizerFMort"
#>     $ Mort           : chr "mizerMort"
#>     $ ERepro         : chr "mizerERepro"
#>     $ EGrowth        : chr "mizerEGrowth"
#>     $ ResourceMort   : chr "mizerResourceMort"
#>     $ RDI            : chr "mizerRDI"
#>     $ RDD            : chr "BevertonHoltRDD"
#>     $ Diffusion      : chr "mizerDiffusion" 
#>  @ sc                  :  num [1:100] 0.001 0.00119 0.00142 0.0017 0.00203 ...
#>  @ initial_n_pp        :  Named num [1:226] 1.09e+38 7.46e+37 5.11e+37 3.51e+37 2.41e+37 ...
#>     - attr(*, "names")= chr [1:226] "2.12e-13" "2.53e-13" "3.02e-13" "3.61e-13" ... 
#>  @ initial_n_other     :  list()
#>  @ species_params      : 'data.frame':	12 obs. of  34 variables:
#>     $ species             : chr  "Sprat" "Sandeel" "N.pout" "Herring" ...
#>     $ w_max               : num  33 36 100 334 324 ...
#>     $ w_mat               : int  13 4 23 99 21 75 78 39 105 165 ...
#>     $ beta                : int  51076 398849 22 280540 191 22 381 283 113 558 ...
#>     $ sigma               : num  0.8 1.9 1.5 3.2 1.9 1.5 1.9 1.8 1.6 2.1 ...
#>     $ k_vb                : num  0.681 1 0.849 0.606 0.536 0.323 0.284 0.266 0.122 0.271 ...
#>     $ gear                : chr  "Industrial" "Industrial" "Industrial" "Pelagic" ...
#>     $ w_min               : num  0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 ...
#>     $ alpha               : num  0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 ...
#>     $ pred_kernel_type    : chr  "lognormal" "lognormal" "lognormal" "lognormal" ...
#>     $ h                   : num  18.2 27.5 32.8 35 30.7 ...
#>     $ k                   : num  0 0 0 0 0 0 0 0 0 0 ...
#>     $ ks                  : num  3.64 5.5 6.57 7.01 6.14 ...
#>     $ z0                  : num  0.1871 0.1817 0.1293 0.0865 0.0874 ...
#>     $ gamma               : num  3.20e-11 1.51e-11 8.63e-11 1.13e-11 4.67e-11 ...
#>     $ m                   : num  1 1 1 1 1 1 1 1 1 1 ...
#>     $ erepro              : num  1 1 1 1 1 1 1 1 1 1 ...
#>     $ R_max               : num  7.38e+11 4.10e+11 1.05e+13 1.11e+12 1.12e+10 ...
#>     $ sel_func            : chr  "knife_edge" "knife_edge" "knife_edge" "knife_edge" ...
#>     $ knife_edge_size     : int  13 4 23 99 21 75 78 39 105 165 ...
#>     $ catchability        : num  1 1 1 1 1 1 1 1 1 1 ...
#>     $ p                   : num  0.7 0.7 0.7 0.7 0.7 0.7 0.7 0.7 0.7 0.7 ...
#>     $ q                   : num  0.8 0.8 0.8 0.8 0.8 0.8 0.8 0.8 0.8 0.8 ...
#>     $ n                   : num  0.667 0.667 0.667 0.667 0.667 ...
#>     $ f0                  : num  0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 ...
#>     $ interaction_resource: num  1 1 1 1 1 1 1 1 1 1 ...
#>     $ w_inf               : num  33 36 100 334 324 ...
#>     $ w_mat25             : num  11.65 3.58 20.61 88.7 18.82 ...
#>     $ w_repro_max         : num  33 36 100 334 324 ...
#>     $ is_background       : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>     $ z_ext               : num  0 0 0 0 0 0 0 0 0 0 ...
#>     $ d                   : num  -0.333 -0.333 -0.333 -0.333 -0.333 ...
#>     $ E_ext               : num  0 0 0 0 0 0 0 0 0 0 ...
#>     $ D_ext               : num  0 0 0 0 0 0 0 0 0 0 ... 
#>  @ given_species_params: 'data.frame':	12 obs. of  27 variables:
#>     $ species             : chr  "Sprat" "Sandeel" "N.pout" "Herring" ...
#>     $ w_max               : num  33 36 100 334 324 ...
#>     $ w_mat               : int  13 4 23 99 21 75 78 39 105 165 ...
#>     $ beta                : int  51076 398849 22 280540 191 22 381 283 113 558 ...
#>     $ sigma               : num  0.8 1.9 1.5 3.2 1.9 1.5 1.9 1.8 1.6 2.1 ...
#>     $ k_vb                : num  0.681 1 0.849 0.606 0.536 0.323 0.284 0.266 0.122 0.271 ...
#>     $ gear                : chr  "Industrial" "Industrial" "Industrial" "Pelagic" ...
#>     $ w_min               : num  0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 ...
#>     $ alpha               : num  0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 ...
#>     $ pred_kernel_type    : chr  "lognormal" "lognormal" "lognormal" "lognormal" ...
#>     $ h                   : num  18.2 27.5 32.8 35 30.7 ...
#>     $ k                   : num  0 0 0 0 0 0 0 0 0 0 ...
#>     $ ks                  : num  3.64 5.5 6.57 7.01 6.14 ...
#>     $ z0                  : num  0.1871 0.1817 0.1293 0.0865 0.0874 ...
#>     $ gamma               : num  3.20e-11 1.51e-11 8.63e-11 1.13e-11 4.67e-11 ...
#>     $ m                   : num  1 1 1 1 1 1 1 1 1 1 ...
#>     $ erepro              : num  1 1 1 1 1 1 1 1 1 1 ...
#>     $ R_max               : num  7.38e+11 4.10e+11 1.05e+13 1.11e+12 1.12e+10 ...
#>     $ sel_func            : chr  "knife_edge" "knife_edge" "knife_edge" "knife_edge" ...
#>     $ knife_edge_size     : int  13 4 23 99 21 75 78 39 105 165 ...
#>     $ catchability        : num  1 1 1 1 1 1 1 1 1 1 ...
#>     $ p                   : num  0.7 0.7 0.7 0.7 0.7 0.7 0.7 0.7 0.7 0.7 ...
#>     $ q                   : num  0.8 0.8 0.8 0.8 0.8 0.8 0.8 0.8 0.8 0.8 ...
#>     $ n                   : num  0.667 0.667 0.667 0.667 0.667 ...
#>     $ f0                  : num  0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 ...
#>     $ interaction_resource: num  1 1 1 1 1 1 1 1 1 1 ...
#>     $ w_inf               : num  33 36 100 334 324 ... 
#>  @ interaction         :  num [1:12, 1:12] 0.7291 0.0341 0.0635 0.2742 0.3624 ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ predator: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ prey    : chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ... 
#>  @ gear_params         : 'data.frame':	12 obs. of  5 variables:
#>     $ gear           : chr  "Industrial" "Industrial" "Industrial" "Pelagic" ...
#>     $ species        : chr  "Sprat" "Sandeel" "N.pout" "Herring" ...
#>     $ sel_func       : chr  "knife_edge" "knife_edge" "knife_edge" "knife_edge" ...
#>     $ knife_edge_size: int  13 4 23 99 21 75 78 39 105 165 ...
#>     $ catchability   : num  1 1 1 1 1 1 1 1 1 1 ... 
#>  @ selectivity         :  num [1:4, 1:12, 1:100] 0 0 0 0 0 0 0 0 0 0 ...
#>     - attr(*, "dimnames")=List of 3
#>      ..$ gear: chr [1:4] "Industrial" "Pelagic" "Beam" "Otter"
#>      ..$ sp  : chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ w   : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>  @ catchability        :  num [1:4, 1:12] 1 0 0 0 1 0 0 0 1 0 ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ gear: chr [1:4] "Industrial" "Pelagic" "Beam" "Otter"
#>      ..$ sp  : chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ... 
#>  @ initial_effort      :  Named num [1:4] 0 1 0.5 0.5
#>     - attr(*, "names")= chr [1:4] "Industrial" "Pelagic" "Beam" "Otter" 
#>  @ A                   :  num [1:12] 1 1 1 1 1 1 1 1 1 1 ...
#>  @ linecolour          :  Named chr [1:17] "#815f00" "#6237e2" "#8da600" "#de53ff" "#0e4300" ...
#>     - attr(*, "names")= chr [1:17] "Sprat" "Sandeel" "N.pout" "Herring" ... 
#>  @ linetype            :  Named chr [1:16] "solid" "solid" "solid" "solid" "solid" "solid" "solid" ...
#>     - attr(*, "names")= chr [1:16] "Sprat" "Sandeel" "N.pout" "Herring" ... 
#>  @ ft_mask             :  logi [1:12, 1:226] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  @ use_predation_diffusion:  logi FALSE
#>  @ second_order_w      : List of 2
#>     $ bin_average: logi FALSE
#>     $ flux       : chr "upwind" 
str(NS_sim)
#> Formal class 'MizerSim' [package "mizer"] with 6 slots
#>  @ params              : Formal class 'MizerParams' [package "mizer"] with 48 slots
#>     @ metadata            :  list()
#>     @ mizer_version       : Classes 'package_version', 'numeric_version'  hidden list of 1
#>        $ : int [1:4] 3 0 0 9003 
#>     @ extensions          :  chr(0) 
#>     @ time_created        :  POSIXct[1:1], format: "2021-09-03 20:30:02"
#>     @ time_modified       :  POSIXct[1:1], format: "2026-06-24 13:50:04"
#>     @ w                   :  num [1:100] 0.001 0.00119 0.00142 0.0017 0.00203 ...
#>     @ dw                  :  num [1:100] 0.000193 0.000231 0.000275 0.000329 0.000392 ...
#>     @ w_full              :  num [1:218] 8.73e-13 1.04e-12 1.24e-12 1.48e-12 1.77e-12 ...
#>     @ dw_full             :  num [1:218] 1.69e-13 2.01e-13 2.40e-13 2.87e-13 3.42e-13 ...
#>     @ w_min_idx           :  Named num [1:12] 1 1 1 1 1 1 1 1 1 1 ...
#>        - attr(*, "names")= chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ... 
#>     @ maturity            :  num [1:12, 1:100] 0 0 0 0 0 0 0 0 0 0 ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>     @ psi                 :  num [1:12, 1:100] 0 0 0 0 0 0 0 0 0 0 ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>     @ initial_n           :  num [1:12, 1:100] 1.67e+13 5.82e+12 1.19e+14 1.36e+13 1.17e+11 ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>     @ intake_max          :  num [1:12, 1:100] 0.145 0.256 0.312 0.284 0.349 ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>     @ search_vol          :  num [1:12, 1:100] 4.39e-13 2.94e-13 7.57e-13 1.95e-13 5.89e-13 ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>     @ metab               :  num [1:12, 1:100] 0.0127 0.0233 0.0268 0.0232 0.03 ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>     @ pred_kernel         :  logi [1(1d)] NA
#>     @ ft_pred_kernel_e    :  cplx [1:12, 1:218] 11.3+0i 26.9+0i 20.8+0i ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ k : chr [1:218] "1" "2" "3" "4" ... 
#>     @ ft_pred_kernel_p    :  cplx [1:12, 1:218] 11.3+0i 26.9+0i 20.8+0i ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ k : chr [1:218] "1" "2" "3" "4" ... 
#>     @ ft_pred_kernel_d    :  cplx [1:12, 1:218] 11.3+0i 26.9+0i 20.8+0i ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ k : chr [1:218] "1" "2" "3" "4" ... 
#>     @ mu_b                :  num [1:12, 1:100] 0.1871 0.1817 0.1293 0.0865 0.0874 ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>     @ ext_encounter       :  num [1:12, 1:100] 0 0 0 0 0 0 0 0 0 0 ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>     @ ext_diffusion       :  num [1:12, 1:100] 0 0 0 0 0 0 0 0 0 0 ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>     @ rr_pp               :  Named num [1:218] 104642 98654 93009 87686 82669 ...
#>        - attr(*, "names")= chr [1:218] "8.73e-13" "1.04e-12" "1.24e-12" "1.48e-12" ... 
#>     @ cc_pp               :  Named num [1:218] 4.88e+35 3.40e+35 2.36e+35 1.64e+35 1.14e+35 ...
#>        - attr(*, "names")= chr [1:218] "8.73e-13" "1.04e-12" "1.24e-12" "1.48e-12" ... 
#>     @ resource_dynamics   :  chr "resource_semichemostat"
#>     @ resource_params     : List of 5
#>        $ kappa      : num 9.27e+10
#>        $ lambda     : num 2.05
#>        $ r_pp       : num 10
#>        $ n          : num 0.667
#>        $ w_pp_cutoff: num 10 
#>     @ other_dynamics      :  list()
#>     @ other_params        :  list()
#>     @ other_encounter     :  list()
#>     @ other_mort          :  list()
#>     @ rates_funcs         : List of 14
#>        $ Rates          : chr "mizerRates"
#>        $ Encounter      : chr "mizerEncounter"
#>        $ FeedingLevel   : chr "mizerFeedingLevel"
#>        $ EReproAndGrowth: chr "mizerEReproAndGrowth"
#>        $ PredRate       : chr "mizerPredRate"
#>        $ PredMort       : chr "mizerPredMort"
#>        $ FMort          : chr "mizerFMort"
#>        $ Mort           : chr "mizerMort"
#>        $ ERepro         : chr "mizerERepro"
#>        $ EGrowth        : chr "mizerEGrowth"
#>        $ ResourceMort   : chr "mizerResourceMort"
#>        $ RDI            : chr "mizerRDI"
#>        $ RDD            : chr "BevertonHoltRDD"
#>        $ Diffusion      : chr "mizerDiffusion" 
#>     @ sc                  :  num [1:100] 0.001 0.00119 0.00142 0.0017 0.00203 ...
#>     @ initial_n_pp        :  Named num [1:218] 4.88e+35 3.40e+35 2.36e+35 1.64e+35 1.14e+35 ...
#>        - attr(*, "names")= chr [1:218] "8.73e-13" "1.04e-12" "1.24e-12" "1.48e-12" ... 
#>     @ initial_n_other     :  list()
#>     @ species_params      : 'data.frame':	12 obs. of  31 variables:
#>        $ species             : chr  "Sprat" "Sandeel" "N.pout" "Herring" ...
#>        $ w_max               : num  33 36 100 334 324 ...
#>        $ w_mat               : int  13 4 23 99 21 75 78 39 105 165 ...
#>        $ beta                : int  51076 398849 22 280540 191 22 381 283 113 558 ...
#>        $ sigma               : num  0.8 1.9 1.5 3.2 1.9 1.5 1.9 1.8 1.6 2.1 ...
#>        $ k_vb                : num  0.681 1 0.849 0.606 0.536 0.323 0.284 0.266 0.122 0.271 ...
#>        $ R_max               : num  7.38e+11 4.10e+11 1.05e+13 1.11e+12 1.12e+10 ...
#>        $ a                   : num  0.007 0.001 0.009 0.002 0.01 0.006 0.008 0.004 0.007 0.005 ...
#>        $ b                   : num  3.01 3.32 2.94 3.43 2.99 ...
#>        $ w_min               : num  0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 ...
#>        $ alpha               : num  0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 ...
#>        $ interaction_resource: num  1 1 1 1 1 1 1 1 1 1 ...
#>        $ n                   : num  0.667 0.667 0.667 0.667 0.667 ...
#>        $ p                   : num  0.7 0.7 0.7 0.7 0.7 0.7 0.7 0.7 0.7 0.7 ...
#>        $ q                   : num  0.717 0.717 0.717 0.717 0.717 ...
#>        $ pred_kernel_type    : chr  "lognormal" "lognormal" "lognormal" "lognormal" ...
#>        $ h                   : num  14.5 25.6 31.2 28.4 34.9 ...
#>        $ k                   : num  0 0 0 0 0 0 0 0 0 0 ...
#>        $ ks                  : num  1.59 2.94 3.37 2.92 3.78 ...
#>        $ z0                  : num  0.1871 0.1817 0.1293 0.0865 0.0874 ...
#>        $ gamma               : num  6.20e-11 4.16e-11 1.07e-10 2.76e-11 8.31e-11 ...
#>        $ w_mat25             : num  11.65 3.58 20.61 88.7 18.82 ...
#>        $ m                   : num  1 1 1 1 1 1 1 1 1 1 ...
#>        $ erepro              : num  1 1 1 1 1 1 1 1 1 1 ...
#>        $ w_repro_max         : num  33 36 100 334 324 ...
#>        $ is_background       : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>        $ z_ext               : num  0 0 0 0 0 0 0 0 0 0 ...
#>        $ d                   : num  -0.333 -0.333 -0.333 -0.333 -0.333 ...
#>        $ E_ext               : num  0 0 0 0 0 0 0 0 0 0 ...
#>        $ D_ext               : num  0 0 0 0 0 0 0 0 0 0 ...
#>        $ w_inf               : num  33 36 100 334 324 ... 
#>     @ given_species_params: 'data.frame':	12 obs. of  25 variables:
#>        $ species             : chr  "Sprat" "Sandeel" "N.pout" "Herring" ...
#>        $ w_max               : num  33 36 100 334 324 ...
#>        $ w_mat               : int  13 4 23 99 21 75 78 39 105 165 ...
#>        $ beta                : int  51076 398849 22 280540 191 22 381 283 113 558 ...
#>        $ sigma               : num  0.8 1.9 1.5 3.2 1.9 1.5 1.9 1.8 1.6 2.1 ...
#>        $ k_vb                : num  0.681 1 0.849 0.606 0.536 0.323 0.284 0.266 0.122 0.271 ...
#>        $ R_max               : num  7.38e+11 4.10e+11 1.05e+13 1.11e+12 1.12e+10 ...
#>        $ a                   : num  0.007 0.001 0.009 0.002 0.01 0.006 0.008 0.004 0.007 0.005 ...
#>        $ b                   : num  3.01 3.32 2.94 3.43 2.99 ...
#>        $ w_min               : num  0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 ...
#>        $ alpha               : num  0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 0.6 ...
#>        $ interaction_resource: num  1 1 1 1 1 1 1 1 1 1 ...
#>        $ n                   : num  0.667 0.667 0.667 0.667 0.667 ...
#>        $ p                   : num  0.7 0.7 0.7 0.7 0.7 0.7 0.7 0.7 0.7 0.7 ...
#>        $ q                   : num  0.717 0.717 0.717 0.717 0.717 ...
#>        $ pred_kernel_type    : chr  "lognormal" "lognormal" "lognormal" "lognormal" ...
#>        $ h                   : num  14.5 25.6 31.2 28.4 34.9 ...
#>        $ k                   : num  0 0 0 0 0 0 0 0 0 0 ...
#>        $ ks                  : num  1.59 2.94 3.37 2.92 3.78 ...
#>        $ z0                  : num  0.1871 0.1817 0.1293 0.0865 0.0874 ...
#>        $ gamma               : num  6.20e-11 4.16e-11 1.07e-10 2.76e-11 8.31e-11 ...
#>        $ w_mat25             : num  11.65 3.58 20.61 88.7 18.82 ...
#>        $ m                   : num  1 1 1 1 1 1 1 1 1 1 ...
#>        $ erepro              : num  1 1 1 1 1 1 1 1 1 1 ...
#>        $ w_inf               : num  33 36 100 334 324 ... 
#>     @ interaction         :  num [1:12, 1:12] 0.7291 0.0341 0.0635 0.2742 0.3624 ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ predator: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ prey    : chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ... 
#>     @ gear_params         : 'data.frame':	12 obs. of  6 variables:
#>        $ species     : chr  "Sprat" "Sandeel" "N.pout" "Herring" ...
#>        $ gear        : chr  "Sprat" "Sandeel" "N.pout" "Herring" ...
#>        $ sel_func    : chr  "sigmoid_length" "sigmoid_length" "sigmoid_length" "sigmoid_length" ...
#>        $ l25         : num  7.6 9.8 8.7 10.1 11.5 19.8 16.4 19.8 11.5 19.1 ...
#>        $ l50         : num  8.1 11.8 12.2 20.8 17 29 25.8 29 17 24.3 ...
#>        $ catchability: num  1.24 0.886 1.24 0.573 0.124 ... 
#>     @ selectivity         :  num [1:12, 1:12, 1:100] 5.9e-08 0.0 0.0 0.0 0.0 ...
#>        - attr(*, "dimnames")=List of 3
#>         ..$ gear: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ sp  : chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ w   : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>     @ catchability        :  num [1:12, 1:12] 1.24 0 0 0 0 ...
#>        - attr(*, "dimnames")=List of 2
#>         ..$ gear: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>         ..$ sp  : chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ... 
#>     @ initial_effort      :  Named num [1:12] 0 0 0 1.808 0.758 ...
#>        - attr(*, "names")= chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ... 
#>     @ A                   :  num [1:12] 1 1 1 1 1 1 1 1 1 1 ...
#>     @ linecolour          :  Named chr [1:17] "#815f00" "#6237e2" "#8da600" "#de53ff" "#0e4300" ...
#>        - attr(*, "names")= chr [1:17] "Sprat" "Sandeel" "N.pout" "Herring" ... 
#>     @ linetype            :  Named chr [1:17] "solid" "solid" "solid" "solid" "solid" "solid" "solid" ...
#>        - attr(*, "names")= chr [1:17] "Sprat" "Sandeel" "N.pout" "Herring" ... 
#>     @ ft_mask             :  logi [1:12, 1:218] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>     @ use_predation_diffusion:  logi FALSE
#>     @ second_order_w      : List of 2
#>        $ bin_average: logi FALSE
#>        $ flux       : chr "upwind"  
#>  @ n                   :  num [1:44, 1:12, 1:100] 1.67e+13 1.67e+13 1.69e+13 1.66e+13 1.65e+13 ...
#>     - attr(*, "dimnames")=List of 3
#>      ..$ time: chr [1:44] "1967" "1968" "1969" "1970" ...
#>      ..$ sp  : chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>      ..$ w   : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ... 
#>  @ effort              :  num [1:44, 1:12] 0 0 0 0 0 0 0 0 0 0 ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ time: chr [1:44] "1967" "1968" "1969" "1970" ...
#>      ..$ gear: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ... 
#>  @ n_pp                :  num [1:44, 1:218] 4.88e+35 4.88e+35 4.88e+35 4.88e+35 4.88e+35 ...
#>     - attr(*, "dimnames")=List of 2
#>      ..$ time: chr [1:44] "1967" "1968" "1969" "1970" ...
#>      ..$ w   : chr [1:218] "8.73e-13" "1.04e-12" "1.24e-12" "1.48e-12" ... 
#>  @ n_other             :  list()
#>     - attr(*, "dim")= int [1:2] 44 0
#>     - attr(*, "dimnames")=List of 2
#>      ..$ time     : chr [1:44] "1967" "1968" "1969" "1970" ...
#>      ..$ component: NULL 
#>  @ sim_params          :  list()
str(getEncounter(NS_params))
#>  'ArraySpeciesBySize' num [1:12, 1:100] 0.299 0.453 0.502 0.575 0.492 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ sp: chr [1:12] "Sprat" "Sandeel" "N.pout" "Herring" ...
#>   ..$ w : chr [1:100] "0.001" "0.00119" "0.00142" "0.0017" ...
#>  - attr(*, "value_name")= chr "Encounter rate"
#>  - attr(*, "units")= chr "g/year"
#>  - attr(*, "representation")= chr "point"
#>  - attr(*, "params")=Formal class 'MizerParams' [package "mizer"] with 48 slots
# }