Description of summary functions

Mizer provides a range of functions to extract information about a simulation from a MizerSim object.

Details

A list of available summary functions is given in the table below.

Function	Returns	Description

getDiet	Three dimensional array (predator x size x prey)	Diet of predator at size, resolved by prey species

getSSB	Two dimensional array (time x species)	Total Spawning Stock Biomass (SSB) of each species through time where SSB is calculated as the sum of weight of all mature individuals.

getBiomass	Two dimensional array (time x species)	Total biomass of each species through time.

getN	Two dimensional array (time x species)	Total abundance of each species through time.

getFeedingLevel	Three dimensional array (time x species x size)	Feeding level of each species by size through time.

getM2	Three dimensional array (time x species x size)	The predation mortality imposed on each species by size through time.

getFMort	Three dimensional array (time x species x size)	Total fishing mortality on each species by size through time.

getFMortGear	Four dimensional array (time x gear x species x size)	Fishing mortality on each species by each gear at size through time.

getYieldGear	Three dimensional array (time x gear x species)	Total yield by gear and species through time.

getYield	Two dimensional array (time x species)	Total yield of each species across all gears through time.

A list of available indicator functions for MizerSim objects is given in the table below

Function	Returns	Description

getProportionOfLargeFish	A vector with values at each time step.	Calculates the proportion of large fish through time. The threshold value can be specified. It is possible to calculation the proportion of large fish based on either length or weight.

getMeanWeight	A vector with values at each saved time step.	The mean weight of the community through time. This is calculated as the total biomass of the community divided by the total abundance.

getMeanMaxWeight	Depends on the measure argument. If measure = “both” then you get a matrix with two columns, one with values by numbers, the other with values by biomass at each saved time step. If measure = “numbers” or “biomass” you get a vector of the respective values at each saved time step	The mean maximum weight of the community through time. This can be calculated by numbers or by biomass. See the help file for more details.

getCommunitySlope	A data.frame with four columns: time step, slope, intercept and the coefficient of determination.	Calculates the slope of the community abundance spectrum through time by performing a linear regression on the logged total numerical abundance and logged body size.

See also