Description of summary functions

Mizer provides a range of functions to summarise the results of a simulation.

Details

A list of available summary functions is given in the table below.

Function	Returns	Description

getDiet()	Three dimensional array (predator x size x prey)	Diet of predator at size, resolved by prey species

getSSB()	Two dimensional array (time x species)	Total Spawning Stock Biomass (SSB) of each species through time where SSB is calculated as the sum of weight of all mature individuals.

getBiomass()	Two dimensional array (time x species)	Total biomass of each species through time.

getN()	Two dimensional array (time x species)	Total abundance of each species through time.

getFeedingLevel()	Three dimensional array (time x species x size)	Feeding level of each species by size through time.

getM2	Three dimensional array (time x species x size)	The predation mortality imposed on each species by size through time.

getFMort()	Three dimensional array (time x species x size)	Total fishing mortality on each species by size through time.

getFMortGear()	Four dimensional array (time x gear x species x size)	Fishing mortality on each species by each gear at size through time.

getYieldGear()	Three dimensional array (time x gear x species)	Total yield by gear and species through time.

getYield()	Two dimensional array (time x species)	Total yield of each species across all gears through time.

See also